Compute the chisquare goodness of fit statistic for an ROC data fitting model

Compute the chisquare goodness of fit statistic for specified ROC data fitting model

ChisqrGoodnessOfFit(fpCounts, tpCounts, parameters, model, lesDistr)

Arguments

fpCounts
The FP counts table

tpCounts
The TP counts table

parameters
The parameters of the model including cutoffs, see details

model
The fitting model: "BINORMAL", "CBM" or "RSM"

lesDistr
The lesion distribution matrix; not needed for "BINORMAL" or "CBM" models. Array [1:maxLL,1:2]. The probability mass function of the lesion distribution for diseased cases. The first column contains the actual numbers of lesions per case. The second column contains the fraction of diseased cases with the number of lesions specified in the first column. The second column must sum to unity.

Value

The return value is a list with the following elements:

chisq

The chi-square statistic

pVal

The p-value of the fit

df

The degrees of freedom

Details

For model = "BINORMAL" the parameters are c(a,b,zetas). For model = "CBM" the parameters are c(mu,alpha,zetas). For model = "RSM" the parameters are c(mu,lambdaP,nuP,zetas).