Extract two arms of a pairing from an MRMC ROC dataset

Extract a paired dataset from a larger dataset. The pairing could be two readers in the same modality, or different readers in different treatments, or the same reader in different treatments. If necessary The data is binned to 5 bins in each condition.

DfExtractCorCbmDataset(dataset, trts = 1, rdrs = 1)

Arguments

dataset

The original dataset from which the pairing is to be extracted

trts

A vector, maximum length 2, contains the indices of the modality or treatments to be extracted

rdrs

A vector, maximum length 2, contains the indices of the reader or readers to be extracted

Value

A 1-modality 2-reader dataset

Details

The desired pairing is contained in the vectors trts and rdrs. If either has length one, the other must have length two and the pairing is implicit. If both are length two, then the pairing is that implied by the first treatement and the second reader, which is one arm, and the other arm is that implied by the second modality paired with the first reader. Using this method any allowed pairing can be extracted and analyzed by FitCorCbm. The utility of this software is in designing a ratings simulator that is statistically matched to a real dataset.

Examples

# \donttest{
## Extract the paired data corresponding to the second and third readers in the first modality
## from the included ROC dataset
dataset11_23 <- DfExtractCorCbmDataset(dataset05, trts = 1, rdrs = c(2,3))

## Extract the paired data corresponding to the third reader in the first and second treatments
dataset12_33 <- DfExtractCorCbmDataset(dataset05, trts = c(1,2), rdrs = 3)

## Extract the data corresponding to the first reader in the first
## modality paired with the data
## from the third reader in the second modality
## (the bin indices are at different positions in the two arrays)
dataset12_13 <- DfExtractCorCbmDataset(dataset05,
trts = c(1,2), rdrs = c(1,3))
# }