Extract a subset of treatments and readers from a dataset

Extract a dataset consisting of a subset of treatments/readers from a larger dataset

DfExtractDataset(dataset, trts, rdrs)

Arguments

dataset
The original dataset from which the subset is to be extracted; can be ROC, FROC or LROC

trts
A vector contains the indices of the treatments to be extracted. If this parameter is not supplied, all treatments are extracted.

rdrs
A vector contains the indices of the readers to be extracted. If this parameter is not supplied, all readers are extracted.

Value

A new dataset containing only the specified treatments and readers that were extracted from the original dataset

Details

Note that trts and rdrs are the vectors of indices not IDs. For example, if the ID of the first reader is "0", the corresponding value in trts should be 1 not 0.

Examples

## Extract the data corresponding to the second reader in the 
## first treatment from an included ROC dataset
dataset1_2 <- DfExtractDataset(dataset05, trts = 1, rdrs = 2)

## Extract the data of the first and third reader in all 
## treatment from the included ROC dataset
datasetA_123 <- DfExtractDataset(dataset05, rdrs = c(1, 3))