Convert an FROC dataset to an AFROC dataset

Converts an FROC dataset to a AFROC dataset, where only the highest rated mark on each non-diseased case is counted and all lesion localizations are counted

DfFroc2Afroc(dataset)

Arguments

dataset
The dataset to be converted, RJafroc-package.

Value

An AFROC dataset

Details

The first list member of the AFROC dataset is NL, whose third dimension has length (K1 + K2), the total number of cases. The ratings of cases (K1 + 1) through (K1 + K2) are -Inf. In an AFROC dataset FPs are only possible on non-diseased cases. The second member of the list is LL. Its third dimension has length K2, the total number of diseased cases. This is because LLs are only possible on diseased cases. The structure is shown below:

NL Ratings array [1:I, 1:J, 1:(K1+K2), 1:maxNL], of non-lesion localizations, NLs

LL Ratings array [1:I, 1:J, 1:K2, 1:maxLL], of lesion localizations, LLs

lesionVector array [1:K2], number of lesions per diseased case

lesionID array [1:K2, 1:maxLL], labels of lesions on diseased cases

lesionWeight array [1:K2, 1:maxLL], weights (or clinical importances) of lesions

dataType "FROC", the data type

modalityID [1:I] inherited modality labels

readerID [1:J] inherited reader labels

Examples

afrocDataSet <- DfFroc2Afroc(dataset05)
p <- PlotEmpiricalOperatingCharacteristics(afrocDataSet, trts = 1, rdrs = 1, opChType = "wAFROC")
## print(p$Plot)
str(afrocDataSet)

#> List of 12
#>  $ NL           : num [1:2, 1:9, 1:92, 1] 6 5 -Inf -Inf 8 ...
#>  $ LL           : num [1:2, 1:9, 1:47, 1:3] 5 9 -Inf 10 8 ...
#>  $ lesionVector : int [1:47] 1 1 1 2 2 1 1 1 1 1 ...
#>  $ lesionID     : num [1:47, 1:3] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ lesionWeight : num [1:47, 1:3] 1 1 1 0.5 0.5 1 1 1 1 1 ...
#>  $ dataType     : chr "FROC"
#>  $ modalityID   : Named chr [1:2] "1" "2"
#>   ..- attr(*, "names")= chr [1:2] "1" "2"
#>  $ readerID     : Named chr [1:9] "1" "2" "3" "4" ...
#>   ..- attr(*, "names")= chr [1:9] "1" "2" "3" "4" ...
#>  $ design       : chr "CROSSED"
#>  $ normalCases  : int [1:45] 48 49 50 51 52 53 54 55 56 57 ...
#>  $ abnormalCases: int [1:47] 1 2 3 4 5 6 7 8 9 10 ...
#>  $ truthTableStr: num [1:2, 1:9, 1:92, 1:4] 1 1 1 1 1 1 1 1 1 1 ...