Read a crossed-treatment data file

Read an crossed-treatment data file, in which the two treatment factors are crossed

DfReadCrossedModalities(fileName, sequentialNames = FALSE)

Arguments

fileName

A string specifying the name of the file that contains the dataset, which must be an extended-JAFROC format data file containing an additional treatment factor.

sequentialNames

If TRUE, consecutive integers (starting from 1) will be used as the treatment and reader IDs. Otherwise, treatment and reader IDs in the original data file will be used. The default is FALSE.

Value

A dataset with the specified structure, similar to a standard

RJafroc dataset (see RJafroc-package). Because of the extra treatment factor,

NL and LL are each five dimensional arrays. There are also two treatment IDS: modalityID1 and modalityID2.

Details

The data format is similar to the JAFROC format (see RJafroc-package). The notable difference is that there are two treatment factors. A sample crossed treatment file "CrossedModalities.xlsx" is in the inst\extdata subdirectory of RJafroc.

References

Thompson JD, Chakraborty DP, Szczepura K, et al. (2016) Effect of reconstruction methods and x-ray tube current-time product on nodule detection in an anthropomorphic thorax phantom: a crossed-treatment JAFROC observer study. Medical Physics. 43(3):1265-1274.

Chakraborty DP (2017) Observer Performance Methods for Diagnostic Imaging - Foundations, Modeling, and Applications with R-Based Examples, CRC Press, Boca Raton, FL. https://www.crcpress.com/Observer-Performance-Methods-for-Diagnostic-Imaging-Foundations-Modeling/Chakraborty/p/book/9781482214840

Examples

# \donttest{
crossedFileName <- system.file("extdata", 
   "CrossedModalities.xlsx", package = "RJafroc", mustWork = TRUE)
crossedData <- DfReadCrossedModalities(crossedFileName)
str(crossedData)
#> List of 3
#>  $ ratings     :List of 3
#>   ..$ NL   : num [1:2, 1:4, 1:11, 1:68, 1:5] -Inf -Inf -Inf -Inf -Inf ...
#>   ..$ LL   : num [1:2, 1:4, 1:11, 1:34, 1:3] -Inf 1 2 8 7 ...
#>   ..$ LL_IL: logi NA
#>  $ lesions     :List of 3
#>   ..$ perCase: int [1:34] 1 2 1 1 1 1 2 2 1 1 ...
#>   ..$ IDs    : num [1:34, 1:3] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ weights: num [1:34, 1:3] 1 0.5 1 1 1 1 0.5 0.5 1 1 ...
#>  $ descriptions:List of 8
#>   ..$ fileName     : chr "DfReadCrossedModalities(CrossedModalities)"
#>   ..$ type         : chr "FROC"
#>   ..$ name         : chr "THOMPSON-X-MOD"
#>   ..$ truthTableStr: logi NA
#>   ..$ design       : chr "FCTRL-X-MOD"
#>   ..$ modalityID1  : Named chr [1:2] "F" "I"
#>   .. ..- attr(*, "names")= chr [1:2] "F" "I"
#>   ..$ modalityID2  : Named chr [1:4] "20" "40" "60" "80"
#>   .. ..- attr(*, "names")= chr [1:4] "20" "40" "60" "80"
#>   ..$ readerID     : Named chr [1:11] "1" "10" "11" "2" ...
#>   .. ..- attr(*, "names")= chr [1:11] "1" "10" "11" "2" ...
# }