Read sample LROC data files

Read the 2 Hupse-Karssemeijer LROC data files; this was a study comparing standalone performance of breast CAD vs. radiologists and residents.

DfReadLrocDataFile(RADIOLOGISTS = TRUE)

Arguments

RADIOLOGISTS

Logical; if TRUE, the default, only radiologists are analyzed otherwise all readers are analyzed.

Value

The LROC dataset.

Details

The data format is similar to the JAFROC format (see RJafroc-package) with the crucial difference that there are two types of LL (TP) events: those representing correct localizations and those representing incorrect localizations. Also, every diseased case has one lesion and NLs are not possible on diseased cases. J is one plus the number of readers. The first treatment is CAD, followed by the readers.

The return value is a list with the following elements:

NL [1, 1:J, 1:K1, 1] array containing the FP ratings

LLCl [1, 1:J, 1:K2, 1] array containing the TP correct localization ratings

LLIl [1, 1:J, 1:K2, 1] array containing the TP incorrect localization ratings

lesionVector array [1:K2], as in standard JAFROC/ROC format dataset, ones

lesionID array [1:K2], as in standard JAFROC/ROC format dataset, ones

lesionWeight array [1:K2], weights (or clinical importances) of lesions

dataType "LROC", the data type

modalityID [1:I], treatment labels

readerID [1:J], reader labels

References

Hupse R, Samulski M, Lobbes M, et al. Standalone computer-aided detection compared to radiologists' performance for the detection of mammographic masses. Eur Radiol 2013.

Chakraborty DP (2017) Observer Performance Methods for Diagnostic Imaging - Foundations, Modeling, and Applications with R-Based Examples, CRC Press, Boca Raton, FL. https://www.routledge.com/Observer-Performance-Methods-for-Diagnostic-Imaging-Foundations-Modeling/Chakraborty/p/book/9781482214840

Examples

radData <- DfReadLrocDataFile()
str(radData)
#> List of 10
#>  $ NL          : num [1, 1:10, 1:120, 1] 28 0 14 0 16 0 31 0 0 0 ...
#>  $ LLCl        : num [1, 1:10, 1:80, 1] 0 0 0 0 0 0 0 0 67 0 ...
#>  $ LLIl        : num [1, 1:10, 1:80, 1] 29 12 13 10 41 67 61 51 0 0 ...
#>  $ lesionVector: num [1:80] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ lesionID    : num [1:80] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ lesionWeight: num [1:80] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ dataType    : chr "LROC"
#>  $ modalityID  : chr "1"
#>  $ readerID    : chr [1:10] "1" "2" "3" "4" ...
#>  $ datasetName : chr "Ignore"
allData <- DfReadLrocDataFile(FALSE)
str(allData)
#> List of 10
#>  $ NL          : num [1, 1:13, 1:120, 1] 28 0 14 36 0 16 0 0 8 31 ...
#>  $ LLCl        : num [1, 1:13, 1:80, 1] 0 0 0 51 0 0 0 0 0 0 ...
#>  $ LLIl        : num [1, 1:13, 1:80, 1] 29 12 13 0 10 41 66 67 21 61 ...
#>  $ lesionVector: num [1:80] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ lesionID    : num [1:80] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ lesionWeight: num [1:80] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ dataType    : chr "LROC"
#>  $ modalityID  : chr "1"
#>  $ readerID    : chr [1:13] "1" "2" "3" "4" ...
#>  $ datasetName : chr "Ignore"