Read a crossed-modality data file

Read a crossed-modality data file, in which the two modality factors are crossed

DfReadXModalities(fileName, sequentialNames = FALSE)

Arguments

fileName

A string specifying the name of the file that contains the dataset, which must be an extended-JAFROC format Excel file containing an additional modality factor.

sequentialNames

If TRUE, consecutive integers (starting from 1) will be used as the modality and reader IDs. Otherwise, modality and reader IDs in the original data file will be used. The default is FALSE.

Value

A dataset with the specified structure, similar to a standard RJafroc dataset (see RJafroc-package). Because of the extra modality factor, NL and LL are each five dimensional arrays. There are also two modality IDS: modalityID1 and modalityID2.

Details

The data format is similar to the JAFROC format (see RJafroc-package). The difference is that there are two modality factors. TBA For an example see ... add reference to FROC book chapter https://dpc10ster.github.io/RJafrocFrocBook/

References

Thompson JD, Chakraborty DP, Szczepura K, et al. (2016) Effect of reconstruction methods and x-ray tube current-time product on nodule detection in an anthropomorphic thorax phantom: a crossed-modality JAFROC observer study. Medical Physics. 43(3):1265-1274.

Chakraborty DP (2017) Observer Performance Methods for Diagnostic Imaging - Foundations, Modeling, and Applications with R-Based Examples, CRC Press, Boca Raton, FL. https://www.routledge.com/Observer-Performance-Methods-for-Diagnostic-Imaging-Foundations-Modeling/Chakraborty/p/book/9781482214840