Save ROC dataset in different formats

Save ROC dataset in other formats so it can be analyzed with alternate software

DfSaveDataFile(
  dataset,
  fileName,
  format = "MRMC",
  dataDescription = "RJafroc dataset converted to imrmc format"
)

Arguments

dataset

The dataset to be saved.

fileName

The file name of the output data file. The extension of the data file must match the corresponding format, see RJafroc-package

format

The format of the data file, which can be "MRMC" or "iMRMC", see RJafroc-package.

dataDescription

An optional string variable describing the data file, the default value is the variable name of dataset The description appears on the first line of *.lrc or *imrmc data file. This parameter is not used when saving dataset in other formats.

Examples

# \donttest{
## DfSaveDataFile(dataset = dataset02, 
##    fileName = "rocData2.csv", format = "MRMC")
## DfSaveDataFile(dataset = dataset02, 
##    fileName = "rocData2.lrc", format = "MRMC", 
##     dataDescription = "ExampleROCdata1")
## DfSaveDataFile(dataset = dataset02, 
##    fileName = "rocData2.txt", format = "MRMC", 
##     dataDescription = "ExampleROCdata2")
##  DfSaveDataFile(dataset = dataset02, 
##    fileName = "dataset05.imrmc", format = "iMRMC", 
##    dataDescription = "ExampleROCdata3") 
# }