Fit CORCBM to a paired ROC dataset

Fit the Correlated Contaminated Binormal Model (CORCBM) to a paired ROC dataset. The ROC dataset has to be formatted as a single treatment, two-reader dataset, even though the actual pairing may be different, see details.

FitCorCbm(dataset)

Arguments

dataset
A paired ROC dataset

Value

The return value is a list containing three objects:

fitCorCbmRet

list(FPCounts,TPCounts, muX,muY,alphaX,alphaY,rhoNor, rhoAbn2,zetaX,zetaY,covMat,fixParam)

stats

list(aucX,aucX,stdAucX, stdAucY,stdErr,areaStat,areaPval)

fittedPlot

The fitted plot with operating points, error bars, for both conditions

Details

The conditions (X, Y) can be two readers interpreting images in the same treatment, the same reader interpreting images in different treatments, or different readers interpreting images in 2 different treatments. Function DfExtractCorCbmDataset can be used to construct a dataset suitable for FitCorCbm. With reference to the returned values, and assuming R bins in condition X and L bins in conditon Y, FPCounts is the R x L matrix containing the counts for non-diseased cases, TPCounts is the R x L matrix containing the counts for diseased cases; muX,muY,alphaX,alphaY,rhoNor,rhoAbn2 are the CORCBM parameters; aucX,aucX are the AUCs in the two conditions; stdAucX,stdAucY are the corresponding standard errors;stdErr contains the standard errors of the parameters of the model; areaStat, areaPval,covMat are the area-statistic, the p-value and the covariance matrix of the parameters. If a parameter approaches a limit, e.g., rhoNor = 0.9999, it is held constant at near the limiting value and the covariance matrix has one less dimension (along each edge) for each parameter that is held constant. The indices of the parameters held fixed are in fitCorCbmRet$fixParam.

References

Zhai X, Chakraborty DP (2017) A bivariate contaminated binormal model for robust fitting of proper ROC curves to a pair of correlated, possibly degenerate, ROC datasets. Medical Physics. 44(6):2207--2222.

Examples

# NOT RUN {
dataset <- DfExtractCorCbmDataset(dataset05, trts = 1, rdrs = c(4,7))
ret <- FitCorCbm(dataset)
print(ret$fitCorCbmRet)
print(ret$stats)
print(ret$fittedPlot)

ret <- FitCorCbm(binnedData123)
print(ret$fitCorCbmRet)
print(ret$stats)
print(ret$fittedPlot)

## Also try two other datasets ending with 124 and 125
# }