Simulate paired binned data for testing FitCorCbm

Simulates single treatment 2-reader binned ROC dataset, simulated according to the CORCBM model, for the purpose of testing the fitting program FitCorCbm.

SimulateCorCbmDataset(
  seed = 123,
  K1 = 50,
  K2 = 50,
  desiredNumBins = 5,
  muX = 1.5,
  muY = 3,
  alphaX = 0.4,
  alphaY = 0.7,
  rhoNor = 0.3,
  rhoAbn2 = 0.8
)

Arguments

seed
The seed variable, default is 123; set to NULL for truly random seed

K1
The number of non-diseased cases, default is 50

K2
The number of diseased cases, default is 50

desiredNumBins
The desired number of bins; default is 5

muX
The CBM mu parameter in condition X

muY
The CBM mu parameter in condition Y

alphaX
The CBM alpha parameter in condition X

alphaY
The CBM alpha parameter in condition Y

rhoNor
The correlation of non-diseased case z-samples

rhoAbn2
The correlation of diseased case z-samples, when disease is visible in both conditions

Value

The return value is the desired dataset, suitable for testing FitCorCbm

Details

X and Y refer to the two arms of the pairing. muX and alphaX refer to the univariate CBM parameters in condition X, rhoNor is the correlation of ratings of non-diseased cases and rhoAbn2 is the correlation of ratings of diseased cases when disease is visible in both conditions. The ROC data is bined to 5 bins in each condition. See referenced publication.

References

Zhai X, Chakraborty DP (2017) A bivariate contaminated binormal model for robust fitting of proper ROC curves to a pair of correlated, possibly degenerate, ROC datasets. Medical Physics. 44(6):2207--2222.

Examples

dataset <- SimulateCorCbmDataset()

# \donttest{
## this takes very long
## dataset <- SimulateCorCbmDataset(K1 = 5000, K2 = 5000)
# }