Simulates an uncorrelated FLROC FrocDataset using the RSM

Simulates an uncorrelated LROC dataset for specified numbers of readers and treatments

SimulateLrocDataset(mu, lambda, nu, zeta1, I, J, K1, K2, lesionVector)

Arguments

mu

The mu parameter of the RSM

lambda

The RSM lambda parameter

nu

The RSM nu parameter

zeta1

The lowest reporting threshold

I

The number of treatments

J

The number of readers

K1

The number of non-diseased cases

K2

The number of diseased cases

lesionVector

A K2 length array containing the numbers of lesions per diseased case

Value

An LROC dataset.

Details

See book chapters on the Radiological Search Model (RSM) for details. The approach is to first simulate an FROC dataset and then convert it to an Lroc dataset. The correlations between FROC ratings on the same case are assumed to be zero.

References

Chakraborty DP (2017) Observer Performance Methods for Diagnostic Imaging - Foundations, Modeling, and Applications with R-Based Examples, CRC Press, Boca Raton, FL. https://www.routledge.com/Observer-Performance-Methods-for-Diagnostic-Imaging-Foundations-Modeling/Chakraborty/p/book/9781482214840

Examples

  set.seed(1)
  K1 <- 5; K2 <- 5; mu <- 2; lambda <- 1; lesionVector <- rep(1, 5); nu <- 0.8; zeta1 <- -3
  frocData <- SimulateFrocDataset(mu, lambda, nu, zeta1, I = 2, J = 5, K1, K2, lesionVector)
  lrocData <- DfFroc2Lroc(frocData)