RSM fitted model for FROC sample size

RSM fitted model for FROC sample size

SsFrocNhRsmModel(dataset, lesDistr)

Arguments

dataset

The pilot dataset object representing a NH (ROC or FROC) dataset.

lesDistr

A 1D array containing the probability mass function of number of lesions per diseased case in the pivotal FROC study.

Value

A list containing:

muMed, the median mu parameter of the NH model.

lambdaMed, the median lambda parameter of the NH model.

nuMed, the median nu parameter of the NH model.

scaleFactor, the scaling factor that multiplies the ROC effect size to get wAFROC effect size.

R2, the R2 of the fit.

Details

If dataset is FROC, it is converted to an ROC dataset. The dataset is automatically binned. The search model is used to fit each treatment-reader combination. The median value for each parameter is computed and returned by the function (3 values). These are used to compute predicted wAFROC and ROC FOMS over a range of values of deltaMu, which are fitted by a straight line constrained to pass through the origin. The scale factor and R2 are returned. The scaling factor is the value by which the ROC effect size must be multiplied to get the wAFROC effect size. Equally weighted lesions is assumed.

Examples

 
# \donttest{
## Examples with CPU or elapsed time > 5s
## user system elapsed
## SsFrocNhRsmModel 8.102  0.023   8.135
 
## SsFrocNhRsmModel(dataset02, c(0.7, 0.2, 0.1))
## the next one should match the vignette 
## SsFrocNhRsmModel(DfExtractDataset(dataset04, trts = c(1,2)), c(0.69, 0.2, 0.11))
# }