Generate a power table

Generate combinations of numbers of readers J and numbers of cases K for desired power and specified generalizations (i.e., RRRC or FRRC or RRFC)

SsPowerTable(effectSize, alpha = 0.05, desiredPower = 0.8,
   method = "DBMH", option = "ALL", ...)

Arguments

effectSize
The postulated effect size

alpha
The The size of the test, default is 0.05

desiredPower
The desired statistical power, default is 0.8

method
Analysis method, "DBMH" or "ORH", the default is "DBMH"

option
Desired generalization; the default is "RRRC", for random-reader random-cases

...
Other necessary parameters, OR or DBM variance components, see details

Value

A data frame containing following three columns.

numReaders

The number of readers in the pivotal study.

numCases

The number of cases in the pivotal study.

power

The calculated statistical power corresponding to the indicated numbers of readers and cases.

Details

Regarding other parameters (...), see details in SsPowerGivenJK

Note

The procedure is valid for ROC studies only; for FROC studies see Online Appendix Chapter 19.

Examples

## Example of sample size calculation with DBM method
retDbm <- StSignificanceTesting (dataset02, FOM = "Wilcoxon", method = "DBMH")
effectSize <- retDbm$ciDiffTrtRRRC$Estimate
varYTR <- retDbm$varComp$varComp[3]
varYTC <- retDbm$varComp$varComp[4]
varYEps <- retDbm$varComp$varComp[6]
powTab <- SsPowerTable(
effectSize = effectSize,
method = "DBMH",
varYTR = varYTR,
varYTC = varYTC,
varYEps = varYEps)
print(powTab)

#> $powerTableRRRC
#>    numReaders numCases power
#> 1           3    >2000  <NA>
#> 2           4     1089   0.8
#> 3           5      344 0.801
#> 4           6      251 0.801
#> 5           7      211 0.801
#> 6           8      188 0.801
#> 7           9      173 0.801
#> 8          10      163 0.802
#> 9          11      155 0.801
#> 10         12      149 0.802
#> 11         13      144 0.801
#> 12         14      140 0.802
#> 13         15      137 0.802
#> 14         16      134 0.802
#> 15         17      131 0.801
#> 16         18      129 0.801
#> 17         19      127 0.801
#> 18         20      126 0.802
#> 19         21      124 0.801
#> 20         22      123 0.802
#> 21         23      122 0.802
#> 22         24      121 0.803
#> 23         25      120 0.802
#> 24         26      119 0.802
#> 25         27      118 0.802
#> 26         28      117 0.801
#> 27         29      117 0.803
#> 28         30      116 0.802
#> 29         31      115 0.801
#> 30         32      115 0.803
#> 31         33      114 0.801
#> 32         34      114 0.803
#> 33         35      113 0.801
#> 34         36      113 0.802
#> 35         37      112   0.8
#> 36         38      112 0.802
#> 37         39      112 0.803
#> 38         40      111 0.801
#> 39         41      111 0.802
#> 40         42      111 0.803
#> 41         43      110 0.801
#> 42         44      110 0.802
#> 43         45      110 0.802
#> 44         46      110 0.803
#> 45         47      109 0.801
#> 46         48      109 0.802
#> 47         49      109 0.802
#> 48         50      109 0.803
#> 49         51      108   0.8
#> 50         52      108 0.801
#> 51         53      108 0.802
#> 52         54      108 0.802
#> 53         55      108 0.803
#> 54         56      107   0.8
#> 55         57      107 0.801
#> 56         58      107 0.801
#> 57         59      107 0.802
#> 58         60      107 0.802
#> 59         61      107 0.803
#> 60         62      107 0.803
#> 61         63      106   0.8
#> 62         64      106 0.801
#> 63         65      106 0.801
#> 64         66      106 0.802
#> 65         67      106 0.802
#> 66         68      106 0.802
#> 67         69      106 0.803
#> 68         70      106 0.803
#> 69         71      106 0.804
#> 70         72      105   0.8
#> 71         73      105 0.801
#> 72         74      105 0.801
#> 73         75      105 0.801
#> 74         76      105 0.802
#> 75         77      105 0.802
#> 76         78      105 0.802
#> 77         79      105 0.803
#> 78         80      105 0.803
#> 79         81      105 0.803
#> 80         82      105 0.803
#> 81         83      104   0.8
#> 82         84      104   0.8
#> 83         85      104 0.801
#> 84         86      104 0.801
#> 85         87      104 0.801
#> 86         88      104 0.801
#> 87         89      104 0.802
#> 88         90      104 0.802
#> 89         91      104 0.802
#> 90         92      104 0.802
#> 91         93      104 0.802
#> 92         94      104 0.803
#> 93         95      104 0.803
#> 94         96      104 0.803
#> 95         97      104 0.803
#> 96         98      104 0.804
#> 97         99      104 0.804
#> 98        100      103   0.8
#> 
#> $powerTableFRRC
#>    numReaders numCases power
#> 1           3      209 0.800
#> 2           4      182 0.801
#> 3           5      166 0.802
#> 4           6      155 0.801
#> 5           7      147 0.801
#> 6           8      141 0.801
#> 7           9      137 0.802
#> 8          10      133 0.801
#> 9          11      130 0.801
#> 10         12      128 0.803
#> 11         13      126 0.803
#> 12         14      124 0.802
#> 13         15      122 0.801
#> 14         16      121 0.802
#> 15         17      120 0.803
#> 16         18      119 0.803
#> 17         19      118 0.803
#> 18         20      117 0.802
#> 19         21      116 0.802
#> 20         22      115 0.801
#> 21         23      115 0.803
#> 22         24      114 0.802
#> 23         25      114 0.803
#> 24         26      113 0.802
#> 25         27      113 0.803
#> 26         28      112 0.801
#> 27         29      112 0.803
#> 28         30      111 0.801
#> 29         31      111 0.802
#> 30         32      111 0.803
#> 31         33      110 0.801
#> 32         34      110 0.802
#> 33         35      110 0.803
#> 34         36      110 0.804
#> 35         37      109 0.801
#> 36         38      109 0.802
#> 37         39      109 0.803
#> 38         40      109 0.803
#> 39         41      108 0.800
#> 40         42      108 0.801
#> 41         43      108 0.802
#> 42         44      108 0.802
#> 43         45      108 0.803
#> 44         46      108 0.804
#> 45         47      107 0.800
#> 46         48      107 0.801
#> 47         49      107 0.801
#> 48         50      107 0.802
#> 49         51      107 0.802
#> 50         52      107 0.803
#> 51         53      107 0.803
#> 52         54      106 0.800
#> 53         55      106 0.800
#> 54         56      106 0.801
#> 55         57      106 0.801
#> 56         58      106 0.802
#> 57         59      106 0.802
#> 58         60      106 0.802
#> 59         61      106 0.803
#> 60         62      106 0.803
#> 61         63      106 0.803
#> 62         64      106 0.804
#> 63         65      105 0.800
#> 64         66      105 0.800
#> 65         67      105 0.801
#> 66         68      105 0.801
#> 67         69      105 0.801
#> 68         70      105 0.802
#> 69         71      105 0.802
#> 70         72      105 0.802
#> 71         73      105 0.802
#> 72         74      105 0.802
#> 73         75      105 0.803
#> 74         76      105 0.803
#> 75         77      105 0.803
#> 76         78      105 0.803
#> 77         79      105 0.804
#> 78         80      105 0.804
#> 79         81      104 0.800
#> 80         82      104 0.800
#> 81         83      104 0.801
#> 82         84      104 0.801
#> 83         85      104 0.801
#> 84         86      104 0.801
#> 85         87      104 0.801
#> 86         88      104 0.801
#> 87         89      104 0.802
#> 88         90      104 0.802
#> 89         91      104 0.802
#> 90         92      104 0.802
#> 91         93      104 0.802
#> 92         94      104 0.802
#> 93         95      104 0.802
#> 94         96      104 0.803
#> 95         97      104 0.803
#> 96         98      104 0.803
#> 97         99      104 0.803
#> 98        100      104 0.803
#> 
#> $powerTableRRFC
#>    numReaders numCases power
#> 1           3    >2000  <NA>
#> 2           4    >2000  <NA>
#> 3           5      289   0.8
#> 4           6      151 0.801
#> 5           7      102   0.8
#> 6           8       78 0.803
#> 7           9       63 0.804
#> 8          10       53 0.805
#> 9          11       45 0.801
#> 10         12       40 0.805
#> 11         13       36 0.808
#> 12         14       32 0.803
#> 13         15       29 0.801
#> 14         16       27 0.806
#> 15         17       25 0.807
#> 16         18       23 0.803
#> 17         19       22 0.811
#> 18         20       20   0.8
#> 19         21      <20  <NA>
#> 

## Example of sample size calculation with OR method
retOR <- StSignificanceTesting (dataset02, FOM = "Wilcoxon", method = "ORH")
effectSize <- retOR$ciDiffTrtRRRC$Estimate
varCompOR <- retOR$varComp
varTR <- varCompOR$varCov[2]
cov1 <- varCompOR$varCov[3]
cov2 <- varCompOR$varCov[4]
cov3 <- varCompOR$varCov[5]
varEps <- varCompOR$varCov[6]
KStar <- length(dataset02$NL[1,1,,1])
powTab <- SsPowerTable(
effectSize = effectSize,
method = "ORH",
KStar = KStar,
varTR = varTR,
cov1 = cov1,
cov2 = cov2,
cov3 = cov3,
varEps = varEps)
print(powTab)

#> $powerTableRRRC
#>    numReaders numCases power
#> 1           3    >2000  <NA>
#> 2           4     1089   0.8
#> 3           5      344 0.801
#> 4           6      251 0.801
#> 5           7      211 0.801
#> 6           8      188 0.801
#> 7           9      173 0.801
#> 8          10      163 0.802
#> 9          11      155 0.801
#> 10         12      149 0.802
#> 11         13      144 0.801
#> 12         14      140 0.802
#> 13         15      137 0.802
#> 14         16      134 0.802
#> 15         17      131 0.801
#> 16         18      129 0.801
#> 17         19      127 0.801
#> 18         20      126 0.802
#> 19         21      124 0.801
#> 20         22      123 0.802
#> 21         23      122 0.802
#> 22         24      121 0.803
#> 23         25      120 0.802
#> 24         26      119 0.802
#> 25         27      118 0.802
#> 26         28      117 0.801
#> 27         29      117 0.803
#> 28         30      116 0.802
#> 29         31      115 0.801
#> 30         32      115 0.803
#> 31         33      114 0.801
#> 32         34      114 0.803
#> 33         35      113 0.801
#> 34         36      113 0.802
#> 35         37      112   0.8
#> 36         38      112 0.802
#> 37         39      112 0.803
#> 38         40      111 0.801
#> 39         41      111 0.802
#> 40         42      111 0.803
#> 41         43      110 0.801
#> 42         44      110 0.802
#> 43         45      110 0.802
#> 44         46      110 0.803
#> 45         47      109 0.801
#> 46         48      109 0.802
#> 47         49      109 0.802
#> 48         50      109 0.803
#> 49         51      108   0.8
#> 50         52      108 0.801
#> 51         53      108 0.802
#> 52         54      108 0.802
#> 53         55      108 0.803
#> 54         56      107   0.8
#> 55         57      107 0.801
#> 56         58      107 0.801
#> 57         59      107 0.802
#> 58         60      107 0.802
#> 59         61      107 0.803
#> 60         62      107 0.803
#> 61         63      106   0.8
#> 62         64      106 0.801
#> 63         65      106 0.801
#> 64         66      106 0.802
#> 65         67      106 0.802
#> 66         68      106 0.802
#> 67         69      106 0.803
#> 68         70      106 0.803
#> 69         71      106 0.804
#> 70         72      105   0.8
#> 71         73      105 0.801
#> 72         74      105 0.801
#> 73         75      105 0.801
#> 74         76      105 0.802
#> 75         77      105 0.802
#> 76         78      105 0.802
#> 77         79      105 0.803
#> 78         80      105 0.803
#> 79         81      105 0.803
#> 80         82      105 0.803
#> 81         83      104   0.8
#> 82         84      104   0.8
#> 83         85      104 0.801
#> 84         86      104 0.801
#> 85         87      104 0.801
#> 86         88      104 0.801
#> 87         89      104 0.802
#> 88         90      104 0.802
#> 89         91      104 0.802
#> 90         92      104 0.802
#> 91         93      104 0.802
#> 92         94      104 0.803
#> 93         95      104 0.803
#> 94         96      104 0.803
#> 95         97      104 0.803
#> 96         98      104 0.804
#> 97         99      104 0.804
#> 98        100      103   0.8
#> 
#> $powerTableFRRC
#>    numReaders numCases power
#> 1           3      209 0.800
#> 2           4      182 0.801
#> 3           5      166 0.802
#> 4           6      155 0.801
#> 5           7      147 0.801
#> 6           8      141 0.801
#> 7           9      137 0.802
#> 8          10      133 0.801
#> 9          11      130 0.801
#> 10         12      128 0.803
#> 11         13      126 0.803
#> 12         14      124 0.802
#> 13         15      122 0.801
#> 14         16      121 0.802
#> 15         17      120 0.803
#> 16         18      119 0.803
#> 17         19      118 0.803
#> 18         20      117 0.802
#> 19         21      116 0.802
#> 20         22      115 0.801
#> 21         23      115 0.803
#> 22         24      114 0.802
#> 23         25      114 0.803
#> 24         26      113 0.802
#> 25         27      113 0.803
#> 26         28      112 0.801
#> 27         29      112 0.803
#> 28         30      111 0.801
#> 29         31      111 0.802
#> 30         32      111 0.803
#> 31         33      110 0.801
#> 32         34      110 0.802
#> 33         35      110 0.803
#> 34         36      110 0.804
#> 35         37      109 0.801
#> 36         38      109 0.802
#> 37         39      109 0.803
#> 38         40      109 0.803
#> 39         41      108 0.800
#> 40         42      108 0.801
#> 41         43      108 0.802
#> 42         44      108 0.802
#> 43         45      108 0.803
#> 44         46      108 0.804
#> 45         47      107 0.800
#> 46         48      107 0.801
#> 47         49      107 0.801
#> 48         50      107 0.802
#> 49         51      107 0.802
#> 50         52      107 0.803
#> 51         53      107 0.803
#> 52         54      106 0.800
#> 53         55      106 0.800
#> 54         56      106 0.801
#> 55         57      106 0.801
#> 56         58      106 0.802
#> 57         59      106 0.802
#> 58         60      106 0.802
#> 59         61      106 0.803
#> 60         62      106 0.803
#> 61         63      106 0.803
#> 62         64      106 0.804
#> 63         65      105 0.800
#> 64         66      105 0.800
#> 65         67      105 0.801
#> 66         68      105 0.801
#> 67         69      105 0.801
#> 68         70      105 0.802
#> 69         71      105 0.802
#> 70         72      105 0.802
#> 71         73      105 0.802
#> 72         74      105 0.802
#> 73         75      105 0.803
#> 74         76      105 0.803
#> 75         77      105 0.803
#> 76         78      105 0.803
#> 77         79      105 0.804
#> 78         80      105 0.804
#> 79         81      104 0.800
#> 80         82      104 0.800
#> 81         83      104 0.801
#> 82         84      104 0.801
#> 83         85      104 0.801
#> 84         86      104 0.801
#> 85         87      104 0.801
#> 86         88      104 0.801
#> 87         89      104 0.802
#> 88         90      104 0.802
#> 89         91      104 0.802
#> 90         92      104 0.802
#> 91         93      104 0.802
#> 92         94      104 0.802
#> 93         95      104 0.802
#> 94         96      104 0.803
#> 95         97      104 0.803
#> 96         98      104 0.803
#> 97         99      104 0.803
#> 98        100      104 0.803
#> 
#> $powerTableRRFC
#>    numReaders numCases power
#> 1           3    >2000  <NA>
#> 2           4    >2000  <NA>
#> 3           5      289   0.8
#> 4           6      151 0.801
#> 5           7      102   0.8
#> 6           8       78 0.803
#> 7           9       63 0.804
#> 8          10       53 0.805
#> 9          11       45 0.801
#> 10         12       40 0.805
#> 11         13       36 0.808
#> 12         14       32 0.803
#> 13         15       29 0.801
#> 14         16       27 0.806
#> 15         17       25 0.807
#> 16         18       23 0.803
#> 17         19       22 0.811
#> 18         20       20   0.8
#> 19         21      <20  <NA>
#>