Number of cases, for specified number of readers, to achieve desired power

Number of cases to achieve the desired power, for specified number of readers J, and specified DBMH or ORH analysis method

SsSampleSizeKGivenJ(
  dataset,
  ...,
  J,
  FOM,
  effectSize = NULL,
  method = "DBMH",
  alpha = 0.05,
  desiredPower = 0.8,
  option = "ALL"
)

Arguments

dataset
The pilot dataset. If set to NULL then variance components must be supplied.

...
Optional variance components, varYTR, varYTC and varYEps. These are needed if dataset is not supplied.

J
The number of readers in the pivotal study.

FOM
The figure of merit. Not needed if variance components are supplied.

effectSize
The effect size to be used in the pivotal study. Default is NULL. Must be supplied if dataset is set to NULL and variance components are supplied.

method
"DBMH" (default) or "ORH".

alpha
The significance level of the study, default is 0.05.

desiredPower
The desired statistical power, default is 0.8.

option
Desired generalization, "RRRC", "FRRC", "RRFC" or "ALL" (the default).

Value

A list of two elements:

K

The minimum number of cases K in the pivotal study to just achieve the desired statistical power. This is calculated for each value of option.

power

The predicted statistical power.

Details

effectSize = NULL uses the observed effect size in the pilot study. A numeric value over-rides the default value. This argument must be supplied if dataset = NULL and variance compenents (the optional ... arguments) are supplied.

References

Hillis SL, Obuchowski NA, Berbaum KS (2011) Power Estimation for Multireader ROC Methods: An Updated and Unified Approach, Acad Radiol, 18, 129--142.

Hillis SL, Obuchowski NA, Schartz KM, Berbaum KS (2005) A comparison of the Dorfman-Berbaum-Metz and Obuchowski-Rockette methods for receiver operating characteristic (ROC) data, Statistics in Medicine, 24:10, 1579--607.

Examples

## the following two should give identical results
SsSampleSizeKGivenJ(dataset02, FOM = "Wilcoxon", effectSize = 0.05, J = 6, method = "DBMH")

#>   KRRRC powerRRRC KFRRC powerFRRC KRRFC powerRRFC
#> 1   170 0.8016186   119 0.8001346    99 0.8020196

a <- UtilVarComponentsDBM(dataset02, FOM = "Wilcoxon")$varComp
SsSampleSizeKGivenJ(dataset = NULL, J = 6, effectSize = 0.05, method = "DBMH",
   list(varYTR = a$varTR, varYTC = a$varTC, varYEps = a$varErr))

#>   KRRRC powerRRRC KFRRC powerFRRC KRRFC powerRRFC
#> 1   170 0.8016186   119 0.8001346    99 0.8020196

## the following two should give identical results
SsSampleSizeKGivenJ(dataset02, FOM = "Wilcoxon", effectSize = 0.05, J = 6, method = "ORH")

#>   KRRRC powerRRRC KFRRC powerFRRC KRRFC powerRRFC
#> 1   170 0.8016186   119 0.8001346    99 0.8020196

a <- UtilVarComponentsOR(dataset02, FOM = "Wilcoxon")$varComp
KStar <- length(dataset02$NL[1,1,,1])
SsSampleSizeKGivenJ(dataset = NULL, J = 6, effectSize = 0.05, method = "ORH",
   list(KStar = KStar, varTR = a$varTR, cov1 = a$cov1, cov2 = a$cov2,
   cov3 = a$cov3, varEps = a$var))

#>   KRRRC powerRRRC KFRRC powerFRRC KRRFC powerRRFC
#> 1   170 0.8016186   119 0.8001346    99 0.8020196

# \donttest{ 
for (J in 6:10) {
 ret <- SsSampleSizeKGivenJ(dataset02, FOM = "Wilcoxon", J = J, option = "RRRC")
 message("# of readers = ", J, " estimated # of cases = ", ret$K,
 ", predicted power = ", signif(ret$powerRRRC,3), "\n")
}

#> # of readers = 6 estimated # of cases = 251, predicted power = 0.801

#> # of readers = 7 estimated # of cases = 211, predicted power = 0.801

#> # of readers = 8 estimated # of cases = 188, predicted power = 0.801

#> # of readers = 9 estimated # of cases = 173, predicted power = 0.801

#> # of readers = 10 estimated # of cases = 163, predicted power = 0.802

# }