Performs DBM or OR significance testing for factorial or crossed modality datasets

Performs DBM or OR significance testing for specified dataset.

St(
  dataset,
  FOM,
  method = "OR",
  covEstMethod = "jackknife",
  analysisOption = "RRRC",
  alpha = 0.05,
  FPFValue = 0.2,
  nBoots = 200,
  seed = NULL
)

Arguments

dataset

The dataset to be analyzed, see RJafroc-package. The dataset design can be "FCTRL" or "FCTRL-X-MOD".

FOM

The figure of merit, see UtilFigureOfMerit

method

The significance testing method to be used: "DBM" for the Dorfman-Berbaum-Metz method or "OR", for the Obuchowski-Rockette method (default).

covEstMethod

The covariance matrix estimation method in ORH analysis (for method = "DBM" the jackknife is always used).

"Jackknife", the default,

"Bootstrap", in which case nBoots is relevant, default is 200,

"DeLong"; requires FOM = "Wilcoxon" or "ROI" or "HrAuc".

analysisOption

Determines which factors are regarded as random and which are fixed:

"RRRC" = random-reader random case, the default,

"FRRC" = fixed-reader random case,

"RRFC" = random-reader fixed case,

"ALL" = all 3 allowed options.

alpha

The significance level (alpha) of the test of the null hypothesis that all modality effects are zero; the default is 0.05

FPFValue

Only needed for LROC data and FOM = "PCL" or "ALROC"; where to evaluate a partial curve based figure of merit. The default is 0.2.

nBoots

The number of bootstraps (defaults to 200), only needed if covEstMethod = "bootstrap" and method = "OR"

seed

For bootstraps the seed of the RNG (defaults to NULL), only needed if covEstMethod = "bootstrap" and method = "OR"

Value

For method = "DBM" the returned list contains 4 dataframes:

FOMs

Contains foms, trtMeans and trtMeanDiffs: see return of UtilFigureOfMerit

ANOVA

Contains TRCAnova, VarCom, IndividualTrt and IndividualRdr ANOVA tables of pseudovalues

RRRC

Contains results of "RRRC" analyses: FTests, ciDiffTrt, ciAvgRdrEachTrt

FRRC

Contains results of "FRRC" analyses: FTests, ciDiffTrt, ciAvgRdrEachTrt, ciDiffTrtEachRdr

RRFC

Contains results of "RRFC" analyses: FTests, ciDiffTrt, ciAvgRdrEachTrt

For method = "OR" the return list contains 4 dataframes:

FOMs

Contains foms, trtMeans and trtMeanDiffs: UtilFigureOfMerit

ANOVA

Contains TRAnova, VarCom, IndividualTrt and IndividualRdr ANOVA tables of FOM values

RRRC

Contains results of "RRRC" analyses - same organization as DBM, see above

FRRC

Contains results of "FRRC" analyses - ditto

RRFC

Contains results of "RRFC" analyses- ditto

References

Dorfman DD, Berbaum KS, Metz CE (1992) ROC characteristic rating analysis: Generalization to the Population of Readers and Patients with the Jackknife method, Invest. Radiol. 27, 723-731.

Obuchowski NA, Rockette HE (1995) Hypothesis Testing of the Diagnostic Accuracy for Multiple Diagnostic Tests: An ANOVA Approach with Dependent Observations, Communications in Statistics: Simulation and Computation 24, 285-308.

Hillis SL (2014) A marginal-mean ANOVA approach for analyzing multireader multicase radiological imaging data, Statistics in medicine 33, 330-360.

Thompson JD, Chakraborty DP, Szczepura K, et al. (2016) Effect of reconstruction methods and x-ray tube current-time product on nodule detection in an anthropomorphic thorax phantom: a crossed-modality JAFROC observer study. Medical Physics. 43(3):1265-1274.

Chakraborty DP (2017) Observer Performance Methods for Diagnostic Imaging - Foundations, Modeling, and Applications with R-Based Examples, CRC Press, Boca Raton, FL. https://www.routledge.com/Observer-Performance-Methods-for-Diagnostic-Imaging-Foundations-Modeling/Chakraborty/p/book/9781482214840

Examples

result <- St(dataset02,FOM = "Wilcoxon", method = "DBM") 
result <- St(dataset02,FOM = "Wilcoxon", method = "OR")
result <- St(datasetX, FOM = "wAFROC", method = "OR", analysisOption = "ALL")

# \donttest{
result <- St(dataset05, FOM = "wAFROC")
result <- St(dataset05, FOM = "HrAuc", method = "DBM") 
# }