Performs OR significance testing for SPLIT-PLOT A or C datasets

Performs Obuchowski-Rockette (OR) significance testing for specified dataset.

StSP(dataset, FOM, alpha = 0.05, analysisOption = "RRRC")

Arguments

dataset

The dataset to be analyzed, see RJafroc-package. Must have two or more treatments and two or more readers. The dataset design must be "SPLIT-PLOT-A" or "SPLIT-PLOT-C".

FOM

The figure of merit

alpha

The significance level of the test, default is 0.05

analysisOption

Determines which factors are regarded as random vs. fixed:

"RRRC" = random-reader random case, the default,

"FRRC" = fixed-reader random case,

"RRFC" = random-reader fixed case,

Value

Results of the analysis

Examples

fileName <- system.file("extdata", "/toyFiles/ROC/rocSpAZP.xlsx", 
package = "RJafroc", mustWork = TRUE)
dsSpA <- DfReadSP(fileName)
ret <- StSP(dsSpA, FOM = "Wilcoxon")

ret <- StSP(datasetFROCSpC, FOM = "wAFROC")