Perform significance testing, DBMH or ORH

Performs Dorfman-Berbaum-Metz (DBM) or Obuchowski-Rockette (OR) significance testing with Hillis' improvements, for specified dataset; significance testing refers to analysis designed to assign a P-value for rejecting a null hypothesis (NH); the most common NH is that the reader-averaged figure of merit (FOM) difference between treatments is zero. The results of the analysis are better visualized in the text or, preferably, Excel-formatted, files produced by UtilOutputReport.

StSignificanceTesting (dataset, FOM = "wJAFROC", alpha = 0.05,
   method = "DBMH", covEstMethod = "Jackknife", nBoots = 200, option = "ALL",
   VarCompFlag = FALSE, FPFValue = 0.2)

Arguments

dataset
The dataset to be analyzed, see RJafroc-package

FOM
The figure of merit, default "wJAFROC", see UtilFigureOfMerit

alpha
The significance level of the test of the null hypothesis that all treatment effects are zero; the default alpha is 0.05

method
The significance testing method to be used. There are two options: "DBMH" (the default) or "ORH", representing the Dorfman-Berbaum-Metz and the Obuchowski-Rockette significance testing methods, respectively.

covEstMethod
The method used to estimate the covariance matrix in ORH analysis; it can be "Jackknife", "Bootstrap" or "DeLong", the last assumes FOM = "Wilcoxon", otherwise an error results. This parameter is not relevant if the analysis method is "DBMH"

nBoots
The number of bootstraps (default is 200), relevant only if the "Bootstrap" method is used to estimate the covariance matrix in the ORH method

option
Determines which factors are regarded as random vs. fixed: "RRRC" = random-reader random case, "FRRC" = fixed-reader random case, "RRFC" = random-reader fixed case, "ALL" outputs the results of "RRRC", "FRRC" and "RRFC" analyses

VarCompFlag
If TRUE, only the appropriate (DBM or OR) variance components (six in all) are returned, default is FALSE

FPFValue
Only needed for LROC data; where to evaluate a partial curve based figure of merit. The default is 0.2.

Value

For method = "DBMH" the returned value is a list with 22 members:

fomArray

The figure of merit array for each treatment-reader combination

anovaY

The ANOVA table of the pseudovalues over all treatments

anovaYi

The ANOVA table of the pseudovalues for each treatment

varComp

The variance components of the pseudovalue model underlying the analysis, 6 values, in the following order: c("Var(R)", "Var(C)", "Var(T*R)", "Var(T*C)", "Var(R*C)", "Var(Error)")

fRRRC

For random-reader random-case (RRRC) analysis, the F-statistic for rejecting the null hypothesis of no treatment effect

ddfRRRC

For RRRC analysis, the denominator degrees of freedom of the F statistic

pRRRC

For RRRC analysis, the p-value of the significance test of the NH

ciDiffTrtRRRC

For RRRC analysis, the confidence intervals and related test statistics for the FOM differences between pairs of treatments

ciAvgRdrEachTrtRRRC

For RRRC analysis, the confidence intervals and related test statistics for rdr. avg. FOM in each treatment

fFRRC

For fixed-reader random-case (FRRC) analysis, the F-statistic for rejecting the NH

ddfFRRC

For FRRC analysis, the denominator degrees of freedom of the F-statistic

pFRRC

For FRRC analysis, the p-value of the significance test of the NH

ciDiffTrtFRRC

For FRRC analysis, the confidence intervals and related test statistics for the FOM differences between pairs of treatments

ciAvgRdrEachTrtFRRC

For FRRC analysis, the confidence intervals and related tests for rdr. avg. FOM in each treatment

ssAnovaEachRdr

The sum of squares table of the ANOVA of the pseudovalues for each reader (based on data for the specified reader)

msAnovaEachRdr

The mean squares table of the ANOVA of the pseudovalues for each reader (based on data for the specified reader)

ciDiffTrtEachRdr

The confidence intervals and related tests of the FOM differences between pairs of treatments for each reader

fRRFC

For random-reader fixed-case (RRFC) analysis, the F statistic

ddfRRFC

For RRFC analysis, the denominator degrees of freedom of the F statistic

pRRFC

For RRFC analysis, the p-value for rejecting the NH

ciDiffTrtRRFC

For RRFC analysis, the confidence intervals and related test statistics for the FOM differences between pairs of treatments

ciAvgRdrEachTrtRRFC

For RRFC analysis, the confidence intervals and related tests for reader averaged FOM in each treatment

For method = "ORH" the return value is a list with with 21 members:
fomArray

Figures of merit array. See the return of UtilFigureOfMerit

msT

Mean square of the figure of merit corresponding to the treatment effect

msTR

Mean square of the figure of merit corresponding to the treatment-reader effect

varComp

The variance components of the pseudovalue model underlying the analysis, 6 values, in the following order: c("Var(R)", "Var(T*R)", "COV1", "COV2", "COV3", "Var(Error)")

fRRRC

Same as DBMH method

ddfRRRC

Same as DBMH method

pRRRC

Same as DBMH method

ciDiffTrtRRRC

Same as DBMH method

ciAvgRdrEachTrtRRRC

Same as DBMH method

fFRRC

Same as DBMH method

ddfFRRC

Same as DBMH method

pFRRC

Same as DBMH method

ciDiffTrtFRRC

Same as DBMH method

ciAvgRdrEachTrtFRRC

Same as DBMH method

ciDiffTrtEachRdr

Same as DBMH method

varCovEachRdr

Obuchowski-Rockette Variance and Cov1 estimates for each reader

fRRFC

Same as DBMH method

ddfRRFC

Same as DBMH method

pRRFC

Same as DBMH method

ciDiffTrtRRFC

Same as DBMH method

ciAvgRdrEachTrtRRFC

Same as DBMH method

References

Dorfman DD, Berbaum KS, Metz CE (1992) ROC characteristic rating analysis: Generalization to the Population of Readers and Patients with the Jackknife method, Invest. Radiol. 27, 723-731.

Obuchowski NA, Rockette HE (1995) Hypothesis Testing of the Diagnostic Accuracy for Multiple Diagnostic Tests: An ANOVA Approach with Dependent Observations, Communications in Statistics: Simulation and Computation 24, 285-308.

Hillis SL (2014) A marginal-mean ANOVA approach for analyzing multireader multicase radiological imaging data, Statistics in medicine 33, 330-360.

Chakraborty DP (2017) Observer Performance Methods for Diagnostic Imaging - Foundations, Modeling, and Applications with R-Based Examples, CRC Press, Boca Raton, FL. https://www.crcpress.com/Observer-Performance-Methods-for-Diagnostic-Imaging-Foundations-Modeling/Chakraborty/p/book/9781482214840

Examples

retDbmRoc  <- StSignificanceTesting(dataset02,
FOM = "Wilcoxon", method = "DBMH")
# NOT RUN {
retDbmwJAFROC  <- StSignificanceTesting(dataset05) # default is weighted JAFROC

retDbmHrAuc  <- StSignificanceTesting(dataset05,
FOM = "HrAuc", method = "DBMH")
print(retDbmHrAuc)

retDbmSongA1  <- StSignificanceTesting(dataset05,
FOM = "SongA1", method = "DBMH")
print(retDbmSongA1)

retDbmSongA2  <- StSignificanceTesting(dataset05,
FOM = "SongA2", method = "DBMH")
print(retDbmSongA2)

retDbmwJafroc1  <- StSignificanceTesting(dataset05,
FOM = "wJAFROC1", method = "DBMH")
print(retDbmwJafroc1)

retDbmJafroc1  <- StSignificanceTesting(dataset05,
FOM = "JAFROC1", method = "DBMH")
print(retDbmJafroc1)

retDbmJAFROC  <- StSignificanceTesting(dataset05,
FOM = "JAFROC", method = "DBMH")
print(retDbmJAFROC)

# }

retOR <- StSignificanceTesting(dataset02,
FOM = "Wilcoxon", method = "ORH")
print(retOR)

#> $fomArray
#>           Rdr - 0   Rdr - 1   Rdr - 2   Rdr - 3   Rdr - 4
#> Trt - 0 0.9196457 0.8587762 0.9038647 0.9731079 0.8297907
#> Trt - 1 0.9478261 0.9053140 0.9217391 0.9993559 0.9299517
#> 
#> $msT
#> [1] 0.004796171
#> 
#> $msTR
#> [1] 0.0005510306
#> 
#> $varComp
#>                  varCov
#> Var(R)     0.0015349993
#> Var(T*R)   0.0002004025
#> COV1       0.0003466137
#> COV2       0.0003440748
#> COV3       0.0002390284
#> Var(Error) 0.0008022883
#> 
#> $fRRRC
#> [1] 4.456319
#> 
#> $ndf
#> [1] 1
#> 
#> $ddfRRRC
#> [1] 15.25967
#> 
#> $pRRRC
#> [1] 0.05166569
#> 
#> $ciDiffTrtRRRC
#>   Treatment    Estimate     StdErr       DF         t     Pr > t   CI Lower
#> 1     0 - 1 -0.04380032 0.02074862 15.25967 -2.110999 0.05166569 -0.0879595
#>       CI Upper
#> 1 0.0003588544
#> 
#> $ciAvgRdrEachTrtRRRC
#>   Treatment      Area     StdErr       DF  CI Lower  CI Upper
#> 1         0 0.8970370 0.03317360 12.74465 0.8252236 0.9688505
#> 2         1 0.9408374 0.02156637 12.71019 0.8941378 0.9875369
#> 
#> $fFRRC
#> [1] 5.475953
#> 
#> $ddfFRRC
#> [1] Inf
#> 
#> $pFRRC
#> [1] 0.01927984
#> 
#> $ciDiffTrtFRRC
#>   Treatment    Estimate     StdErr  DF         t     Pr > t    CI Lower
#> 1     0 - 1 -0.04380032 0.01871748 Inf -2.340075 0.01927984 -0.08048591
#>      CI Upper
#> 1 -0.00711473
#> 
#> $ciAvgRdrEachTrtFRRC
#>   Treatment      Area     StdErr  DF  CI Lower  CI Upper
#> 1         0 0.8970370 0.02428971 Inf 0.8494301 0.9446440
#> 2         1 0.9408374 0.01677632 Inf 0.9079564 0.9737183
#> 
#> $ciDiffTrtEachRdr
#>   Reader Treatment    Estimate     StdErr  DF          t     Pr > t    CI Lower
#> 1      0     0 - 1 -0.02818035 0.02551213 Inf -1.1045864 0.26933885 -0.07818322
#> 2      1     0 - 1 -0.04653784 0.02630183 Inf -1.7693768 0.07683102 -0.09808848
#> 3      2     0 - 1 -0.01787440 0.03120965 Inf -0.5727202 0.56683414 -0.07904418
#> 4      3     0 - 1 -0.02624799 0.01729129 Inf -1.5179891 0.12901715 -0.06013829
#> 5      4     0 - 1 -0.10016103 0.04405746 Inf -2.2734182 0.02300099 -0.18651207
#>       CI Upper
#> 1  0.021822507
#> 2  0.005012792
#> 3  0.043295388
#> 4  0.007642316
#> 5 -0.013809995
#> 
#> $varCovEachRdr
#>   Reader          Var         Cov1
#> 0      0 0.0006989006 3.734661e-04
#> 1      1 0.0011060528 7.601598e-04
#> 2      2 0.0008423434 3.553224e-04
#> 3      3 0.0001505777 1.083399e-06
#> 4      4 0.0012135668 2.430368e-04
#> 
#> $fRRFC
#> [1] 8.704
#> 
#> $ddfRRFC
#> [1] 4
#> 
#> $pRRFC
#> [1] 0.04195875
#> 
#> $ciDiffTrtRRFC
#>   Treatment    Estimate     StdErr DF         t     Pr > t    CI Lower
#> 1     0 - 1 -0.04380032 0.01484629  4 -2.950254 0.04195875 -0.08502022
#>      CI Upper
#> 1 -0.00258042
#> 
#> $ciAvgRdrEachTrtRRFC
#>   Treatment      Area     StdErr DF  CI Lower  CI Upper
#> 1         0 0.8970370 0.02482994  4 0.8280981 0.9659760
#> 2         1 0.9408374 0.01615303  4 0.8959894 0.9856854
#>