Perform significance testing for single fixed factor analysis

Significance testing for datasets with a single reader in multiple (at least two) treatments, or a single treatment with multiple (at least two) readers, where reader or treatment, respectively, is regarded as a fixed factor and a common case-set, regarded as random, is assumed.

StSignificanceTestingSingleFixedFactor(
  dataset,
  FOM,
  FPFValue = 0.2,
  alpha = 0.05
)

Arguments

dataset
A single-treatment or single-reader dataset.

FOM
The figure of merit, see UtilFigureOfMerit.

FPFValue
Only needed for LROC data and FOM = "PCL" or "ALROC"; at which FPF to evaluate a partial curve based figure of merit. The default is FPF = 0.2.

alpha
The significance level (alpha, default 0.05) of the test of the null hypothesis that FOMs of all levels of the fixed factor are identical.

Value

The return value is a list containing:

f

The observed F-statistic for testing the null hypothesis of no treatment effect.

ddf

The denominator degrees of freedom of the F statistic. The numerator degrees of freedom is always the number of levels of the fixed factor minus one.

pValue

The p-value for rejecting the NH.

fomStats

Statistics for FOM for each level of the fixed factor.

diffFomStats

Statistics for FOM-differences for all distinct pairings of the levels of the fixed factor

Details

This function performs implements Hillis et al. 2005, Eqn. 23. Following an overall F-test, reader-pairings are compared using paired t-tests. In order for a specific pairing to be declared significant, the F-test must also be significant.

References

Hillis SL, Obuchowski NA, Schartz KM, Berbaum KS (2005) A comparison of the Dorfman-Berbaum-Metz and Obuchowski-Rockette methods for receiver operating characteristic (ROC) data, Statistics in Medicine, 24(10), 1579-607.

Hillis SL (2007) A comparison of denominator degrees of freedom methods for multiple observer ROC studies, Statistics in Medicine. 26:596-619.

Hillis SL (2014) A marginal mean ANOVA approach for analyzing multireader multicase radiological imaging data, Statistics in medicine 33, 330-360.

Examples

## Create a single treatment ROC dataset with one treatment and four readers
singleFactorData <- DfExtractDataset(dataset02, 1, 1:4)

## Performs single treatment fixed reader analysis
StSignificanceTestingSingleFixedFactor(singleFactorData, FOM = "Wilcoxon")

#> $f
#> [1] 5.20948
#> 
#> $ddf
#> [1] 339
#> 
#> $pValue
#> [1] 0.001576444
#> 
#> $fomStats
#>   Reader              Area             stdErr  DF           CILower
#> 1      0 0.919645732689211 0.0301255163700399 113 0.859961651339663
#> 2      1  0.85877616747182 0.0363753334990573 113 0.786710071176329
#> 3      2 0.903864734299517 0.0282594117842767 113 0.847877742703897
#> 4      3 0.973107890499195  0.017338833155504 113 0.938756534796343
#>             CIUpper
#> 1 0.979329814038759
#> 2  0.93084226376731
#> 3 0.959851725895137
#> 4  1.00745924620205
#> 
#> $diffFomStats
#>   Reader  Difference          t  DF        PrGTt      CILower       CIUpper
#> 1  0 - 1  0.06086957  2.0545714 113 0.0422268728  0.002174313  0.1195648174
#> 2  0 - 2  0.01578100  0.6104293 113 0.5428032056 -0.035437058  0.0669990546
#> 3  0 - 3 -0.05346216 -1.9761947 113 0.0505695484 -0.107059194  0.0001348783
#> 4  1 - 2 -0.04508857 -1.4867784 113 0.1398583124 -0.105170543  0.0149934093
#> 5  1 - 3 -0.11433172 -3.4526967 113 0.0007819153 -0.179936026 -0.0487274197
#> 6  2 - 3 -0.06924316 -2.3985979 113 0.0180950060 -0.126436228 -0.0120500848
#>