Significance testing for single random factor

Significance testing for datasets with multiple readers in a single treatment, compare average FOM to specified NH value.

StSingleTreatmentRandomReader(dataset, fomNh, FOM, FPFValue = 0.2,
  alpha = 0.05)

Arguments

dataset
A single-treatment multipe-reader dataset.

fomNh
The comparison value that the reader average FOM is compared to.

FOM
The figure of merit, see UtilFigureOfMerit.

FPFValue
Only needed for LROC data and FOM = "PCL" or "ALROC"; where to evaluate a partial curve based figure of merit. The default is 0.2.

alpha
The significance level (alpha, default 0.05) of the test of the null hypothesis that the reader averaged FOMs and the specified NH value fomNh are identical.

Value

The return value is a list containing:

fom Observed reader FOMs.

avgFom Average reader FOM.

CIAvgFom Confidence interval of the reader averaged FOM.

vaR Reader variance term of the Obuchowski-Rockette model.

cov2 cov2 of the Obuchowski-Rockette model.

var Error term of the Obuchowski-Rockette model.

Tstat The observed value of the t-statistic.

df The degrees of freedom associated with the t-statistic.

pValue The p-value for rejecting the NH.

Details

This function performs implements Hillis et al. 2005, Eqn. 23.

References

Hillis SL, Obuchowski NA, Schartz KM, Berbaum KS (2005) A comparison of the Dorfman-Berbaum-Metz and Obuchowski-Rockette methods for receiver operating characteristic (ROC) data, Statistics in Medicine, 24(10), 1579-607.

Hillis SL (2007) A comparison of denominator degrees of freedom methods for multiple observer ROC studies, Statistics in Medicine. 26:596-619.

Hillis SL (2014) A marginal mean ANOVA approach for analyzing multireader multicase radiological imaging data, Statistics in medicine 33, 330-360.

Examples

## Create a single treatment ROC dataset with four readers
singleFactorData <- DfExtractDataset(dataset02, trts = 1, rdrs = seq(1,4))

## Perform single-treatment random-reader analysis
StSingleTreatmentRandomReader(singleFactorData, fomNh = 0.8, FOM = "Wilcoxon")

#> $fom
#>           Rdr0      Rdr1      Rdr2      Rdr3
#> Trt0 0.9196457 0.8587762 0.9038647 0.9731079
#> 
#> $avgFom
#> [1] 0.9138486
#> 
#> $CIAvgFom
#> [1] 0.8436480 0.9840493
#> 
#> $varR
#> [1] 0.002225974
#> 
#> $cov2
#> [1] 0.0004051924
#> 
#> $var
#> [1] 0.0008324853
#> 
#> $Tstat
#> [1] 3.671227
#> 
#> $df
#> [1] 8.959168
#> 
#> $pval
#> [1] 0.005185003
#>