RSM ROC/AFROC/wAFROC AUC calculator

Returns the ROC, AFROC and wAFROC AUCs corresponding to specified RSM parameters. See also UtilAucPROPROC, UtilAucBIN and UtilAucCBM

UtilAnalyticalAucsRSM(mu, lambda, nu, zeta1 = -Inf, lesDistr, relWeights = 0)

Arguments

mu

The mean of the Gaussian distribution for the ratings of latent LLs (continuous ratings of lesions that are found by the search mechanism). The NLs are assumed to be distributed as N(0,1).

lambda

The RSM lambda parameter.

nu

The RSM nu parameters.

zeta1

The lowest reporting threshold, the default is -Inf.

lesDistr

The lesion distribution 1D array, i.e., the probability mass function (pmf) of the numbers of lesions for diseased cases.

relWeights

The relative weights of the lesions; a vector of length maxLL; if zero, the default, equal weights are assumed.

Value

The ROC, AFROC and wAFROC AUCs corresponding to the specified parameters

References

Chakraborty DP (2017) Observer Performance Methods for Diagnostic Imaging - Foundations, Modeling, and Applications with R-Based Examples, CRC Press, Boca Raton, FL. https://www.routledge.com/Observer-Performance-Methods-for-Diagnostic-Imaging-Foundations-Modeling/Chakraborty/p/book/9781482214840

Chakraborty DP (2006) A search model and figure of merit for observer data acquired according to the free-response paradigm, Phys Med Biol 51, 3449-3462.

Chakraborty DP (2006) ROC Curves predicted by a model of visual search, Phys Med Biol 51, 3463–3482.

Examples

mu <- 1;lambda <- 1;nu <- 0.9
lesDistr <- c(0.9, 0.1) 
## i.e., 90% of dis. cases have one lesion, and 10% have two lesions
relWeights <- c(0.05, 0.95)
## i.e., lesion 1 has weight 5 percent while lesion two has weight 95 percent

UtilAnalyticalAucsRSM(mu, lambda, nu, zeta1 = -Inf, lesDistr)
#> $aucROC
#> [1] 0.8147209
#> 
#> $aucAFROC
#> [1] 0.7448364
#> 
#> $aucwAFROC
#> [1] 0.7448364
#> 
UtilAnalyticalAucsRSM(mu, lambda, nu, zeta1 = -Inf, lesDistr, relWeights)
#> $aucROC
#> [1] 0.8147209
#> 
#> $aucAFROC
#> [1] 0.7448364
#> 
#> $aucwAFROC
#> [1] 0.785221
#>