RSM ROC/AFROC AUC calculator

Returns the ROC and AFROC AUCs corresponding to specified RSM parameters. See also UtilAucPROPROC, UtilAucBinormal and UtilAucCBM

UtilAucsRSM(mu, lambdaP, nuP, lesDistr)

Arguments

mu
The mean(s) of the Gaussian distribution(s) for the ratings of latent LLs (continuous ratings of lesions that are found by the search mechanism)

lambdaP
The Poisson distribution parameter(s), which describes the random number of latent NLs (suspicious regions that do not correspond to actual lesions) per case; these are the physical parameters.

nuP
The physical nuP parameters, each of which is the success probability of the binomial distribution(s) describing the random number of latent LLs (suspicious regions that correspond to actual lesions) per diseased case.

lesDistr
See PlotRsmOperatingCharacteristics.

Value

A list containing the ROC and AFROC AUCs corresponding to the specified parameters

Details

The RSM parameters (mu, lambdaP and nuP) can be vectors, provided they are of the same length; the first parameter of each array is used, followed by the second, etc; a common lesion distribution is assumed.

References

Chakraborty DP (2017) Observer Performance Methods for Diagnostic Imaging - Foundations, Modeling, and Applications with R-Based Examples, CRC Press, Boca Raton, FL. https://www.crcpress.com/Observer-Performance-Methods-for-Diagnostic-Imaging-Foundations-Modeling/Chakraborty/p/book/9781482214840

Chakraborty DP (2006) A search model and figure of merit for observer data acquired according to the free-response paradigm, Phys Med Biol 51, 3449-3462.

Chakraborty DP (2006) ROC Curves predicted by a model of visual search, Phys Med Biol 51, 3463--3482.

Examples

mu <- 1;lambdaP <- 1;nuP <- 1
lesDistr <- rbind(c(1, 0.9), c(2, 0.1)) 
## i.e., 90% of dis. cases have one lesion, and 10% have two lesions
UtilAucsRSM(mu, lambdaP, nuP, lesDistr)$aucROC

#> [1] 0.84700653

UtilAucsRSM(mu, lambdaP, nuP, lesDistr)$aucAFROC

#> [1] 0.80715827

mu <- c(1,2);lambdaP <- c(1,0.5);nuP <- c(1, 0.8)
lesDistr <- rbind(c(1, 0.9), c(2, 0.1))
## i.e., 90% of dis. cases have one lesion, and 10% have two lesions
UtilAucsRSM(mu, lambdaP, nuP, lesDistr)$aucROC

#> [1] 0.84700653 0.88149316

UtilAucsRSM(mu, lambdaP, nuP, lesDistr)$aucAFROC

#> [1] 0.80715827 0.83133709