Calculate empirical figures of merit (FOMs) for specified dataset

Calculate the specified empirical figure of merit for each treatment-reader combination in the ROC, FROC or LROC dataset

UtilFigureOfMerit(dataset, FOM = "wAFROC", FPFValue = 0.2)

Arguments

dataset
The dataset to be analyzed, see RJafroc-package

FOM
The figure of merit to be used in the calculation. The default is "wAFROC"

FPFValue
Only needed for LROC data; where to evaluate a partial curve based figure of merit. The default is 0.2.

Value

An c(I, J) array, where the row names are the IDs of the treatments and column names are the IDs of the readers.

Details

The allowed FOMs depend on the type of dataset (i.e., dataType field of dataset object). For ROC datasets: the "Wilcoxon" is allowed. For FROC datasets: The following FOMs are allowed: "AFROC1", "AFROC", "wAFROC1", "wAFROC" (the default), "HrAuc", "SongA1", "SongA2", "HrSe", "HrSp", "MaxLLF", "MaxNLF", "MaxNLFAllCases", and "ExpTrnsfmSp". The "MaxLLF", "MaxNLF" and "MaxNLFAllCases" FOMs correspond to ordinate, and abscissa, respectively, of the highest point on the FROC operating characteristic obtained by counting all the marks). The "ExpTrnsfmSp" FOM is described in the paper by Popescu. Given the large number of FOMs possible with FROC data, it is appropriate to make a recommendation: it is recommended that one use the wAFROC FOM. For LROC datasets: The following FOMs are allowed:"Wilcoxon" for ROC data inferred from LROC data, which ignores localization information; or "PCL" or "ALROC", in which case one needs to specify an additional argument, FPFValue: the desired FPF at which to evaluate PCL or ALROC; the default is 0.2.

References

Chakraborty DP (2017) Observer Performance Methods for Diagnostic Imaging - Foundations, Modeling, and Applications with R-Based Examples, CRC Press, Boca Raton, FL. https://www.crcpress.com/Observer-Performance-Methods-for-Diagnostic-Imaging-Foundations-Modeling/Chakraborty/p/book/9781482214840

Chakraborty DP, Berbaum KS (2004) Observer studies involving detection and localization: modeling, analysis, and validation, Medical Physics, 31(8), 1--18.

Song T, Bandos AI, Rockette HE, Gur D (2008) On comparing methods for discriminating between actually negative and actually positive subjects with FROC type data, Medical Physics 35 1547--1558.

Popescu LM (2011) Nonparametric signal detectability evaluation using an exponential transformation of the FROC curve, Medical Physics, 38(10), 5690.

Obuchowski NA, Lieber ML, Powell KA (2000) Data Analysis for Detection and Localization of Multiple Abnormalities with Application to Mammography, Acad Radiol, 7:7 553--554.

Swensson RG (1996) Unified measurement of observer performance in detecting and localizing target objects on images, Med Phys 23:10, 1709--1725.

Examples

# ROC data
UtilFigureOfMerit(dataset = dataset02, FOM = "Wilcoxon")

#>           Rdr - 0   Rdr - 1   Rdr - 2   Rdr - 3   Rdr - 4
#> Trt - 0 0.9196457 0.8587762 0.9038647 0.9731079 0.8297907
#> Trt - 1 0.9478261 0.9053140 0.9217391 0.9993559 0.9299517

# FROC dataset, converted to ROC, Wilcoxon FOM
UtilFigureOfMerit(DfFroc2Roc(dataset01), FOM = "Wilcoxon")

#>            Rdr - 1   Rdr - 2   Rdr - 3   Rdr - 4   Rdr - 5
#> Trt - BT 0.7959972 0.8652271 0.8411751 0.8311096 0.8465590
#> Trt - DM 0.6689490 0.7211493 0.7841175 0.7979284 0.7061096

# FROC dataset, default wAFROC FOM
UtilFigureOfMerit(dataset = dataset01)

#>            Rdr - 1   Rdr - 2   Rdr - 3   Rdr - 4   Rdr - 5
#> Trt - BT 0.7602704 0.8406191 0.8171524 0.8153090 0.8278324
#> Trt - DM 0.6425854 0.7049977 0.7518434 0.7724426 0.6836962

#LROC data
UtilFigureOfMerit(dataset = datasetCadLroc, FOM = "ALROC", FPFValue = 0.2)

#>           Rdr - 1   Rdr - 2   Rdr - 3   Rdr - 4   Rdr - 5   Rdr - 6   Rdr - 7
#> Trt - 1 0.1005382 0.1075944 0.1089236 0.1447396 0.1189063 0.1000335 0.1345511
#>           Rdr - 8   Rdr - 9  Rdr - 10
#> Trt - 1 0.1222135 0.1060417 0.1054167