Get the lesion distribution vector of a dataset

The lesion distribution vector for a dataset.

UtilLesionDistrVector(dataset)

Arguments

dataset

A dataset object.

Value

lesDistr The 1D lesion distribution array.

Details

Two characteristics of an FROC dataset, apart from ratings, affect the FOM: the distribution of lesions per case and the distribution of lesion weights. This function addresses the distribution of lesions per case. The distribution of weights is addressed in UtilLesionWeightsMatrix. lesDistr is a 1D-array containing the fraction of unique values of lesions per diseased case in the dataset. For ROC or LROC data this vector is c(1), since all diseased cases contain one lesion. For FROC data the length of the vector equals the maximum number of lesions per diseased case. The first entry is the fraction of dis. cases containing one lesion, the second entry is the fraction of dis. cases containing two lesions, etc. See PlotRsmOperatingCharacteristics for a function that depends on lesDistr.

Examples

UtilLesionDistrVector (dataset01) # FROC dataset ## [1] 0.93258427 0.06741573
#> [1] 0.93258427 0.06741573
UtilLesionDistrVector (dataset02) # ROC dataset  ## 1
#> [1] 1