Calculate mean squares for factorial dataset

Calculates the mean squares used in the DBM and ORH methods for factorial dataset

UtilMeanSquares(dataset, FOM = "Wilcoxon", FPFValue = 0.2, method = "DBM")

Arguments

dataset
The dataset to be analyzed, see RJafroc-package.

FOM
The figure of merit to be used in the calculation. The default is "FOM_wAFROC". See UtilFigureOfMerit.

FPFValue
Only needed for LROC data and FOM = "PCL" or "ALROC"; where to evaluate a partial curve based figure of merit. The default is 0.2.

method
The method, in which the mean squares are calculated. The two valid choices are "DBM" (default) and "OR".

Value

A list containing all possible mean squares

Details

For DBM method, msT, msTR, msTC, msTRC will not be available if the dataset contains only one treatment. Similarly, msR, msTR, msRC, msTRC will not be returned for single reader dataset. For ORH method, msT, msR, msTR will be returned for multiple reader multiple treatment dataset. msT is not available for single treatment dataset, and msR is not available for single reader dataset.

Examples

UtilMeanSquares(dataset02, FOM = "Wilcoxon")

#> $msT
#> [1] 0.54676344
#> 
#> $msR
#> [1] 0.4373268
#> 
#> $msC
#> [1] 0.39686988
#> 
#> $msTR
#> [1] 0.062817491
#> 
#> $msTC
#> [1] 0.099848084
#> 
#> $msRC
#> [1] 0.06450106
#> 
#> $msTRC
#> [1] 0.039971603
#> 
#> $msCSingleT
#> [1] 0.33629430 0.16042367
#> 
#> $msCSingleR
#> [1] 0.122249800 0.212748238 0.136533908 0.017289369 0.166052811
#> 

UtilMeanSquares(dataset05, FOM = "wAFROC", method = "OR")

#> $msT
#> [1] 0.027397181
#> 
#> $msR
#> [1] 0.018337407
#> 
#> $msTR
#> [1] 0.00092775374
#>