Utility for estimating Obuchowski-Rockette variance components for factorial datasets

Utility for estimating Obuchowski-Rockette variance components for factorial datasets

UtilORVarComponentsFactorial(
  dataset,
  FOM,
  FPFValue = 0.2,
  covEstMethod = "jackknife",
  nBoots = 200,
  seed = NULL
)

Arguments

dataset

The factorial dataset object

FOM

The figure of merit

FPFValue

Only needed for LROC data and FOM = "PCL" or "ALROC"; where to evaluate a partial curve based figure of merit. The default is 0.2.

covEstMethod

The covariance estimation method, "jackknife" (the default) or "bootstrap" or "DeLong" (DeLongt is applicable only for FOM = Wilcoxon).

nBoots

Only needed for bootstrap covariance estimation method. The number of bootstraps, defaults to 200.

seed

Only needed for the bootstrap covariance estimation method. The initial seed for the random number generator, the default is NULL, as if no seed has been specified.

Value

A list object containing the following data.frames:

foms: the figures of merit for different treatment-reader combinations

TRanova: the OR treatment-reader ANOVA table

VarCom: the OR variance-components Cov1, Cov2, Cov3, Var and correlations rho1, rho2 and rho3

IndividualTrt: the individual treatment mean-squares, Var and Cov2 values

IndividualRdr: the individual reader mean-squares, Var and Cov1 values

Details

The variance components are obtained using StSignificanceTesting with method = "OR".

Examples

## use the default jackknife for covEstMethod
vc <- UtilORVarComponentsFactorial(dataset02, FOM = "Wilcoxon")
str(vc) 
#> List of 4
#>  $ TRanova      :'data.frame':	3 obs. of  3 variables:
#>   ..$ SS: num [1:3] 0.0048 0.0153 0.0022
#>   ..$ DF: num [1:3] 1 4 4
#>   ..$ MS: num [1:3] 0.004796 0.003836 0.000551
#>  $ VarCom       :'data.frame':	6 obs. of  2 variables:
#>   ..$ Estimates: num [1:6] 0.001535 0.0002 0.000347 0.000344 0.000239 ...
#>   ..$ Rhos     : num [1:6] NA NA 0.432 0.429 0.298 ...
#>  $ IndividualTrt:'data.frame':	2 obs. of  4 variables:
#>   ..$ DF         : num [1:2] 4 4
#>   ..$ msREachTrt : num [1:2] 0.00308 0.0013
#>   ..$ varEachTrt : num [1:2] 0.00101 0.00059
#>   ..$ cov2EachTrt: num [1:2] 0.000484 0.000204
#>  $ IndividualRdr:'data.frame':	5 obs. of  4 variables:
#>   ..$ DF         : num [1:5] 1 1 1 1 1
#>   ..$ msTEachRdr : num [1:5] 0.000397 0.001083 0.00016 0.000344 0.005016
#>   ..$ varEachRdr : num [1:5] 0.000699 0.001106 0.000842 0.000151 0.001214
#>   ..$ cov1EachRdr: num [1:5] 3.73e-04 7.60e-04 3.55e-04 1.08e-06 2.43e-04

UtilORVarComponentsFactorial(dataset02, FOM = "Wilcoxon", 
   covEstMethod = "bootstrap", nBoots = 2000, seed = 100)$VarCom 
#>           Estimates       Rhos
#> VarR  0.00155082365         NA
#> VarTR 0.00019451356         NA
#> Cov1  0.00032833333 0.41690698
#> Cov2  0.00033926769 0.43079107
#> Cov3  0.00023657230 0.30039181
#> Var   0.00078754579         NA

UtilORVarComponentsFactorial(dataset02, FOM = "Wilcoxon", covEstMethod = "DeLong")$VarCom 
#>           Estimates       Rhos
#> VarR  0.00153642538         NA
#> VarTR 0.00020458400         NA
#> Cov1  0.00034200896 0.43175729
#> Cov2  0.00033952653 0.42862343
#> Cov3  0.00023584965 0.29774017
#> Var   0.00079213245         NA