Generate a text- or Excel-formatted report file (one-treatment factorial datasets only)

Generates a formatted report resulting from the analysis, for a factorial dataset, and saves it to a text- or an Excel-formatted file.

UtilOutputReport(
  dataset,
  FOM,
  method = "OR",
  covEstMethod = "jackknife",
  analysisOption = "ALL",
  ReportFileExt = "txt",
  ReportFileBaseName = NULL,
  alpha = 0.05,
  nBoots = 200,
  sequentialNames = FALSE,
  overWrite = FALSE
)

Arguments

dataset

The dataset object to be analyzed, not the Excel input file name!, see Dataset in RJafroc-package. This R-object is created by DfReadDataFile.

FOM

The figure of merit; see St.

method

The significance testing method, "DBM" or "OR" (the default).

covEstMethod

The default is "Jackknife"; only needed for method = "OR" when "Bootstrap" or "DeLong" are possible.

analysisOption

"RRRC", "FRRC", "RRFC or "ALL" (the default); see St.

ReportFileExt

The file extension determines the type of output. "txt" (default) for a text file or "xlsx" for an Excel file.

ReportFileBaseName

The report file base name, i.e., without the extension (which is specified by option ReportFileExt)). The output file is created in the user's directory. The default is NULL, see details.

alpha

See St; the default is 0.05.

nBoots

See St; only needed for "OR" analysis; the default is 200.

sequentialNames

A logical variable: if TRUE, consecutive integers (starting from 1) will be used as the modality and reader IDs in the output report. Otherwise, modality and reader IDs in the original dataset are used. This option is needed for aesthetics, as long names can mess up the output. The default is FALSE.

overWrite

A logical variable: if FALSE, a warning will be issued if the report file already exists and the program will wait until the user inputs "y" or "n" to determine whether to overwrite the existing file. If TRUE, the existing file will be silently overwritten. The default is FALSE.

Value

StResult: the object returned by St.

Details

This function is available only for factorial datasets. ReportFileBaseName: The report file base name. The default value NULL generates a temporary text file, whose very long name is displayed. However, the file is very hard to locate. This is so that the package passes CRAN checks, as writing to the project directory, or any of its sub-directories, is frowned upon.

Examples

# \donttest{
 # text output is created in a temporary file
ret <- UtilOutputReport(dataset03, FOM = "Wilcoxon")
 # Excel output is created in a temporary file
ret <- UtilOutputReport(dataset03, FOM = "Wilcoxon", ReportFileExt = "xlsx") 

# }