Utility for Obuchowski-Rockette variance components

Utility for Obuchowski-Rockette variance components

UtilVarComponentsOR(
  dataset,
  FOM,
  FPFValue = 0.2,
  covEstMethod = "jackknife",
  nBoots = 200,
  seed = NULL
)

Arguments

dataset
The dataset object

FOM
The figure of merit

FPFValue
Only needed for LROC data and FOM = "PCL" or "ALROC"; where to evaluate a partial curve based figure of merit. The default is 0.2.

covEstMethod
The covariance estimation method, "jackknife" (the default) or "bootstrap".

nBoots
Only needed for bootstrap covariance estimation method. The number of bootstraps, defaults to 200.

seed
Only needed for the bootstrap covariance estimation method. The initial seed for the random number generator, the default is NULL, as if no seed has been specified.

Value

A list object containing the variance components.

Details

The variance components are obtained using StSignificanceTesting with method = "ORH".

Examples

UtilVarComponentsOR(dataset02, FOM = "Wilcoxon")$varComp

#>          varR        varTR         cov1         cov2         cov3          var
#> 1 0.001534999 0.0002004025 0.0003466137 0.0003440748 0.0002390284 0.0008022883

## uses the default jackknife for covEstMethod

UtilVarComponentsOR(dataset02, FOM = "Wilcoxon",
   covEstMethod = "bootstrap", nBoots = 2000, seed = 100)$varComp

#>          varR        varTR         cov1         cov2         cov3          var
#> 1 0.001550824 0.0001945136 0.0003283333 0.0003392677 0.0002365723 0.0007875458

UtilVarComponentsOR(dataset02, FOM = "Wilcoxon",
   covEstMethod = "DeLong")$varComp

#>          varR       varTR        cov1         cov2         cov3          var
#> 1 0.001536425 0.000204584 0.000342009 0.0003395265 0.0002358497 0.0007921325