Binned dataset suitable for checking FitCorCbm; seed = 123

A binned dataset suitable for analysis by FitCorCbm. It was generated by DfCreateCorCbmDataset by setting the seed variable to 123. Note the formatting of the data as a single treatment two reader dataset, even though the actual pairing might be different, see FitCorCbm. The dataset is intentionally large so as to demonstrate the asymptotic convergence of ML estimates, produced by FitCorCbm, to the population values. The data was generated by the following argument values to DfCreateCorCbmDataset: seed = 123, K1 = 5000, K2 = 5000, desiredNumBins = 5, muX = 1.5, muY = 3, alphaX = 0.4, alphaY = 0.7, rhoNor = 0.3, rhoAbn2 = 0.8.

binnedData123

Format

A list with 8 elements:

NL Ratings array [1, 1:2, 1:10000, 1], of non-lesion localizations, NLs

LL Ratings array [1, 1:2, 1:5000, 1], of lesion localizations, LLs

lesionNum array [1:5000], number of lesions per diseased case, all set to one

lesionID array [1:5000, 1], lesions labels on diseased cases, all set to one

lesionWeight array [1:5000, 1], weights, all set to one

dataType "ROC", the data type

modalityID "1", treatment label

readerID [1:2] "1" "2", reader labels

References

Zhai X, Chakraborty DP (2017). A bivariate contaminated binormal model for robust fitting of proper ROC curves to a pair of correlated, possibly degenerate, ROC datasets. Medical Physics. 44(6):2207--2222.

Examples

str(binnedData123)

#> List of 8
#>  $ NL          : num [1, 1:2, 1:10000, 1] 1 2 1 1 3 1 2 3 1 3 ...
#>  $ LL          : num [1, 1:2, 1:5000, 1] 2 3 2 1 3 2 1 1 2 1 ...
#>  $ lesionNum   : num [1:5000] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ lesionID    : num [1:5000, 1] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ lesionWeight: num [1:5000, 1] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ dataType    : chr "ROC"
#>  $ modalityID  : chr "1"
#>  $ readerID    : chr [1:2] "1" "2"