Dobbins 1 FROC dataset

This is referred to in the book as the "DOB1" dataset. Dobbins et al conducted a multi-institutional, MRMC study to compare the performance of digital tomosynthesis (GE's VolumeRad device), dual-energy (DE) imaging, and conventional chest radiography for pulmonary nodule detection and management. All study images were obtained with a flat-panel detector developed by GE. The case set consisted of 158 subjects, of which 43 were non-diseased and the rest had 1 - 20 pulmonary nodules independently verified, using with CT images, by 3 experts who did not participate in the observer study. The study used FROC paradigm data collection. There are 4 treatments labeled 1 - 4 (conventional chest x-ray, CXR, CXR augmented with dual-energy (CXR+DE), VolumeRad digital tomosynthesis images and VolumeRad augmented with DE (VolumeRad+DE).

dataset11

Format

A list with 8 elements:

NL, num [1:4, 1:5, 1:158, 1:4], ratings of non-lesion localizations, NLs

LL, num [1:4, 1:5, 1:115, 1:20], ratings of lesion localizations, LLs

lesionVector, int [1:115], number of lesions per diseased case

lesionID, num [1:115, 20], labels of lesions on diseased cases, all set to 1

lesionWeight, num [1:115, 20] weights (or clinical importances) of lesions, all set to 1

dataType, chr "FROC", the data type

modalityID, chr [1:4] "1" "2" ..., treatment labels

readerID, chr [1:5] "1" "2" ..., reader labels

design, "CROSSED", study design

normalCases, int [1:43] 6 9 14 27 62 66 70 71 83 91 ..., normal case labels

abnormalCases, int [1:115] 1 2 3 5 7 8 10 11 13 17 ..., abnormal case labels

truthTableStr, num [1:4, 1:5, 1:158, 1:21] 1 1 1 1 1 1 ..., truth table structure

References

Dobbins III JT et al. Multi-Institutional Evaluation of Digital Tomosynthesis, Dual-Energy Radiography, and Conventional Chest Radiography for the Detection and Management of Pulmonary Nodules. Radiology. 2016;282(1):236-250.

Examples

str(dataset11)

#> List of 12
#>  $ NL           : num [1:4, 1:5, 1:158, 1:4] -Inf -Inf -Inf -Inf -Inf ...
#>  $ LL           : num [1:4, 1:5, 1:115, 1:20] -Inf -Inf -Inf -Inf -Inf ...
#>  $ lesionVector : int [1:115] 6 4 7 1 3 3 3 8 11 2 ...
#>  $ lesionID     : num [1:115, 1:20] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ lesionWeight : num [1:115, 1:20] 0.167 0.25 0.143 1 0.333 ...
#>  $ dataType     : chr "FROC"
#>  $ modalityID   : Named chr [1:4] "1" "2" "3" "4"
#>   ..- attr(*, "names")= chr [1:4] "1" "2" "3" "4"
#>  $ readerID     : Named chr [1:5] "1" "2" "3" "4" ...
#>   ..- attr(*, "names")= chr [1:5] "1" "2" "3" "4" ...
#>  $ design       : chr "CROSSED"
#>  $ normalCases  : int [1:43] 6 9 14 27 62 66 70 71 83 91 ...
#>  $ abnormalCases: int [1:115] 1 2 3 5 7 8 10 11 13 17 ...
#>  $ truthTableStr: num [1:4, 1:5, 1:158, 1:21] 1 1 1 1 1 1 1 1 1 1 ...