Dobbins 2 ROC dataset

This is referred to in the code as the "DOB2" dataset. It contains actionability ratings, i.e., do you recommend further follow up on the patient, one a 1 (definitely not) to 5 (definitely yes), effectively an ROC dataset using a 5-point rating scale.

dataset12

Format

A list with 12 elements:

NL, num [1:4, 1:5, 1:152, 1], ratings of false positives, FPs

LL, num [1:4, 1:5, 1:88, 1], ratings of true positives, TPs

lesionVector, int [1:88], number of lesions per diseased case, all set to 1

lesionID, num [1:88, 1], labels of lesions on diseased cases, all set to 1

lesionWeight, num [1:88, 1], weights (or clinical importances) of lesions, all set to 1

dataType, chr "ROC", the data type

modalityID, chr [1:4] "1" "2" "3" "4", treatment labels

readerID, chr [1:5] "1" "2" "3" "4" ..., reader labels

design, "CROSSED", study design

normalCases, int [1:64] 6 9 14 20 27 33 ..., normal case labels

abnormalCases, int [1:88] 1 2 3 5 7 8 10 11 13 18 ..., abnormal case labels

truthTableStr, num [1:4, 1:5, 1:152, 1:2] 1 1 1 1 1 1 ..., truth table structure

References

Dobbins III JT et al. Multi-Institutional Evaluation of Digital Tomosynthesis, Dual-Energy Radiography, and Conventional Chest Radiography for the Detection and Management of Pulmonary Nodules. Radiology. 2016;282(1):236-250.

Examples

str(dataset11)

#> List of 12
#>  $ NL           : num [1:4, 1:5, 1:158, 1:4] -Inf -Inf -Inf -Inf -Inf ...
#>  $ LL           : num [1:4, 1:5, 1:115, 1:20] -Inf -Inf -Inf -Inf -Inf ...
#>  $ lesionVector : int [1:115] 6 4 7 1 3 3 3 8 11 2 ...
#>  $ lesionID     : num [1:115, 1:20] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ lesionWeight : num [1:115, 1:20] 0.167 0.25 0.143 1 0.333 ...
#>  $ dataType     : chr "FROC"
#>  $ modalityID   : Named chr [1:4] "1" "2" "3" "4"
#>   ..- attr(*, "names")= chr [1:4] "1" "2" "3" "4"
#>  $ readerID     : Named chr [1:5] "1" "2" "3" "4" ...
#>   ..- attr(*, "names")= chr [1:5] "1" "2" "3" "4" ...
#>  $ design       : chr "CROSSED"
#>  $ normalCases  : int [1:43] 6 9 14 27 62 66 70 71 83 91 ...
#>  $ abnormalCases: int [1:115] 1 2 3 5 7 8 10 11 13 17 ...
#>  $ truthTableStr: num [1:4, 1:5, 1:158, 1:21] 1 1 1 1 1 1 1 1 1 1 ...