Dobbins 3 FROC dataset

This is referred to in the code as the "DOB3" dataset. This is a subset of DOB1 which includes data for lesions not-visible on CXR, but visible to truth panel on all treatments.

dataset13

Format

A list with 8 elements:

NL Ratings array [1:4, 1:5, 1:158, 1:4], of non-lesion localizations, NLs

LL Ratings array [1:4, 1:5, 1:106, 1:15], of lesion localizations, LLs

lesionNum array [1:106], number of lesions per diseased case, all set to 1

lesionID array [1:106, 15], labels of lesions on diseased cases, all set to 1

lesionWeight array [1:106, 15] weights (or clinical importance) of lesions, all set to 1

dataType "FROC", the data type

modalityID [1:4] "1" "2" ..., treatment labels

readerID [1:5] "1" "2" ..., reader labels

References

Dobbins III JT et al. Multi-Institutional Evaluation of Digital Tomosynthesis, Dual-Energy Radiography, and Conventional Chest Radiography for the Detection and Management of Pulmonary Nodules. Radiology. 2016;282(1):236-250.

Examples

str(dataset13)

#> List of 8
#>  $ NL          : num [1:4, 1:5, 1:158, 1:4] -Inf 3 -Inf 4 5 ...
#>  $ LL          : num [1:4, 1:5, 1:106, 1:15] -Inf -Inf -Inf -Inf -Inf ...
#>  $ lesionNum   : int [1:106] 3 1 7 2 3 6 10 1 4 5 ...
#>  $ lesionID    : num [1:106, 1:15] 2 3 1 2 1 2 1 2 1 2 ...
#>  $ lesionWeight: num [1:106, 1:15] 0.333 1 0.143 0.5 0.333 ...
#>  $ dataType    : chr "FROC"
#>  $ modalityID  : Named chr [1:4] "1" "2" "3" "4"
#>   ..- attr(*, "names")= chr [1:4] "1" "2" "3" "4"
#>  $ readerID    : Named chr [1:5] "1" "2" "3" "4" ...
#>   ..- attr(*, "names")= chr [1:5] "1" "2" "3" "4" ...