Federica Zanca real (as opposed to inferred) ROC dataset

This is referred to in the book as the "FZR" dataset. It is a real ROC study, conducted on the same images and using the same radiologists, on treatments "4" and "5" of dataset04. This was compared to highest rating inferred ROC data from dataset04 to conclude, erroneously, that the highest rating assumption is invalid. See book Section 13.6 and run "~/GitHub/RJafroc/inst/InferredVsReal/InferredVsReal.R".

dataset14

Format

A list with 3 elements: $ratings, $lesions and $descriptions; $ratings contain 3 elements, $NL, $LL and $LL_IL as sub-lists; $lesions contain 3 elements, $perCase, $IDs and $weights as sub-lists; $descriptions contain 7 elements, $fileName, $type, $name, $truthTableStr, $design, $modalityID and $readerID as sub-lists;

rating$NL num Array[1:2, 1:4, 1:200, 1] ratings of non-lesion localizations NLs

rating$LL num Array[1:2, 1:4, 1:100, 1] ratings of lesion localizations LLs

rating$LL_IL NA This placeholder is used only for LROC data

lesions$perCase int Array[1:100], number of lesions per diseased case

lesions$IDs num Array[1:100, 1] , numeric labels of lesions on diseased cases

lesions$weights num Array[1:100, 1], weights (or clinical importance) of lesions

descriptions$fileName chr "dataset14", base name of dataset in `data` folder

descriptions$type chr "ROC" the data type

descriptions$name chr "FEDERICA-REAL-ROC", the name of the dataset

descriptions$truthTableStr num Array[1:2, 1:4, 1:200, 1:2] truth table structure

descriptions$design chr "FCTRL", study design factorial dataset

descriptions$modalityID chr Array[1:2] modality label(s)

descriptions$readerID chr Array[1:4] reader labels

References

Zanca F, Hillis SL, Claus F, et al (2012) Correlation of free-response and receiver-operating-characteristic area-under-the-curve estimates: Results from independently conducted FROC/ROC studies in mammography. Med Phys. 39(10):5917-5929.

Examples

res <- str(dataset14)
#> List of 3
#>  $ ratings     :List of 3
#>   ..$ NL   : num [1:2, 1:4, 1:200, 1] 2 2 2 2 1 3 2 2 3 1 ...
#>   ..$ LL   : num [1:2, 1:4, 1:100, 1] 6 5 6 4 5 5 5 5 5 4 ...
#>   ..$ LL_IL: logi NA
#>  $ lesions     :List of 3
#>   ..$ perCase: int [1:100] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ IDs    : num [1:100, 1] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ weights: num [1:100, 1] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ descriptions:List of 7
#>   ..$ fileName     : chr "dataset14"
#>   ..$ type         : chr "ROC"
#>   ..$ name         : chr "FEDERICA-REAL-ROC"
#>   ..$ truthTableStr: num [1:2, 1:4, 1:200, 1:2] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ design       : chr "FCTRL"
#>   ..$ modalityID   : Named chr [1:2] "4" "5"
#>   .. ..- attr(*, "names")= chr [1:2] "4" "5"
#>   ..$ readerID     : Named chr [1:4] "1" "2" "3" "4"
#>   .. ..- attr(*, "names")= chr [1:4] "1" "2" "3" "4"