Nico Karssemeijer LROC dataset (CAD vs. radiologists)

This is the actual LROC data corresponding to dataset09, which was the inferred ROC data. Note that the LL field is split into two, LLCl, representing true positives where the lesions were correctly localized, and LLIl, representing true positives where the lesions were incorrectly localized. The first reader is CAD and the remaining readers are radiologists. The function StSignificanceTestingSingleFixedFactor analyzes such datasets.

datasetCadLroc

Format

A list with 9 elements:

NL Ratings array [1, 1:10, 1:200, 1], of false positives, FPs

LLCl Ratings array [1, 1:10, 1:80, 1], of true positives with correct localization, TPCls

LLIl Ratings array [1, 1:10, 1:80, 1], of true positives with incorrect localization, TPIls

lesionNum array [1:80], number of lesions per diseased case, all set to 1

lesionID array [1:80, 1], labels of lesions on diseased cases, all set to 1

lesionWeight array [1:80, 1], weights (or clinical importance) of lesions, all set to 1

dataType "LROC", the data type

modalityID [1:2] "0" "1", treatment labels

readerID [1:10] "1" "2" ..., reader labels

References

Hupse R et al. Standalone computer-aided detection compared to radiologists' performance for the detection of mammographic masses. Eur Radiol. 2013;23(1):93-100.

Examples

str(datasetCadLroc)

#> List of 9
#>  $ NL          : num [1, 1:10, 1:200, 1] 28 0 14 0 16 0 31 0 0 0 ...
#>  $ LLCl        : num [1, 1:10, 1:80, 1] 0 0 0 0 0 0 0 0 67 0 ...
#>  $ LLIl        : num [1, 1:10, 1:80, 1] 29 12 13 10 41 67 61 51 0 0 ...
#>  $ lesionNum   : num [1:80] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ lesionID    : num [1:80, 1] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ lesionWeight: num [1:80, 1] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ dataType    : chr "LROC"
#>  $ modalityID  : chr "1"
#>  $ readerID    : chr [1:10] "1" "2" "3" "4" ...