Simulated degenerate ROC dataset (for testing purposes)

A simulated degenerated dataset. A degenerate dataset is defined as one with no interior operating points on the ROC plot. Such data tend to be observed with expert level radiologists. This dataset is used to illustrate the robustness of two fitting models, namely CBM and RSM. The widely used binormal model and PROPROC fail on such datasets.

datasetDegenerate

Format

A list with 8 elements:

NL Ratings array [1, 1, 1:15, 1], of false positives, FPs

LL Ratings array [1, 1, 1:10, 1], of true positives, TPs

lesionVector array [1:10], number of lesions per diseased case, all set to 1

lesionID array [1:10, 1], labels of lesions on diseased cases, all set to 1

lesionWeight array [1:10, 1], weights (or clinical importance) of lesions, all set to 1

dataType "ROC", the data type

modalityID "1", treatment label

readerID "1", reader label

Examples

str(datasetDegenerate)

#> List of 8
#>  $ NL          : num [1, 1, 1:15, 1] 1 1 1 1 1 ...
#>  $ LL          : num [1, 1, 1:10, 1] 1 1 1 1 1 2 2 2 2 2
#>  $ lesionVector: num [1:10] 1 1 1 1 1 1 1 1 1 1
#>  $ lesionID    : num [1:10, 1] 1 1 1 1 1 1 1 1 1 1
#>  $ lesionWeight: num [1:10, 1] 1 1 1 1 1 1 1 1 1 1
#>  $ dataType    : chr "ROC"
#>  $ modalityID  : chr "1"
#>  $ readerID    : chr "1"