John Thompson crossed modality FROC dataset

This is a crossed modality dataset, see book Section 18.5. There are two modality factors. The first modality factor modalityID1 can be "F" or "I", which represent two CT reconstruction algorithms. The second modality factor modalityID2 can be "20" "40" "60" "80", which represent the mAs values of the image acquisition. The factors are fully crossed.

datasetX

Format

A list with 3 elements: $ratings, $lesions and $descriptions; $ratings contain 3 elements, $NL, $LL and $LL_IL as sub-lists; $lesions contain 3 elements, $perCase, $IDs and $weights as sub-lists; $descriptions contain 7 elements, $fileName, $type, $name, $truthTableStr, $design, $modalityID and $readerID as sub-lists;

rating$NL num Array[1:2, 1:4, 1:11, 1:68, 1:5] ratings of non-lesion localizations NLs

rating$LL num Array[1:2, 1:4, 1:11, 1:34, 1:3] ratings of lesion localizations LLs

rating$LL_IL NA This placeholder is used only for LROC data

lesions$perCase int Array[1:34], number of lesions per diseased case

lesions$IDs num Array[1:34, 1:3] , numeric labels of lesions on diseased cases

lesions$weights num Array[1:34, 1:3], weights (or clinical importance) of lesions

descriptions$fileName chr "datasetX", base name of dataset in `data` folder

descriptions$type chr "FROC" the data type

descriptions$name chr "THOMPSON-X-MOD", the name of the dataset

descriptions$truthTableStr NA truth table structure

descriptions$design chr "FCTRL-X-MOD", study design factorial dataset

descriptions$modalityID chr Array[1:2] "F" "I", modality label(s)

descriptions$readerID chr Array[1:4] "20" "40" "60" "80", reader labels

References

Thompson JD, Chakraborty DP, et al. (2016) Effect of reconstruction methods and x-ray tube current-time product on nodule detection in an anthropomorphic thorax phantom: a crossed-modality JAFROC observer study. Medical Physics. 43(3):1265-1274.

Examples

res <- str(datasetX)
#> List of 3
#>  $ ratings     :List of 3
#>   ..$ NL   : num [1:2, 1:4, 1:11, 1:68, 1:5] -Inf -Inf -Inf -Inf -Inf ...
#>   ..$ LL   : num [1:2, 1:4, 1:11, 1:34, 1:3] -Inf 1 2 8 7 ...
#>   ..$ LL_IL: logi NA
#>  $ lesions     :List of 3
#>   ..$ perCase: int [1:34] 1 2 1 1 1 1 2 2 1 1 ...
#>   ..$ IDs    : num [1:34, 1:3] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ weights: num [1:34, 1:3] 1 0.5 1 1 1 1 0.5 0.5 1 1 ...
#>  $ descriptions:List of 8
#>   ..$ fileName     : chr "DfReadXModalities(XModDataFile)"
#>   ..$ type         : chr "FROC"
#>   ..$ name         : chr "THOMPSON-X-MOD"
#>   ..$ truthTableStr: logi NA
#>   ..$ design       : chr "FCTRL-X-MOD"
#>   ..$ modalityID1  : Named chr [1:2] "F" "I"
#>   .. ..- attr(*, "names")= chr [1:2] "F" "I"
#>   ..$ modalityID2  : Named chr [1:4] "20" "40" "60" "80"
#>   .. ..- attr(*, "names")= chr [1:4] "20" "40" "60" "80"
#>   ..$ readerID     : Named chr [1:11] "1" "2" "3" "4" ...
#>   .. ..- attr(*, "names")= chr [1:11] "1" "2" "3" "4" ...