Convert an LROC dataset to a ROC dataset

Converts an LROC dataset to an ROC dataset

DfLroc2Roc(dataset)

Arguments

dataset

The LROC dataset to be converted.

Value

An ROC dataset

Details

For the diseased cases one takes the maximum rating on each diseased case, which could be a LL ("true positive" correct localization) or a LL_IL ("true positive" incorrect localization) rating, whichever has the higher rating. For non-diseased cases the NL arrays are identical.

Examples

rocDataSet <- DfLroc2Roc(datasetCadLroc)