RSM fitted model for FROC sample size

RSM fitted model for FROC sample size

SsFrocNhRsmModel(dataset, lesionPmf)

Arguments

dataset
The pilot dataset object representing a NH ROC (or FROC) dataset.

lesionPmf
An array containing the probability mass function of number of lesions per diseased case in the proposed pivotal FROC study.

Value

A list containing:

muMed, the median mu parameter of the NH model.

lambdaMed, the median lambda parameter of the NH model.

nuMed, the median nu parameter of the NH model.

lesDistr, the lesion distribution array.

lesWghtDistr, the lesion weight distribution array.

scaleFactor, the scaling factor that multiplies the ROC effect size to get wAFROC effect size.

R2, the R2 of the fit.

Details

If dataset is FROC, it is converted to an ROC dataset. The search model is used to fit each treatment-reader combination in the pilot dataset. The median value for each parameter is computed and are returned by the function (3 vaalues). These are used to compute predicted wAFROC and ROC FOMS over a range of values of deltaMu, which are fitted by a straight line constrained to pass throught the origin. The scaleFactor (scaling factor) and R2 are returned. The scaling factor is the value by which the ROC effect size must be multiplied to get the wAFROC effect size. Also returned are the lesDist and lesWghtDist arrays, which are needed for computing FOMs. See 2nd FROC SS vignette. Equally weighted lesions is assumed.

Examples

# \donttest{
SsFrocNhRsmModel(dataset02, c(0.7, 0.2, 0.1))

#> $muMed
#> [1] 3.476166
#> 
#> $lambdaMed
#> [1] 6.513176
#> 
#> $nuMed
#> [1] 0.5589415
#> 
#> $scaleFactor
#> [1] 2.216458
#> 
#> $lesDistr
#>      [,1] [,2]
#> [1,]    1  0.7
#> [2,]    2  0.2
#> [3,]    3  0.1
#> 
#> $lesWghtDistr
#>           [,1]      [,2]      [,3]
#> [1,] 1.0000000      -Inf      -Inf
#> [2,] 0.5000000 0.5000000      -Inf
#> [3,] 0.3333333 0.3333333 0.3333333
#> 
#> $R2
#> [1] 0.9999906
#> 

## the next one should match the vignette 
SsFrocNhRsmModel(DfExtractDataset(dataset04, trts = c(1,2)), c(0.69, 0.2, 0.11))

#> $muMed
#> [1] 3.310556
#> 
#> $lambdaMed
#> [1] 5.675511
#> 
#> $nuMed
#> [1] 0.3673814
#> 
#> $scaleFactor
#> [1] 2.168861
#> 
#> $lesDistr
#>      [,1] [,2]
#> [1,]    1 0.69
#> [2,]    2 0.20
#> [3,]    3 0.11
#> 
#> $lesWghtDistr
#>           [,1]      [,2]      [,3]
#> [1,] 1.0000000      -Inf      -Inf
#> [2,] 0.5000000 0.5000000      -Inf
#> [3,] 0.3333333 0.3333333 0.3333333
#> 
#> $R2
#> [1] 0.9999904
#> 

# }