Number of cases, for specified number of readers, to achieve desired power

Number of cases to achieve the desired power, for specified number of readers J, and specified DBMH or ORH analysis method

SsSampleSizeKGivenJ(dataset, J, alpha = 0.05, effectSize = NULL,
  desiredPower = 0.8, option = "ALL", method = "DBMH", ...)

Arguments

dataset
The pilot ROC dataset to be used to extrapolate to the pivotal study. If missing, then variance components and effectSize must be passed as additional parameters.

J
The number of readers in the pivotal study.

alpha
The significance level of the study, default is 0.05.

effectSize
The effect size to be used in the pivotal study, default value is NULL.

desiredPower
The desired statistical power, default is 0.8.

option
Desired generalization, "RRRC", "FRRC", "RRFC" or "ALL" (the default).

method
"DBMH" (default) or "ORH".

...
Other parameters, OR or DBM variance components, passed internally, see details

Value

A list of two elements:

K

The minimum number of cases K in the pivotal study to just achieve the desired statistical power.

power

The predicted statistical power.

Details

Other parameters ... are reserved for internal use. The default effectSize uses the observed effect size in the pilot study. A numeric value over-rides the default value.

Note

The procedure is valid for ROC studies only; for FROC studies see Online Appendix Chapter 19.

References

Hillis SL, Obuchowski NA, Berbaum KS (2011) Power Estimation for Multireader ROC Methods: An Updated and Unified Approach, Acad Radiol, 18, 129--142.

Hillis SL, Obuchowski NA, Schartz KM, Berbaum KS (2005) A comparison of the Dorfman-Berbaum-Metz and Obuchowski-Rockette methods for receiver operating characteristic (ROC) data, Statistics in Medicine, 24:10, 1579--607.

Examples

## An example using the DBM method
SsSampleSizeKGivenJ(dataset02, J = 6, method = "DBMH")

#> $KRRRC
#> [1] 251
#> 
#> $powerRRRC
#> [1] 0.8005403
#> 
#> $KFRRC
#> [1] 155
#> 
#> $powerFRRC
#> [1] 0.8014626
#> 
#> $KRRFC
#> [1] 151
#> 
#> $powerRRFC
#> [1] 0.8008112
#> 

## An example using the OR method
SsSampleSizeKGivenJ(dataset02, J = 6, method = "ORH")

#> $KRRRC
#> [1] 251
#> 
#> $powerRRRC
#> [1] 0.8005403
#> 
#> $KFRRC
#> [1] 155
#> 
#> $powerFRRC
#> [1] 0.8014626
#> 
#> $KRRFC
#> [1] 151
#> 
#> $powerRRFC
#> [1] 0.8008112
#> 

for (J in 6:10) {
 ret <- SsSampleSizeKGivenJ(dataset02, J = J, option = "RRRC")
 message("# of readers = ", J, " estimated # of cases = ", ret$K, ", predicted power = ",
    signif(ret$powerRRRC,3), "\n")
}

#> # of readers = 6 estimated # of cases = 251, predicted power = 0.801

#> # of readers = 7 estimated # of cases = 211, predicted power = 0.801

#> # of readers = 8 estimated # of cases = 188, predicted power = 0.801

#> # of readers = 9 estimated # of cases = 173, predicted power = 0.801

#> # of readers = 10 estimated # of cases = 163, predicted power = 0.802