DBM or OR significance testing for a one treatment factorial or two-treatment crossed modality factorial dataset (not SPLIT_PLOT)

Performs DBM or OR significance testing for the dataset.

St(
  dataset,
  FOM,
  method = "OR",
  covEstMethod = "jackknife",
  analysisOption = "RRRC",
  alpha = 0.05,
  FPFValue = 0.2,
  nBoots = 200,
  seed = NULL,
  details = 0
)

Arguments

dataset

The dataset to be analyzed, see RJafroc-package. The dataset design can be "FCTRL" or "FCTRL-X-MOD".

FOM

The figure of merit, see UtilFigureOfMerit

method

The significance testing method to be used: "DBM" for the Dorfman-Berbaum-Metz method or "OR" for the Obuchowski-Rockette method (default).

covEstMethod

The covariance matrix estimation method in ORH analysis (for method = "DBM" the jackknife is always used).

"Jackknife" (default),

"Bootstrap", in which case nBoots is relevant, default 200,

"DeLong"; requires FOM = "Wilcoxon" or "ROI" or "HrAuc".

analysisOption

Determines which factors are regarded as random and which are fixed:

"RRRC" = random-reader random case (default),

"FRRC" = fixed-reader random case,

"RRFC" = random-reader fixed case,

alpha

The significance level (alpha) of the test of the null hypothesis that all modality effects are zero (default: alpha = 0.05).

FPFValue

Only needed for LROC data and FOM = "PCL" or "ALROC"; where to evaluate a partial curve based figure of merit. (default: FPFValue = 0.2).

nBoots

The number of bootstraps (defaults to 200), only needed if covEstMethod = "bootstrap" and method = "OR"

seed

For bootstraps the seed of the RNG (default: seed = NULL), only needed if method = "OR" and covEstMethod = "bootstrap".

details

Amount of explanations in output, default is 0 for no explanations and 1 for explanations.

Value

A list containing the results of the analysis.

Note

details = 0 should suffice for factorial dataset analysis since the names of the output lists are self-explanatory. For cross-modality analysis details = 1 is suggested to better understand the output.

References

Dorfman DD, Berbaum KS, Metz CE (1992) ROC characteristic rating analysis: Generalization to the Population of Readers and Patients with the Jackknife method, Invest. Radiol. 27, 723-731.

Obuchowski NA, Rockette HE (1995) Hypothesis Testing of the Diagnostic Accuracy for Multiple Diagnostic Tests: An ANOVA Approach with Dependent Observations, Communications in Statistics: Simulation and Computation 24, 285-308.

Hillis SL (2014) A marginal-mean ANOVA approach for analyzing multireader multicase radiological imaging data, Statistics in medicine 33, 330-360.

Thompson JD, Chakraborty DP, Szczepura K, et al. (2016) Effect of reconstruction methods and x-ray tube current-time product on nodule detection in an anthropomorphic thorax phantom: a crossed-modality JAFROC observer study. Medical Physics. 43(3):1265-1274.

Chakraborty DP (2017) Observer Performance Methods for Diagnostic Imaging - Foundations, Modeling, and Applications with R-Based Examples, CRC Press, Boca Raton, FL. https://www.routledge.com/Observer-Performance-Methods-for-Diagnostic-Imaging-Foundations-Modeling/Chakraborty/p/book/9781482214840

Examples

result <- St(dataset02,FOM = "Wilcoxon", method = "DBM")
result <- St(dataset02,FOM = "Wilcoxon", method = "OR")
result <- St(datasetX, FOM = "wAFROC", method = "OR", analysisOption = "RRRC")

# \donttest{
result <- St(dataset05, FOM = "wAFROC")
result <- St(dataset05, FOM = "HrAuc", method = "DBM")
# }