Performs DBM or OR significance testing for factorial or split-plot A,C datasets

Performs Dorfman-Berbaum-Metz (DBM) or Obuchowski-Rockette (OR) significance testing, for specified dataset; significance testing refers to analysis designed to assign a P-value, and other statistics, for rejecting the null hypothesis (NH) that the reader-averaged figure of merit (FOM) differences between treatments is zero. The results of the analysis are best visualized in the text or Excel-formatted files produced by UtilOutputReport.

StSignificanceTesting(
  dataset,
  FOM,
  FPFValue = 0.2,
  alpha = 0.05,
  method = "DBM",
  covEstMethod = "jackknife",
  nBoots = 200,
  analysisOption = "ALL",
  tempOrgCode = FALSE
)

Arguments

dataset

The dataset to be analyzed, see RJafroc-package. Must have two or more treatments and two or more readers. The dataset design can be "FCTRL", "SPLIT-PLOT-A" or "SPLIT-PLOT-C".

FOM

The figure of merit, see UtilFigureOfMerit

FPFValue

Only needed for LROC data and FOM = "PCL" or "ALROC"; where to evaluate a partial curve based figure of merit. The default is 0.2.

alpha

The significance level of the test of the null hypothesis that all treatment effects are zero; the default is 0.05

method

The significance testing method to be used: "DBM" (the default), representing the Dorfman-Berbaum-Metz method or "OR", representing the Obuchowski-Rockette method. and the Obuchowski-Rockette significance testing methods, respectively.

covEstMethod

The covariance matrix estimation method in ORH analysis (for method = "DBM" the jackknife is always used).

"Jackknife", the default,

"Bootstrap", in which case nBoots (above) is relevant,

"DeLong"; requires FOM = "Wilcoxon" or "ROI" or "HrAuc", otherwise an error results.

nBoots

The number of bootstraps (defaults to 200), relevant only if covEstMethod = "bootstrap" and method = "OR"

analysisOption

Determines which factors are regarded as random vs. fixed:

"RRRC" = random-reader random case,

"FRRC" = fixed-reader random case,

"RRFC" = random-reader fixed case,

"ALL" = outputs results of "RRRC", "FRRC" and "RRFC" analyses - this is the default.

tempOrgCode,

default FALSE; if TRUE, then code from version 0.0.1 of RJafroc is used (see RJafroc_0.0.1.tar). This is intended to check against errors that crept in subsequent to the original version as I attempted to improve the organization of the code and the output. As implicit in the name of this temporary flag, it will eventually be removed.

Value

For method = "DBM" the returned list contains 4 dataframes:

FOMs

Contains foms, trtMeans and trtMeanDiffs: see return of UtilFigureOfMerit

ANOVA

Contains TRCAnova, VarCom, IndividualTrt and IndividualRdr ANOVA tables of pseudovalues

RRRC

Contains results of "RRRC" analyses: FTests, ciDiffTrt, ciAvgRdrEachTrt

FRRC

Contains results of "FRRC" analyses: FTests, ciDiffTrt, ciAvgRdrEachTrt, ciDiffTrtEachRdr

RRFC

Contains results of "RRFC" analyses: FTests, ciDiffTrt, ciAvgRdrEachTrt

For method = "OR" the return list contains 4 dataframes:

FOMs

Contains foms, trtMeans and trtMeanDiffs: UtilFigureOfMerit

ANOVA

Contains TRAnova, VarCom, IndividualTrt and IndividualRdr ANOVA tables of FOM values

RRRC

Contains results of "RRRC" analyses - same organization as DBM, see above

FRRC

Contains results of "FRRC" analyses - ditto

RRFC

Contains results of "RRFC" analyses- ditto

References

Dorfman DD, Berbaum KS, Metz CE (1992) ROC characteristic rating analysis: Generalization to the Population of Readers and Patients with the Jackknife method, Invest. Radiol. 27, 723-731.

Obuchowski NA, Rockette HE (1995) Hypothesis Testing of the Diagnostic Accuracy for Multiple Diagnostic Tests: An ANOVA Approach with Dependent Observations, Communications in Statistics: Simulation and Computation 24, 285-308.

Hillis SL (2014) A marginal-mean ANOVA approach for analyzing multireader multicase radiological imaging data, Statistics in medicine 33, 330-360.

Chakraborty DP (2017) Observer Performance Methods for Diagnostic Imaging - Foundations, Modeling, and Applications with R-Based Examples, CRC Press, Boca Raton, FL. https://www.routledge.com/Observer-Performance-Methods-for-Diagnostic-Imaging-Foundations-Modeling/Chakraborty/p/book/9781482214840

Examples

StSignificanceTesting(dataset02,FOM = "Wilcoxon", method = "DBM") 
#> $FOMs
#> $FOMs$foms
#>            rdr0       rdr1       rdr2       rdr3       rdr4
#> trt0 0.91964573 0.85877617 0.90386473 0.97310789 0.82979066
#> trt1 0.94782609 0.90531401 0.92173913 0.99935588 0.92995169
#> 
#> $FOMs$trtMeans
#>        Estimate
#> trt0 0.89703704
#> trt1 0.94083736
#> 
#> $FOMs$trtMeanDiffs
#>               Estimate
#> trt0-trt1 -0.043800322
#> 
#> 
#> $ANOVA
#> $ANOVA$TRCanova
#>                 SS   DF          MS
#> T       0.54676344    1 0.546763441
#> R       1.74930720    4 0.437326799
#> C      44.84629692  113 0.396869884
#> TR      0.25126996    4 0.062817491
#> TC     11.28283352  113 0.099848084
#> RC     29.15447929  452 0.064501060
#> TRC    18.06716464  452 0.039971603
#> Total 105.89811497 1139          NA
#> 
#> $ANOVA$VarCom
#>            Estimates
#> VarR   0.00153499935
#> VarC   0.02724923428
#> VarTR  0.00020040252
#> VarTC  0.01197529621
#> VarRC  0.01226472859
#> VarErr 0.03997160319
#> 
#> $ANOVA$IndividualTrt
#>       DF        Trt0        Trt1
#> msR    4 0.351419668 0.148724622
#> msC  113 0.336294298 0.160423670
#> msRC 452 0.060436071 0.044036593
#> 
#> $ANOVA$IndividualRdr
#>       DF        rdr0        rdr1        rdr2        rdr3       rdr4
#> msT    1 0.045265545 0.123448933 0.018211160 0.039270539 0.57183723
#> msC  113 0.122249800 0.212748238 0.136533908 0.017289369 0.16605281
#> msTC 113 0.037099528 0.039431808 0.055520398 0.017042354 0.11064041
#> 
#> 
#> $RRRC
#> $RRRC$FTests
#>                  DF         MS     FStat           p
#> Treatment  1.000000 0.54676344 4.4563187 0.051665686
#> Error     15.259675 0.12269397        NA          NA
#> 
#> $RRRC$ciDiffTrt
#>               Estimate      StdErr        DF          t       PrGTt
#> trt0-trt1 -0.043800322 0.020748618 15.259675 -2.1109995 0.051665686
#>                CILower       CIUpper
#> trt0-trt1 -0.087959499 0.00035885444
#> 
#> $RRRC$ciAvgRdrEachTrt
#>        Estimate      StdErr        DF    CILower    CIUpper
#> trt0 0.89703704 0.033173597 12.744648 0.82522360 0.96885048
#> trt1 0.94083736 0.021566368 12.710190 0.89413783 0.98753689
#> 
#> 
#> $FRRC
#> $FRRC$FTests
#>            DF          MS     FStat           p
#> Treatment   1 0.546763441 5.4759532 0.021034969
#> Error     113 0.099848084        NA          NA
#> 
#> $FRRC$ciDiffTrt
#>               Estimate      StdErr  DF          t       PrGTt      CILower
#> trt0-trt1 -0.043800322 0.018717483 113 -2.3400755 0.021034969 -0.080883031
#>                 CIUpper
#> trt0-trt1 -0.0067176131
#> 
#> $FRRC$ciAvgRdrEachTrt
#>        Estimate      StdErr  DF    CILower    CIUpper
#> trt0 0.89703704 0.024289710 113 0.84891474 0.94515933
#> trt1 0.94083736 0.016776324 113 0.90760044 0.97407428
#> 
#> $FRRC$ciDiffTrtEachRdr
#>                     Estimate      StdErr  DF           t       PrGTt
#> rdr0::trt0-trt1 -0.028180354 0.025512133 113 -1.10458638 0.271685324
#> rdr1::trt0-trt1 -0.046537842 0.026301827 113 -1.76937679 0.079528660
#> rdr2::trt0-trt1 -0.017874396 0.031209647 113 -0.57272023 0.567971968
#> rdr3::trt0-trt1 -0.026247987 0.017291289 113 -1.51798907 0.131809983
#> rdr4::trt0-trt1 -0.100161031 0.044057460 113 -2.27341816 0.024890494
#>                      CILower       CIUpper
#> rdr0::trt0-trt1 -0.078724490  0.0223637817
#> rdr1::trt0-trt1 -0.098646505  0.0055708210
#> rdr2::trt0-trt1 -0.079706336  0.0439575435
#> rdr3::trt0-trt1 -0.060505148  0.0080091742
#> rdr4::trt0-trt1 -0.187446806 -0.0128752553
#> 
#> 
#> $RRFC
#> $RRFC$FTests
#>           DF          MS FStat           p
#> Treatment  1 0.546763441 8.704 0.041958752
#> Error      4 0.062817491    NA          NA
#> 
#> $RRFC$ciDiffTrt
#>               Estimate      StdErr DF          t       PrGTt      CILower
#> trt0-trt1 -0.043800322 0.014846287  4 -2.9502542 0.041958752 -0.085020224
#>                 CIUpper
#> trt0-trt1 -0.0025804202
#> 
#> $RRFC$ciAvgRdrEachTrt
#>        Estimate      StdErr DF    CILower    CIUpper
#> trt0 0.89703704 0.024829936  4 0.82809808 0.96597599
#> trt1 0.94083736 0.016153030  4 0.89598936 0.98568536
#> 
#> 
StSignificanceTesting(dataset02,FOM = "Wilcoxon", method = "OR")
#> $FOMs
#> $FOMs$foms
#>            rdr0       rdr1       rdr2       rdr3       rdr4
#> trt0 0.91964573 0.85877617 0.90386473 0.97310789 0.82979066
#> trt1 0.94782609 0.90531401 0.92173913 0.99935588 0.92995169
#> 
#> $FOMs$trtMeans
#>        Estimate
#> trt0 0.89703704
#> trt1 0.94083736
#> 
#> $FOMs$trtMeanDiffs
#>               Estimate
#> trt0-trt1 -0.043800322
#> 
#> 
#> $ANOVA
#> $ANOVA$TRanova
#>              SS DF            MS
#> T  0.0047961705  1 0.00479617053
#> R  0.0153448000  4 0.00383619999
#> TR 0.0022041225  4 0.00055103062
#> 
#> $ANOVA$VarCom
#>           Estimates       Rhos
#> VarR  0.00153499935         NA
#> VarTR 0.00020040252         NA
#> Cov1  0.00034661371 0.43203138
#> Cov2  0.00034407483 0.42886683
#> Cov3  0.00023902837 0.29793328
#> Var   0.00080228827         NA
#> 
#> $ANOVA$IndividualTrt
#>      DF   msREachTrt    varEachTrt   cov2EachTrt
#> trt0  4 0.0030826287 0.00101410277 0.00048396180
#> trt1  4 0.0013046019 0.00059047376 0.00020418785
#> 
#> $ANOVA$IndividualRdr
#>      DF    msTEachRdr    varEachRdr   cov1EachRdr
#> rdr0  1 0.00039706618 0.00069890056 3.7346610e-04
#> rdr1  1 0.00108288538 0.00110605283 7.6015978e-04
#> rdr2  1 0.00015974702 0.00084234345 3.5532241e-04
#> rdr3  1 0.00034447841 0.00015057773 1.0833995e-06
#> rdr4  1 0.00501611603 0.00121356676 2.4303685e-04
#> 
#> 
#> $RRRC
#> $RRRC$FTests
#>                  DF           MS     FStat           p
#> Treatment  1.000000 0.0047961705 4.4563187 0.051665686
#> Error     15.259675 0.0010762629        NA          NA
#> 
#> $RRRC$ciDiffTrt
#>               Estimate      StdErr        DF          t       PrGTt
#> trt0-trt1 -0.043800322 0.020748618 15.259675 -2.1109995 0.051665686
#>                CILower       CIUpper
#> trt0-trt1 -0.087959499 0.00035885444
#> 
#> $RRRC$ciAvgRdrEachTrt
#>        Estimate      StdErr        DF    CILower    CIUpper          Cov2
#> trt0 0.89703704 0.033173597 12.744648 0.82522360 0.96885048 0.00048396180
#> trt1 0.94083736 0.021566368 12.710190 0.89413783 0.98753689 0.00020418785
#> 
#> 
#> $FRRC
#> $FRRC$FTests
#>                      MS     Chisq DF           p
#> Treatment 0.00479617053 5.4759532  1 0.019279843
#> Error     0.00087586039        NA NA          NA
#> 
#> $FRRC$ciDiffTrt
#>               Estimate      StdErr          z       PrGTz      CILower
#> trt0-trt1 -0.043800322 0.018717483 -2.3400755 0.019279843 -0.080485914
#>                 CIUpper
#> trt0-trt1 -0.0071147303
#> 
#> $FRRC$ciAvgRdrEachTrt
#>        Estimate      StdErr  DF    CILower    CIUpper
#> trt0 0.89703704 0.024289710 113 0.84943008 0.94464399
#> trt1 0.94083736 0.016776324 113 0.90795637 0.97371835
#> 
#> $FRRC$ciDiffTrtEachRdr
#>                     Estimate      StdErr           z       PrGTz      CILower
#> rdr0::trt0-trt1 -0.028180354 0.025512133 -1.10458638 0.269338854 -0.078183215
#> rdr1::trt0-trt1 -0.046537842 0.026301827 -1.76937679 0.076831017 -0.098088476
#> rdr2::trt0-trt1 -0.017874396 0.031209647 -0.57272023 0.566834139 -0.079044180
#> rdr3::trt0-trt1 -0.026247987 0.017291289 -1.51798907 0.129017153 -0.060138290
#> rdr4::trt0-trt1 -0.100161031 0.044057460 -2.27341816 0.023000993 -0.186512066
#>                       CIUpper
#> rdr0::trt0-trt1  0.0218225068
#> rdr1::trt0-trt1  0.0050127916
#> rdr2::trt0-trt1  0.0432953879
#> rdr3::trt0-trt1  0.0076423157
#> rdr4::trt0-trt1 -0.0138099949
#> 
#> $FRRC$IndividualRdrVarCov1
#>         varEachRdr   cov1EachRdr
#> rdr0 0.00069890056 3.7346610e-04
#> rdr1 0.00110605283 7.6015978e-04
#> rdr2 0.00084234345 3.5532241e-04
#> rdr3 0.00015057773 1.0833995e-06
#> rdr4 0.00121356676 2.4303685e-04
#> 
#> 
#> $RRFC
#> $RRFC$FTests
#>    DF            MS     F           p
#> T   1 0.00479617053 8.704 0.041958752
#> TR  4 0.00055103062    NA          NA
#> 
#> $RRFC$ciDiffTrt
#>               Estimate      StdErr DF          t       PrGTt      CILower
#> trt0-trt1 -0.043800322 0.014846287  4 -2.9502542 0.041958752 -0.085020224
#>                 CIUpper
#> trt0-trt1 -0.0025804202
#> 
#> $RRFC$ciAvgRdrEachTrt
#>        Estimate      StdErr DF    CILower    CIUpper
#> Trt0 0.89703704 0.024829936  4 0.82809808 0.96597599
#> Trt1 0.94083736 0.016153030  4 0.89598936 0.98568536
#> 
#> 
## following is split-plot-c analysis using a simulated split-plot-c dataset
StSignificanceTesting(datasetFROCSpC, FOM = "wAFROC", method = "OR")
#> $FOMs
#> $FOMs$foms
#>            rdr1       rdr3       rdr4       rdr5
#> trt4 0.87681159 0.79202279 0.57312253 0.78758818
#> trt5 0.73460145 0.72863248 0.56571146 0.68992504
#> 
#> $FOMs$trtMeans
#>        Estimate
#> trt4 0.75738627
#> trt5 0.67971761
#> 
#> $FOMs$trtMeanDiffs
#>              Estimate
#> trt4-trt5 0.077668666
#> 
#> 
#> $ANOVA
#> $ANOVA$TRanova
#>              SS DF           MS
#> T  0.0120648434  1 0.0120648434
#> R  0.0639811804  3 0.0213270601
#> TR 0.0048526915  3 0.0016175638
#> 
#> $ANOVA$VarCom
#>       Estimates Rhos
#> VarR         NA   NA
#> VarTR        NA   NA
#> Cov1         NA   NA
#> Cov2          0   NA
#> Cov3          0   NA
#> Var          NA   NA
#> 
#> $ANOVA$IndividualTrt
#>      DF   msREachTrt varEachTrt cov2EachTrt
#> trt4  3 0.0167757928         NA           0
#> trt5  3 0.0061688312         NA           0
#> 
#> $ANOVA$IndividualRdr
#>      DF    msTEachRdr varEachRdr cov1EachRdr
#> rdr1  1 1.0111863e-02         NA          NA
#> rdr3  1 2.0091659e-03         NA          NA
#> rdr4  1 2.7461958e-05         NA          NA
#> rdr5  1 4.7690444e-03         NA          NA
#> 
#> 
#> $RRRC
#> $RRRC$FTests
#>           DF           MS     FStat           p
#> Treatment  1 0.0120648434 7.4586506 0.071876984
#> Error      3 0.0016175638        NA          NA
#> 
#> $RRRC$ciDiffTrt
#>              Estimate      StdErr DF        t       PrGTt      CILower
#> trt4-trt5 0.077668666 0.028439091  3 2.731053 0.071876984 -0.012837215
#>              CIUpper
#> trt4-trt5 0.16817455
#> 
#> $RRRC$ciAvgRdrEachTrt
#>        Estimate      StdErr DF    CILower    CIUpper Cov2
#> trt4 0.75738627 0.064760699  3 0.55128883 0.96348372    0
#> trt5 0.67971761 0.039270954  3 0.55473991 0.80469531    0
#> 
#> 
#> $FRRC
#> $FRRC$FTests
#>                    MS Chisq DF  p
#> Treatment 0.012064843    NA  1 NA
#> Error              NA    NA NA NA
#> 
#> $FRRC$ciDiffTrt
#>              Estimate StdErr  z PrGTz CILower CIUpper
#> trt4-trt5 0.077668666     NA NA    NA      NA      NA
#> 
#> $FRRC$ciAvgRdrEachTrt
#>        Estimate StdErr  DF CILower CIUpper
#> trt4 0.75738627     NA 199      NA      NA
#> trt5 0.67971761     NA 199      NA      NA
#> 
#> $FRRC$ciDiffTrtEachRdr
#>                     Estimate StdErr  z PrGTz CILower CIUpper
#> rdr1::trt4-trt5 0.1422101449     NA NA    NA      NA      NA
#> rdr3::trt4-trt5 0.0633903134     NA NA    NA      NA      NA
#> rdr4::trt4-trt5 0.0074110672     NA NA    NA      NA      NA
#> rdr5::trt4-trt5 0.0976631393     NA NA    NA      NA      NA
#> 
#> $FRRC$IndividualRdrVarCov1
#>      varEachRdr cov1EachRdr
#> rdr1         NA          NA
#> rdr3         NA          NA
#> rdr4         NA          NA
#> rdr5         NA          NA
#> 
#> 
#> $RRFC
#> $RRFC$FTests
#>    DF           MS         F           p
#> T   1 0.0120648434 7.4586506 0.071876984
#> TR  3 0.0016175638        NA          NA
#> 
#> $RRFC$ciDiffTrt
#>              Estimate      StdErr DF        t       PrGTt      CILower
#> trt4-trt5 0.077668666 0.028439091  3 2.731053 0.071876984 -0.012837215
#>              CIUpper
#> trt4-trt5 0.16817455
#> 
#> $RRFC$ciAvgRdrEachTrt
#>        Estimate      StdErr DF    CILower    CIUpper
#> Trt4 0.75738627 0.064760699  3 0.55128883 0.96348372
#> Trt5 0.67971761 0.039270954  3 0.55473991 0.80469531
#> 
#> 

# \donttest{
StSignificanceTesting(dataset05, FOM = "wAFROC")
#> $FOMs
#> $FOMs$foms
#>            rdr1       rdr2       rdr3       rdr4       rdr5       rdr6
#> trt1 0.72446809 0.88096927 0.80957447 0.61331757 0.57730496 0.84933018
#> trt2 0.80236407 0.96855792 0.84598109 0.75141844 0.72092199 0.87186761
#>            rdr7       rdr8       rdr9
#> trt1 0.89278960 0.80260047 0.76398739
#> trt2 0.93695823 0.89948779 0.81903073
#> 
#> $FOMs$trtMeans
#>        Estimate
#> trt1 0.76826022
#> trt2 0.84628754
#> 
#> $FOMs$trtMeanDiffs
#>               Estimate
#> trt1-trt2 -0.078027318
#> 
#> 
#> $ANOVA
#> $ANOVA$TRCanova
#>                 SS   DF          MS
#> T       2.52054062    1 2.520540621
#> R      13.49633191    8 1.687041489
#> C      70.13689242   91 0.770735082
#> TR      0.68282675    8 0.085353344
#> TC     34.21044848   91 0.375938994
#> RC     91.67293479  728 0.125924361
#> TRC    54.43725576  728 0.074776450
#> Total 267.15723074 1655          NA
#> 
#> $ANOVA$VarCom
#>            Estimates
#> VarR   0.00842684910
#> VarC   0.01909156536
#> VarTR  0.00011496624
#> VarTC  0.03346250490
#> VarRC  0.02557395538
#> VarErr 0.07477645022
#> 
#> $ANOVA$IndividualTrt
#>       DF       Trt1        Trt2
#> msR    8 1.15087970 0.621515134
#> msC   91 0.75365687 0.393017207
#> msRC 728 0.10448481 0.096216005
#> 
#> $ANOVA$IndividualRdr
#>      DF       rdr1        rdr2        rdr3       rdr4       rdr5        rdr6
#> msT   1 0.27911806 0.352901514 0.060970329 0.87730507 0.94878904 0.023365047
#> msC  91 0.28285163 0.074277854 0.196690575 0.28828435 0.33493737 0.141461841
#> msTC 91 0.13584273 0.052890217 0.094720971 0.12444654 0.15027163 0.119121220
#>             rdr7        rdr8       rdr9
#> msT  0.089739950 0.431808964 0.13936939
#> msC  0.069053606 0.157958695 0.23261405
#> msTC 0.067979253 0.085070876 0.14380716
#> 
#> 
#> $RRRC
#> $RRRC$FTests
#>                  DF         MS     FStat           p
#> Treatment   1.00000 2.52054062 6.5211824 0.011570576
#> Error     164.05279 0.38651589        NA          NA
#> 
#> $RRRC$ciDiffTrt
#>               Estimate      StdErr        DF          t       PrGTt     CILower
#> trt1-trt2 -0.078027318 0.030555086 164.05279 -2.5536606 0.011570576 -0.13835925
#>                CIUpper
#> trt1-trt2 -0.017695389
#> 
#> $RRRC$ciAvgRdrEachTrt
#>        Estimate      StdErr        DF    CILower    CIUpper
#> trt1 0.76826022 0.046625911 19.570419 0.67086323 0.86565721
#> trt2 0.84628754 0.033302818 17.465105 0.77616700 0.91640808
#> 
#> 
#> $FRRC
#> $FRRC$FTests
#>           DF         MS     FStat           p
#> Treatment  1 2.52054062 6.7046533 0.011196091
#> Error     91 0.37593899        NA          NA
#> 
#> $FRRC$ciDiffTrt
#>               Estimate     StdErr DF          t       PrGTt     CILower
#> trt1-trt2 -0.078027318 0.03013412 91 -2.5893345 0.011196091 -0.13788505
#>               CIUpper
#> trt1-trt2 -0.01816959
#> 
#> $FRRC$ciAvgRdrEachTrt
#>        Estimate      StdErr DF    CILower    CIUpper
#> trt1 0.76826022 0.030169747 91 0.70833173 0.82818872
#> trt2 0.84628754 0.021786658 91 0.80301102 0.88956406
#> 
#> $FRRC$ciDiffTrtEachRdr
#>                     Estimate      StdErr DF           t        PrGTt
#> rdr1::trt1-trt2 -0.077895981 0.054342459 91 -1.43342761 0.1551626262
#> rdr2::trt1-trt2 -0.087588652 0.033908514 91 -2.58308730 0.0113864383
#> rdr3::trt1-trt2 -0.036406619 0.045377875 91 -0.80229890 0.4244703693
#> rdr4::trt1-trt2 -0.138100867 0.052013071 91 -2.65511851 0.0093592862
#> rdr5::trt1-trt2 -0.143617021 0.057155704 91 -2.51273294 0.0137406464
#> rdr6::trt1-trt2 -0.022537431 0.050888031 91 -0.44288275 0.6589015625
#> rdr7::trt1-trt2 -0.044168637 0.038442293 91 -1.14895949 0.2535837070
#> rdr8::trt1-trt2 -0.096887313 0.043004266 91 -2.25296982 0.0266642339
#> rdr9::trt1-trt2 -0.055043341 0.055912813 91 -0.98444951 0.3275047715
#>                     CILower      CIUpper
#> rdr1::trt1-trt2 -0.18584060  0.030048638
#> rdr2::trt1-trt2 -0.15494375 -0.020233555
#> rdr3::trt1-trt2 -0.12654419  0.053730955
#> rdr4::trt1-trt2 -0.24141844 -0.034783291
#> rdr5::trt1-trt2 -0.25714981 -0.030084236
#> rdr6::trt1-trt2 -0.12362025  0.078545392
#> rdr7::trt1-trt2 -0.12052953  0.032192255
#> rdr8::trt1-trt2 -0.18231000 -0.011464622
#> rdr9::trt1-trt2 -0.16610728  0.056020593
#> 
#> 
#> $RRFC
#> $RRFC$FTests
#>           DF          MS    FStat             p
#> Treatment  1 2.520540621 29.53066 0.00062022543
#> Error      8 0.085353344       NA            NA
#> 
#> $RRFC$ciDiffTrt
#>               Estimate      StdErr DF         t         PrGTt     CILower
#> trt1-trt2 -0.078027318 0.014358534  8 -5.434212 0.00062022543 -0.11113816
#>                CIUpper
#> trt1-trt2 -0.044916479
#> 
#> $RRFC$ciAvgRdrEachTrt
#>        Estimate      StdErr DF    CILower    CIUpper
#> trt1 0.76826022 0.037282051  8 0.68228766 0.85423279
#> trt2 0.84628754 0.027397484  8 0.78310883 0.90946625
#> 
#> 
StSignificanceTesting(dataset05, FOM = "HrAuc", method = "DBM") 
#> $FOMs
#> $FOMs$foms
#>            rdr1       rdr2       rdr3       rdr4       rdr5       rdr6
#> trt1 0.85886525 0.89314421 0.88226950 0.73026005 0.71418440 0.92907801
#> trt2 0.89148936 0.97612293 0.88274232 0.84822695 0.83995272 0.90898345
#>            rdr7       rdr8      rdr9
#> trt1 0.91891253 0.87919622 0.8144208
#> trt2 0.95248227 0.93380615 0.8569740
#> 
#> $FOMs$trtMeans
#>        Estimate
#> trt1 0.84670344
#> trt2 0.89897557
#> 
#> $FOMs$trtMeanDiffs
#>              Estimate
#> trt1-trt2 -0.05227213
#> 
#> 
#> $ANOVA
#> $ANOVA$TRCanova
#>                 SS   DF          MS
#> T       1.13120350    1 1.131203500
#> R       5.27623066    8 0.659528832
#> C      55.82349116   91 0.613444958
#> TR      0.89314716    8 0.111643395
#> TC     20.79022710   91 0.228464034
#> RC     71.32075503  728 0.097968070
#> TRC    48.61282976  728 0.066775865
#> Total 203.84788437 1655          NA
#> 
#> $ANOVA$VarCom
#>            Estimates
#> VarR   0.00280811539
#> VarC   0.01965492882
#> VarTR  0.00048769055
#> VarTC  0.01796535211
#> VarRC  0.01559610252
#> VarErr 0.06677586505
#> 
#> $ANOVA$IndividualTrt
#>       DF        Trt1        Trt2
#> msR    8 0.561057769 0.210114459
#> msC   91 0.563067149 0.278841843
#> msRC 728 0.090956216 0.073787719
#> 
#> $ANOVA$IndividualRdr
#>      DF        rdr1        rdr2          rdr3        rdr4       rdr5
#> msT   1 0.048959308 0.316731553 1.0283409e-05 0.640144750 0.72761285
#> msC  91 0.155401358 0.069235322 1.5183444e-01 0.278912322 0.27074065
#> msTC 91 0.084330397 0.052667612 5.9205668e-02 0.098475734 0.15521211
#>             rdr6        rdr7        rdr8        rdr9
#> msT  0.018574407 0.051838662 0.137183240 0.083295609
#> msC  0.096225736 0.058295318 0.110097721 0.206446651
#> msTC 0.057541357 0.061413832 0.067745186 0.126079061
#> 
#> 
#> $RRRC
#> $RRRC$FTests
#>                  DF         MS     FStat           p
#> Treatment  1.000000 1.13120350 4.1385762 0.047467309
#> Error     47.951643 0.27333156        NA          NA
#> 
#> $RRRC$ciDiffTrt
#>              Estimate      StdErr        DF          t       PrGTt     CILower
#> trt1-trt2 -0.05227213 0.025694769 47.951643 -2.0343491 0.047467309 -0.10393627
#>                  CIUpper
#> trt1-trt2 -0.00060798802
#> 
#> $RRRC$ciAvgRdrEachTrt
#>        Estimate      StdErr        DF    CILower    CIUpper
#> trt1 0.84670344 0.035324047 27.127971 0.77424048 0.91916640
#> trt2 0.89897557 0.022392216 31.233962 0.85332021 0.94463094
#> 
#> 
#> $FRRC
#> $FRRC$FTests
#>           DF         MS     FStat           p
#> Treatment  1 1.13120350 4.9513417 0.028541895
#> Error     91 0.22846403        NA          NA
#> 
#> $FRRC$ciDiffTrt
#>              Estimate      StdErr DF          t       PrGTt      CILower
#> trt1-trt2 -0.05227213 0.023491392 91 -2.2251611 0.028541895 -0.098934894
#>                 CIUpper
#> trt1-trt2 -0.0056093661
#> 
#> $FRRC$ciAvgRdrEachTrt
#>        Estimate      StdErr DF    CILower    CIUpper
#> trt1 0.84670344 0.026077438 91 0.79490381 0.89850307
#> trt2 0.89897557 0.018351172 91 0.86252322 0.93542792
#> 
#> $FRRC$ciDiffTrtEachRdr
#>                       Estimate      StdErr DF            t       PrGTt
#> rdr1::trt1-trt2 -0.03262411348 0.042816697 91 -0.761948401 0.448061123
#> rdr2::trt1-trt2 -0.08297872340 0.033837082 91 -2.452301424 0.016102240
#> rdr3::trt1-trt2 -0.00047281324 0.035875894 91 -0.013179135 0.989513727
#> rdr4::trt1-trt2 -0.11796690307 0.046268529 91 -2.549614272 0.012456947
#> rdr5::trt1-trt2 -0.12576832151 0.058087660 91 -2.165146972 0.032992556
#> rdr6::trt1-trt2  0.02009456265 0.035368051 91  0.568155775 0.571328496
#> rdr7::trt1-trt2 -0.03356973995 0.036538792 91 -0.918742476 0.360659113
#> rdr8::trt1-trt2 -0.05460992908 0.038376053 91 -1.423021001 0.158149342
#> rdr9::trt1-trt2 -0.04255319149 0.052353120 91 -0.812810997 0.418447847
#>                      CILower      CIUpper
#> rdr1::trt1-trt2 -0.117674221  0.052425994
#> rdr2::trt1-trt2 -0.150191929 -0.015765518
#> rdr3::trt1-trt2 -0.071735869  0.070790242
#> rdr4::trt1-trt2 -0.209873653 -0.026060153
#> rdr5::trt1-trt2 -0.241152323 -0.010384320
#> rdr6::trt1-trt2 -0.050159727  0.090348852
#> rdr7::trt1-trt2 -0.106149561  0.039010081
#> rdr8::trt1-trt2 -0.130839244  0.021619386
#> rdr9::trt1-trt2 -0.146546233  0.061439850
#> 
#> 
#> $RRFC
#> $RRFC$FTests
#>           DF        MS     FStat           p
#> Treatment  1 1.1312035 10.132292 0.012936472
#> Error      8 0.1116434        NA          NA
#> 
#> $RRFC$ciDiffTrt
#>              Estimate      StdErr DF          t       PrGTt      CILower
#> trt1-trt2 -0.05227213 0.016421633  8 -3.1831262 0.012936472 -0.090140484
#>                CIUpper
#> trt1-trt2 -0.014403776
#> 
#> $RRFC$ciAvgRdrEachTrt
#>        Estimate      StdErr DF    CILower    CIUpper
#> trt1 0.84670344 0.026030866  8 0.78667616 0.90673073
#> trt2 0.89897557 0.015929891  8 0.86224118 0.93570997
#> 
#> 
StSignificanceTesting(dataset05, FOM = "SongA1", method = "DBM") 
#> $FOMs
#> $FOMs$foms
#>            rdr1       rdr2       rdr3       rdr4       rdr5       rdr6
#> trt1 0.86879433 0.89243499 0.88770686 0.72907801 0.72765957 0.92174941
#> trt2 0.90661939 0.97612293 0.89574468 0.84680851 0.83947991 0.91702128
#>            rdr7       rdr8       rdr9
#> trt1 0.91891253 0.88486998 0.82316785
#> trt2 0.95130024 0.93427896 0.86548463
#> 
#> $FOMs$trtMeans
#>        Estimate
#> trt1 0.85048595
#> trt2 0.90365117
#> 
#> $FOMs$trtMeanDiffs
#>               Estimate
#> trt1-trt2 -0.053165222
#> 
#> 
#> $ANOVA
#> $ANOVA$TRCanova
#>                 SS   DF          MS
#> T       1.17018790    1 1.170187902
#> R       5.07364865    8 0.634206082
#> C      52.96901221   91 0.582077057
#> TR      0.67747324    8 0.084684155
#> TC     21.35270676   91 0.234645129
#> RC     70.11562972  728 0.096312678
#> TRC    48.36423043  728 0.066434382
#> Total 199.72288892 1655          NA
#> 
#> $ANOVA$VarCom
#>            Estimates
#> VarR   0.00282415017
#> VarC   0.01764186846
#> VarTR  0.00019836709
#> VarTC  0.01869008298
#> VarRC  0.01493914786
#> VarErr 0.06643438246
#> 
#> $ANOVA$IndividualTrt
#>       DF        Trt1        Trt2
#> msR    8 0.517450975 0.201439261
#> msC   91 0.557364867 0.259357319
#> msRC 728 0.089158233 0.073588828
#> 
#> $ANOVA$IndividualRdr
#>      DF        rdr1        rdr2         rdr3       rdr4       rdr5         rdr6
#> msT   1 0.065813814 0.322168905 0.0029719051 0.63758161 0.57517418 0.0010283409
#> msC  91 0.128880350 0.069967830 0.1423505957 0.27928988 0.27313401 0.1004461284
#> msTC 91 0.087796745 0.053476379 0.0600038713 0.10036741 0.14682451 0.0551701902
#>             rdr7        rdr8        rdr9
#> msT  0.048252324 0.112297392 0.082372673
#> msC  0.059009429 0.106752317 0.192747936
#> msTC 0.062176139 0.066718794 0.133586147
#> 
#> 
#> $RRRC
#> $RRRC$FTests
#>                  DF        MS     FStat           p
#> Treatment  1.000000 1.1701879 4.6271708 0.034857042
#> Error     71.345374 0.2528949        NA          NA
#> 
#> $RRRC$ciDiffTrt
#>               Estimate      StdErr        DF         t       PrGTt     CILower
#> trt1-trt2 -0.053165222 0.024715526 71.345374 -2.151086 0.034857042 -0.10244245
#>                 CIUpper
#> trt1-trt2 -0.0038879957
#> 
#> $RRRC$ciAvgRdrEachTrt
#>        Estimate      StdErr        DF    CILower    CIUpper
#> trt1 0.85048595 0.034502286 29.027105 0.77992371 0.92104818
#> trt2 0.90365117 0.021625037 29.559007 0.85945931 0.94784303
#> 
#> 
#> $FRRC
#> $FRRC$FTests
#>           DF         MS     FStat           p
#> Treatment  1 1.17018790 4.9870539 0.027989516
#> Error     91 0.23464513        NA          NA
#> 
#> $FRRC$ciDiffTrt
#>               Estimate      StdErr DF          t       PrGTt   CILower
#> trt1-trt2 -0.053165222 0.023807051 91 -2.2331713 0.027989516 -0.100455
#>                 CIUpper
#> trt1-trt2 -0.0058754399
#> 
#> $FRRC$ciAvgRdrEachTrt
#>        Estimate      StdErr DF    CILower    CIUpper
#> trt1 0.85048595 0.025945057 91 0.79894928 0.90202262
#> trt2 0.90365117 0.017698403 91 0.86849547 0.93880687
#> 
#> $FRRC$ciDiffTrtEachRdr
#>                      Estimate      StdErr DF           t       PrGTt
#> rdr1::trt1-trt2 -0.0378250591 0.043687812 91 -0.86580347 0.388874868
#> rdr2::trt1-trt2 -0.0836879433 0.034095894 91 -2.45448744 0.016010857
#> rdr3::trt1-trt2 -0.0080378251 0.036116921 91 -0.22255012 0.824384408
#> rdr4::trt1-trt2 -0.1177304965 0.046710816 91 -2.52041191 0.013463974
#> rdr5::trt1-trt2 -0.1118203310 0.056496346 91 -1.97924889 0.050810497
#> rdr6::trt1-trt2  0.0047281324 0.034631661 91  0.13652629 0.891706828
#> rdr7::trt1-trt2 -0.0323877069 0.036764864 91 -0.88094185 0.380670077
#> rdr8::trt1-trt2 -0.0494089835 0.038084230 91 -1.29736071 0.197786500
#> rdr9::trt1-trt2 -0.0423167849 0.053889207 91 -0.78525528 0.434343606
#>                      CILower        CIUpper
#> rdr1::trt1-trt2 -0.124605530  0.04895541196
#> rdr2::trt1-trt2 -0.151415248 -0.01596063871
#> rdr3::trt1-trt2 -0.079779652  0.06370400179
#> rdr4::trt1-trt2 -0.210515794 -0.02494519895
#> rdr5::trt1-trt2 -0.224043383  0.00040272062
#> rdr6::trt1-trt2 -0.064063408  0.07351967309
#> rdr7::trt1-trt2 -0.105416592  0.04064117822
#> rdr8::trt1-trt2 -0.125058628  0.02624066076
#> rdr9::trt1-trt2 -0.149361075  0.06472750571
#> 
#> 
#> $RRFC
#> $RRFC$FTests
#>           DF          MS     FStat            p
#> Treatment  1 1.170187902 13.818263 0.0058942719
#> Error      8 0.084684155        NA           NA
#> 
#> $RRFC$ciDiffTrt
#>               Estimate      StdErr DF          t        PrGTt      CILower
#> trt1-trt2 -0.053165222 0.014302136  8 -3.7172924 0.0058942719 -0.086146007
#>                CIUpper
#> trt1-trt2 -0.020184437
#> 
#> $RRFC$ciAvgRdrEachTrt
#>        Estimate      StdErr DF    CILower    CIUpper
#> trt1 0.85048595 0.024998816  8 0.79283857 0.90813332
#> trt2 0.90365117 0.015597568  8 0.86768311 0.93961923
#> 
#> 
StSignificanceTesting(dataset05, FOM = "SongA2", method = "DBM") 
#> $FOMs
#> $FOMs$foms
#>            rdr1       rdr2       rdr3       rdr4       rdr5       rdr6
#> trt1 0.85248227 0.89078014 0.88250591 0.72411348 0.72482270 0.90330969
#> trt2 0.84278960 0.97565012 0.88061466 0.83735225 0.83262411 0.88037825
#>            rdr7       rdr8       rdr9
#> trt1 0.91678487 0.87825059 0.81702128
#> trt2 0.94940898 0.92340426 0.84539007
#> 
#> $FOMs$trtMeans
#>        Estimate
#> trt1 0.84334121
#> trt2 0.88529025
#> 
#> $FOMs$trtMeanDiffs
#>               Estimate
#> trt1-trt2 -0.041949041
#> 
#> 
#> $ANOVA
#> $ANOVA$TRCanova
#>                 SS   DF          MS
#> T       0.72852493    1 0.728524933
#> R       5.06973296    8 0.633716620
#> C      55.55549864   91 0.610499985
#> TR      0.93568277    8 0.116960346
#> TC     21.47190847   91 0.235955038
#> RC     76.97612318  728 0.105736433
#> TRC    55.57793192  728 0.076343313
#> Total 216.31540287 1655          NA
#> 
#> $ANOVA$VarCom
#>            Estimates
#> VarR   0.00264871279
#> VarC   0.01917510150
#> VarTR  0.00044148949
#> VarTC  0.01773463612
#> VarRC  0.01469655993
#> VarErr 0.07634331307
#> 
#> $ANOVA$IndividualTrt
#>       DF        Trt1       Trt2
#> msR    8 0.494860755 0.25581621
#> msC   91 0.567009247 0.27944578
#> msRC 728 0.093744766 0.08833498
#> 
#> $ANOVA$IndividualRdr
#>      DF         rdr1        rdr2          rdr3       rdr4       rdr5
#> msT   1 0.0043216024 0.331333993 0.00016453454 0.58985888 0.53457271
#> msC  91 0.1530192374 0.071322435 0.15120206785 0.27815432 0.27333141
#> msTC 91 0.1219576976 0.054532872 0.06559641050 0.10765930 0.14966359
#>             rdr6        rdr7        rdr8        rdr9
#> msT  0.024189148 0.048959308 0.093787256 0.037020271
#> msC  0.143353943 0.061070559 0.111764647 0.213172836
#> msTC 0.073227539 0.062126058 0.073579302 0.138358779
#> 
#> 
#> $RRRC
#> $RRRC$FTests
#>                  DF         MS     FStat          p
#> Treatment  1.000000 0.72852493 2.6341233 0.11161718
#> Error     44.733136 0.27657207        NA         NA
#> 
#> $RRRC$ciDiffTrt
#>               Estimate      StdErr        DF          t      PrGTt      CILower
#> trt1-trt2 -0.041949041 0.025846634 44.733136 -1.6229982 0.11161718 -0.094015407
#>               CIUpper
#> trt1-trt2 0.010117325
#> 
#> $RRRC$ciAvgRdrEachTrt
#>        Estimate      StdErr        DF    CILower    CIUpper
#> trt1 0.84334121 0.034194055 30.618721 0.77356676 0.91311566
#> trt2 0.88529025 0.023232884 24.417830 0.83738332 0.93319719
#> 
#> 
#> $FRRC
#> $FRRC$FTests
#>           DF         MS     FStat           p
#> Treatment  1 0.72852493 3.0875583 0.082256459
#> Error     91 0.23595504        NA          NA
#> 
#> $FRRC$ciDiffTrt
#>               Estimate     StdErr DF          t       PrGTt      CILower
#> trt1-trt2 -0.041949041 0.02387341 91 -1.7571449 0.082256459 -0.089370637
#>               CIUpper
#> trt1-trt2 0.005472555
#> 
#> $FRRC$ciAvgRdrEachTrt
#>        Estimate      StdErr DF    CILower    CIUpper
#> trt1 0.84334121 0.026168565 91 0.79136057 0.89532185
#> trt2 0.88529025 0.018371034 91 0.84879845 0.92178206
#> 
#> $FRRC$ciDiffTrtEachRdr
#>                      Estimate      StdErr DF            t       PrGTt
#> rdr1::trt1-trt2  0.0096926714 0.051490332 91  0.188242550 0.851105664
#> rdr2::trt1-trt2 -0.0848699764 0.034431050 91 -2.464925580 0.015580869
#> rdr3::trt1-trt2  0.0018912530 0.037762533 91  0.050082788 0.960166181
#> rdr4::trt1-trt2 -0.1132387707 0.048377882 91 -2.340713685 0.021431872
#> rdr5::trt1-trt2 -0.1078014184 0.057039953 91 -1.889928245 0.061950680
#> rdr6::trt1-trt2  0.0229314421 0.039898660 91  0.574742168 0.566883894
#> rdr7::trt1-trt2 -0.0326241135 0.036750054 91 -0.887729671 0.377026527
#> rdr8::trt1-trt2 -0.0451536643 0.039994375 91 -1.129000365 0.261865154
#> rdr9::trt1-trt2 -0.0283687943 0.054843409 91 -0.517268978 0.606223474
#>                      CILower       CIUpper
#> rdr1::trt1-trt2 -0.092586548  0.1119718907
#> rdr2::trt1-trt2 -0.153263028 -0.0164769249
#> rdr3::trt1-trt2 -0.073119381  0.0769018865
#> rdr4::trt1-trt2 -0.209335490 -0.0171420509
#> rdr5::trt1-trt2 -0.221104278  0.0055014413
#> rdr6::trt1-trt2 -0.056322344  0.1021852278
#> rdr7::trt1-trt2 -0.105623581  0.0403753545
#> rdr8::trt1-trt2 -0.124597578  0.0342902491
#> rdr9::trt1-trt2 -0.137308490  0.0805709011
#> 
#> 
#> $RRFC
#> $RRFC$FTests
#>           DF         MS     FStat         p
#> Treatment  1 0.72852493 6.2288199 0.0371869
#> Error      8 0.11696035        NA        NA
#> 
#> $RRFC$ciDiffTrt
#>               Estimate     StdErr DF          t     PrGTt      CILower
#> trt1-trt2 -0.041949041 0.01680812  8 -2.4957604 0.0371869 -0.080708637
#>                CIUpper
#> trt1-trt2 -0.003189446
#> 
#> $RRFC$ciAvgRdrEachTrt
#>        Estimate      StdErr DF    CILower    CIUpper
#> trt1 0.84334121 0.024447043  8 0.78696623 0.89971620
#> trt2 0.88529025 0.017577166  8 0.84475724 0.92582327
#> 
#> 
StSignificanceTesting(dataset05, FOM = "wAFROC1", method = "DBM")
#> $FOMs
#> $FOMs$foms
#>            rdr1       rdr2       rdr3       rdr4       rdr5       rdr6
#> trt1 0.74277290 0.88222710 0.78902636 0.64594897 0.63787388 0.81076549
#> trt2 0.80974406 0.96927999 0.87208989 0.76034921 0.75580096 0.87207061
#>            rdr7       rdr8       rdr9
#> trt1 0.87806429 0.81024514 0.77515032
#> trt2 0.94320459 0.88430851 0.83356460
#> 
#> $FOMs$trtMeans
#>        Estimate
#> trt1 0.77467494
#> trt2 0.85560138
#> 
#> $FOMs$trtMeanDiffs
#>              Estimate
#> trt1-trt2 -0.08092644
#> 
#> 
#> $ANOVA
#> $ANOVA$TRCanova
#>                 SS   DF          MS
#> T       2.71132271    1 2.711322708
#> R       9.47251796    8 1.184064745
#> C      63.90117742   91 0.702210741
#> TR      0.18006655    8 0.022508319
#> TC     29.38566701   91 0.322919418
#> RC     71.58323438  728 0.098328619
#> TRC    44.01195423  728 0.060455981
#> Total 221.24594026 1655          NA
#> 
#> $ANOVA$VarCom
#>             Estimates
#> VarR    0.00610697711
#> VarC    0.01896770476
#> VarTR  -0.00041247459
#> VarTC   0.02916260407
#> VarRC   0.01893631879
#> VarErr  0.06045598108
#> 
#> $ANOVA$IndividualTrt
#>       DF        Trt1        Trt2
#> msR    8 0.705334435 0.501238629
#> msC   91 0.657336677 0.367793482
#> msRC 728 0.084660507 0.074124093
#> 
#> $ANOVA$IndividualRdr
#>      DF       rdr1        rdr2        rdr3       rdr4       rdr5        rdr6
#> msT   1 0.20631632 0.348597405 0.317379235 0.60202116 0.63971256 0.172882609
#> msC  91 0.25566141 0.068830839 0.164789008 0.22215560 0.27945815 0.133831185
#> msTC 91 0.12257664 0.043919858 0.079477558 0.12843517 0.12317862 0.088443061
#>             rdr7       rdr8        rdr9
#> msT  0.195189913 0.25232759 0.156962476
#> msC  0.063285806 0.15760943 0.143218267
#> msTC 0.063554079 0.07241945 0.084562819
#> 
#> 
#> $RRRC
#> $RRRC$FTests
#>                  DF         MS     FStat            p
#> Treatment    1.0000 2.71132271 9.5143559 0.0020822949
#> Error     1282.3527 0.28497176        NA           NA
#> 
#> $RRRC$ciDiffTrt
#>              Estimate      StdErr        DF         t        PrGTt   CILower
#> trt1-trt2 -0.08092644 0.026236188 1282.3527 -3.084535 0.0020822949 -0.132397
#>                CIUpper
#> trt1-trt2 -0.029455875
#> 
#> $RRRC$ciAvgRdrEachTrt
#>        Estimate      StdErr        DF    CILower    CIUpper
#> trt1 0.77467494 0.039287289 26.264475 0.69395832 0.85539156
#> trt2 0.85560138 0.030984413 20.120310 0.79099379 0.92020897
#> 
#> 
#> $FRRC
#> $FRRC$FTests
#>           DF         MS     FStat            p
#> Treatment  1 2.71132271 8.3962827 0.0047100297
#> Error     91 0.32291942        NA           NA
#> 
#> $FRRC$ciDiffTrt
#>              Estimate      StdErr DF         t        PrGTt     CILower
#> trt1-trt2 -0.08092644 0.027928455 91 -2.897634 0.0047100297 -0.13640288
#>                CIUpper
#> trt1-trt2 -0.025449995
#> 
#> $FRRC$ciAvgRdrEachTrt
#>        Estimate      StdErr DF    CILower    CIUpper
#> trt1 0.77467494 0.028175963 91 0.71870685 0.83064303
#> trt2 0.85560138 0.021075935 91 0.81373663 0.89746614
#> 
#> $FRRC$ciDiffTrtEachRdr
#>                     Estimate      StdErr DF          t        PrGTt     CILower
#> rdr1::trt1-trt2 -0.066971169 0.051620826 91 -1.2973672 0.1977842596 -0.16950960
#> rdr2::trt1-trt2 -0.087052883 0.030899507 91 -2.8172904 0.0059398084 -0.14843096
#> rdr3::trt1-trt2 -0.083063521 0.041566489 91 -1.9983290 0.0486657734 -0.16563025
#> rdr4::trt1-trt2 -0.114400247 0.052840032 91 -2.1650298 0.0330018019 -0.21936048
#> rdr5::trt1-trt2 -0.117927074 0.051747427 91 -2.2788974 0.0250121886 -0.22071698
#> rdr6::trt1-trt2 -0.061305119 0.043848321 91 -1.3981178 0.1654761161 -0.14840442
#> rdr7::trt1-trt2 -0.065140302 0.037170020 91 -1.7524957 0.0830568113 -0.13897398
#> rdr8::trt1-trt2 -0.074063367 0.039677901 91 -1.8666150 0.0651755767 -0.15287864
#> rdr9::trt1-trt2 -0.058414277 0.042875659 91 -1.3624112 0.1764306854 -0.14358151
#>                        CIUpper
#> rdr1::trt1-trt2  0.03556726026
#> rdr2::trt1-trt2 -0.02567480868
#> rdr3::trt1-trt2 -0.00049679751
#> rdr4::trt1-trt2 -0.00944001359
#> rdr5::trt1-trt2 -0.01513716778
#> rdr6::trt1-trt2  0.02579418408
#> rdr7::trt1-trt2  0.00869337705
#> rdr8::trt1-trt2  0.00475190967
#> rdr9::trt1-trt2  0.02675295290
#> 
#> 
#> $RRFC
#> $RRFC$FTests
#>           DF          MS     FStat             p
#> Treatment  1 2.711322708 120.45869 4.2196838e-06
#> Error      8 0.022508319        NA            NA
#> 
#> $RRFC$ciDiffTrt
#>              Estimate       StdErr DF          t         PrGTt     CILower
#> trt1-trt2 -0.08092644 0.0073734606  8 -10.975368 4.2196838e-06 -0.09792967
#>                CIUpper
#> trt1-trt2 -0.063923209
#> 
#> $RRFC$ciAvgRdrEachTrt
#>        Estimate      StdErr DF    CILower    CIUpper
#> trt1 0.77467494 0.029186524  8 0.70737070 0.84197919
#> trt2 0.85560138 0.024604078  8 0.79886427 0.91233849
#> 
#> 
StSignificanceTesting(dataset05, FOM = "AFROC1", method = "DBM")
#> $FOMs
#> $FOMs$foms
#>            rdr1       rdr2       rdr3       rdr4       rdr5       rdr6
#> trt1 0.74714444 0.88568534 0.79412307 0.65521371 0.63697494 0.81180914
#> trt2 0.79707074 0.96361459 0.87693441 0.74714444 0.75469786 0.86726234
#>            rdr7       rdr8       rdr9
#> trt1 0.88190862 0.81457259 0.77846352
#> trt2 0.93671702 0.88494842 0.82212601
#> 
#> $FOMs$trtMeans
#>        Estimate
#> trt1 0.77843282
#> trt2 0.85005732
#> 
#> $FOMs$trtMeanDiffs
#>               Estimate
#> trt1-trt2 -0.071624498
#> 
#> 
#> $ANOVA
#> $ANOVA$TRCanova
#>                 SS   DF          MS
#> T       2.12384848    1 2.123848476
#> R       9.63346477    8 1.204183097
#> C      53.09534053   91 0.583465281
#> TR      0.20703861    8 0.025879826
#> TC     22.15400366   91 0.243450590
#> RC     64.26551399  728 0.088276805
#> TRC    42.56495922  728 0.058468351
#> Total 194.04416926 1655          NA
#> 
#> $ANOVA$VarCom
#>            Estimates
#> VarR    0.0062418197
#> VarC    0.0172336798
#> VarTR  -0.0003542231
#> VarTC   0.0205535821
#> VarRC   0.0149042272
#> VarErr  0.0584683506
#> 
#> $ANOVA$IndividualTrt
#>       DF       Trt1        Trt2
#> msR    8 0.70206158 0.528001343
#> msC   91 0.49992187 0.326994004
#> msRC 728 0.07299324 0.073751915
#> 
#> $ANOVA$IndividualRdr
#>      DF       rdr1        rdr2        rdr3       rdr4       rdr5        rdr6
#> msT   1 0.11466127 0.279356569 0.315455095 0.38875792 0.63749953 0.141452669
#> msC  91 0.21160764 0.075256727 0.136766482 0.17495482 0.21188376 0.118382593
#> msTC 91 0.12245599 0.031792892 0.064034987 0.12527849 0.10769078 0.077798308
#>             rdr7        rdr8        rdr9
#> msT  0.138182197 0.227826836 0.087695003
#> msC  0.062966692 0.123387440 0.174473570
#> msTC 0.042387980 0.059576966 0.080180998
#> 
#> 
#> $RRRC
#> $RRRC$FTests
#>                 DF         MS     FStat            p
#> Treatment   1.0000 2.12384848 10.072217 0.0015923932
#> Error     531.0852 0.21086206        NA           NA
#> 
#> $RRRC$ciDiffTrt
#>               Estimate      StdErr       DF          t        PrGTt     CILower
#> trt1-trt2 -0.071624498 0.022568311 531.0852 -3.1736756 0.0015923932 -0.11595861
#>                CIUpper
#> trt1-trt2 -0.027290387
#> 
#> $RRRC$ciAvgRdrEachTrt
#>        Estimate      StdErr        DF    CILower    CIUpper
#> trt1 0.77843282 0.036925800 20.688067 0.70157085 0.85529479
#> trt2 0.85005732 0.030716945 17.514296 0.78539489 0.91471974
#> 
#> 
#> $FRRC
#> $FRRC$FTests
#>           DF         MS     FStat            p
#> Treatment  1 2.12384848 8.7239406 0.0039964818
#> Error     91 0.24345059        NA           NA
#> 
#> $FRRC$ciDiffTrt
#>               Estimate      StdErr DF          t        PrGTt     CILower
#> trt1-trt2 -0.071624498 0.024249637 91 -2.9536318 0.0039964818 -0.11979342
#>                CIUpper
#> trt1-trt2 -0.023455573
#> 
#> $FRRC$ciAvgRdrEachTrt
#>        Estimate      StdErr DF    CILower    CIUpper
#> trt1 0.77843282 0.024571739 91 0.72962408 0.82724156
#> trt2 0.85005732 0.019872602 91 0.81058283 0.88953180
#> 
#> $FRRC$ciDiffTrtEachRdr
#>                     Estimate      StdErr DF          t        PrGTt     CILower
#> rdr1::trt1-trt2 -0.049926308 0.051595414 91 -0.9676501 0.3357838932 -0.15241426
#> rdr2::trt1-trt2 -0.077929256 0.026289728 91 -2.9642473 0.0038730163 -0.13015057
#> rdr3::trt1-trt2 -0.082811349 0.037310387 91 -2.2195253 0.0289362274 -0.15692385
#> rdr4::trt1-trt2 -0.091930730 0.052186641 91 -1.7615759 0.0814995986 -0.19559308
#> rdr5::trt1-trt2 -0.117722918 0.048384955 91 -2.4330480 0.0169274354 -0.21383369
#> rdr6::trt1-trt2 -0.055453206 0.041125024 91 -1.3484054 0.1808745513 -0.13714301
#> rdr7::trt1-trt2 -0.054808401 0.030355853 91 -1.8055299 0.0742984922 -0.11510657
#> rdr8::trt1-trt2 -0.070375829 0.035988213 91 -1.9555244 0.0535890418 -0.14186199
#> rdr9::trt1-trt2 -0.043662491 0.041750032 91 -1.0458074 0.2984199863 -0.12659380
#>                       CIUpper
#> rdr1::trt1-trt2  0.0525616442
#> rdr2::trt1-trt2 -0.0257079398
#> rdr3::trt1-trt2 -0.0086988491
#> rdr4::trt1-trt2  0.0117316225
#> rdr5::trt1-trt2 -0.0216121496
#> rdr6::trt1-trt2  0.0262366019
#> rdr7::trt1-trt2  0.0054897723
#> rdr8::trt1-trt2  0.0011103343
#> rdr9::trt1-trt2  0.0392688190
#> 
#> 
#> $RRFC
#> $RRFC$FTests
#>           DF          MS     FStat             p
#> Treatment  1 2.123848476 82.065795 1.7661339e-05
#> Error      8 0.025879826        NA            NA
#> 
#> $RRFC$ciDiffTrt
#>               Estimate       StdErr DF          t         PrGTt      CILower
#> trt1-trt2 -0.071624498 0.0079064313  8 -9.0590173 1.7661339e-05 -0.089856762
#>                CIUpper
#> trt1-trt2 -0.053392235
#> 
#> $RRFC$ciAvgRdrEachTrt
#>        Estimate      StdErr DF    CILower    CIUpper
#> trt1 0.77843282 0.029118730  8 0.71128490 0.84558073
#> trt2 0.85005732 0.025252382  8 0.79182522 0.90828941
#> 
#> 
StSignificanceTesting(dataset05, FOM = "AFROC", method = "DBM")
#> $FOMs
#> $FOMs$foms
#>            rdr1       rdr2       rdr3       rdr4       rdr5       rdr6
#> trt1 0.73013183 0.88436911 0.81412429 0.62485876 0.57890772 0.85065913
#> trt2 0.78926554 0.96290019 0.85122411 0.73841808 0.71977401 0.86704331
#>           rdr7       rdr8       rdr9
#> trt1 0.8960452 0.80734463 0.76760829
#> trt2 0.9306968 0.89962335 0.80772128
#> 
#> $FOMs$trtMeans
#>        Estimate
#> trt1 0.77267211
#> trt2 0.84074074
#> 
#> $FOMs$trtMeanDiffs
#>               Estimate
#> trt1-trt2 -0.068068634
#> 
#> 
#> $ANOVA
#> $ANOVA$TRCanova
#>                 SS   DF          MS
#> T       1.91820230    1 1.918202297
#> R      13.54411580    8 1.693014475
#> C      57.88453619   91 0.636093804
#> TR      0.62895302    8 0.078619127
#> TC     26.57493585   91 0.292032262
#> RC     80.71706011  728 0.110875083
#> TRC    51.76949055  728 0.071111938
#> Total 233.03729382 1655          NA
#> 
#> $ANOVA$VarCom
#>            Estimates
#> VarR   8.5577837e-03
#> VarC   1.6905467e-02
#> VarTR  8.1599889e-05
#> VarTC  2.4546703e-02
#> VarRC  1.9881573e-02
#> VarErr 7.1111938e-02
#> 
#> $ANOVA$IndividualTrt
#>       DF        Trt1        Trt2
#> msR    8 1.126175756 0.645457847
#> msC   91 0.583592460 0.344533607
#> msRC 728 0.089804085 0.092182935
#> 
#> $ANOVA$IndividualRdr
#>      DF       rdr1        rdr2        rdr3       rdr4       rdr5        rdr6
#> msT   1 0.16085260 0.283687957 0.063314217 0.59320310 0.91279234 0.012348303
#> msC  91 0.23221438 0.080045517 0.164567266 0.22547486 0.25205361 0.128630322
#> msTC 91 0.13325647 0.040085514 0.079211449 0.11781107 0.13311867 0.103095166
#>             rdr7        rdr8        rdr9
#> msT  0.055233738 0.391706654 0.074016407
#> msC  0.069804069 0.124260614 0.246043828
#> msTC 0.045446884 0.072216896 0.136685641
#> 
#> 
#> $RRRC
#> $RRRC$FTests
#>                  DF         MS     FStat           p
#> Treatment   1.00000 1.91820230 6.4038386 0.012725723
#> Error     116.12925 0.29953945        NA          NA
#> 
#> $RRRC$ciDiffTrt
#>               Estimate      StdErr        DF          t       PrGTt     CILower
#> trt1-trt2 -0.068068634 0.026898425 116.12925 -2.5305807 0.012725723 -0.12134373
#>                CIUpper
#> trt1-trt2 -0.014793539
#> 
#> $RRRC$ciAvgRdrEachTrt
#>        Estimate      StdErr        DF    CILower    CIUpper
#> trt1 0.77267211 0.044232097 16.553447 0.67915840 0.86618582
#> trt2 0.84074074 0.032928860 15.478243 0.77074294 0.91073854
#> 
#> 
#> $FRRC
#> $FRRC$FTests
#>           DF         MS     FStat          p
#> Treatment  1 1.91820230 6.5684602 0.01202163
#> Error     91 0.29203226        NA         NA
#> 
#> $FRRC$ciDiffTrt
#>               Estimate      StdErr DF          t      PrGTt     CILower
#> trt1-trt2 -0.068068634 0.026559216 91 -2.5629007 0.01202163 -0.12082525
#>                CIUpper
#> trt1-trt2 -0.015312014
#> 
#> $FRRC$ciAvgRdrEachTrt
#>        Estimate      StdErr DF    CILower    CIUpper
#> trt1 0.77267211 0.026548480 91 0.71993681 0.82540740
#> trt2 0.84074074 0.020398612 91 0.80022140 0.88126008
#> 
#> $FRRC$ciDiffTrtEachRdr
#>                     Estimate      StdErr DF           t        PrGTt
#> rdr1::trt1-trt2 -0.059133710 0.053822670 91 -1.09867663 0.2748074951
#> rdr2::trt1-trt2 -0.078531073 0.029519895 91 -2.66027615 0.0092275229
#> rdr3::trt1-trt2 -0.037099812 0.041496844 91 -0.89403937 0.3736593002
#> rdr4::trt1-trt2 -0.113559322 0.050607413 91 -2.24392666 0.0272625766
#> rdr5::trt1-trt2 -0.140866290 0.053794834 91 -2.61858399 0.0103425503
#> rdr6::trt1-trt2 -0.016384181 0.047341306 91 -0.34608637 0.7300766851
#> rdr7::trt1-trt2 -0.034651601 0.031432082 91 -1.10242781 0.2731828185
#> rdr8::trt1-trt2 -0.092278719 0.039622374 91 -2.32895483 0.0220757793
#> rdr9::trt1-trt2 -0.040112994 0.054510797 91 -0.73587246 0.4637008986
#>                      CILower      CIUpper
#> rdr1::trt1-trt2 -0.166045833  0.047778413
#> rdr2::trt1-trt2 -0.137168719 -0.019893428
#> rdr3::trt1-trt2 -0.119528194  0.045328570
#> rdr4::trt1-trt2 -0.214084730 -0.013033914
#> rdr5::trt1-trt2 -0.247723119 -0.034009461
#> rdr6::trt1-trt2 -0.110421869  0.077653508
#> rdr7::trt1-trt2 -0.097087569  0.027784368
#> rdr8::trt1-trt2 -0.170983698 -0.013573741
#> rdr9::trt1-trt2 -0.148391996  0.068166007
#> 
#> 
#> $RRFC
#> $RRFC$FTests
#>           DF          MS     FStat           p
#> Treatment  1 1.918202297 24.398672 0.001135776
#> Error      8 0.078619127        NA          NA
#> 
#> $RRFC$ciDiffTrt
#>               Estimate      StdErr DF          t       PrGTt      CILower
#> trt1-trt2 -0.068068634 0.013780467  8 -4.9395012 0.001135776 -0.099846448
#>                CIUpper
#> trt1-trt2 -0.036290819
#> 
#> $RRFC$ciAvgRdrEachTrt
#>        Estimate      StdErr DF    CILower    CIUpper
#> trt1 0.77267211 0.036879746  8 0.68762726 0.85771695
#> trt2 0.84074074 0.027920216  8 0.77635661 0.90512487
#> 
#> 
# }