Perform DBM or OR significance testing

Performs Dorfman-Berbaum-Metz (DBM) or Obuchowski-Rockette (OR) significance testing (with Hillis' improvements), for specified dataset; significance testing refers to analysis designed to assign a P-value, and other statistics, for rejecting the null hypothesis (NH) that the reader-averaged figure of merit (FOM) difference between treatments is zero. The results of the analysis are better visualized in the text or Excel-formatted files produced by UtilOutputReport.

StSignificanceTesting(dataset, FOM, FPFValue = 0.2, alpha = 0.05,
  method = "DBMH", covEstMethod = "Jackknife", nBoots = 200,
  option = "ALL", tempOrgCode = FALSE)

Arguments

dataset
The dataset to be analyzed, see RJafroc-package. Must have two or more treatments and two or more readers.

FOM
The figure of merit, see UtilFigureOfMerit

FPFValue
Only needed for LROC data and FOM = "PCL" or "ALROC"; where to evaluate a partial curve based figure of merit. The default is 0.2.

alpha
The significance level of the test of the null hypothesis that all treatment effects are zero; the default is 0.05

method
The significance testing method to be used. There are two options: "DBMH" (the default) or "ORH", representing the Dorfman-Berbaum-Metz and the Obuchowski-Rockette significance testing methods, respectively.

covEstMethod
The covariance matrix estimation method in ORH analysis (for method = "DBMH" the jackknife is always used).

"Jackknife", the default,

"Bootstrap", in which case nBoots (above) is relevant

"DeLong"; requires FOM = "Wilcoxon", otherwise an error results.

nBoots
The number of bootstraps (defaults to 200), relevant only if covEstMethod = "Bootstrap" and method = "ORH"

option
Determines which factors are regarded as random vs. fixed:

"RRRC" = random-reader random case,

"FRRC" = fixed-reader random case,

"RRFC" = random-reader fixed case,

"ALL" = outputs the results of "RRRC", "FRRC" and "RRFC" analyses

tempOrgCode,
default FALSE; if TRUE, then code from version 0.0.1 of RJafroc is used (see RJafroc_0.0.1.tar). This is intended to check against errors that crept in subsequent to the original version as I attempted to improve the organization of the code and the output. As implicit in the name of this temporary flag, it will eventually be removed.

Value

For method = "DBMH" the returned list has 15 members:

fomArray

Figure of merit array: see return of UtilFigureOfMerit

anovaY

ANOVA table of the pseudovalues, over all treatments

anovaYi

ANOVA table of the pseudovalues, for each treatment

varComp

The variance components of the DBM pseudovalue mode; 6 values, in the following order: varR, varC, varTR, varTC, varRC and varErr

FTestStatsRRRC

Results of the F-test for RRRC random reader random case analysis; contains the following items: fRRRC - the value of the F-statistic, ndfRRRC - the numerator degrees of freedom, ddfRRRC - the denominator degrees of freedom and pRRRC - the pvalue

ciDiffTrtRRRC

For RRRC analysis, the confidence intervals and related statistics for the FOM differences between pairs of treatments

ciAvgRdrEachTrtRRRC

For RRRC analysis, the confidence intervals and related test statistics for rdr. avg. FOM in each treatment

FTestStatsFRRC

As for FTestStatsRRRC except that this is for fixed-reader random-case (FRRC) analysis

ciDiffTrtFRRC

For FRRC analysis, the confidence intervals and related test statistics for the FOM differences between pairs of treatments

ciAvgRdrEachTrtFRRC

For FRRC analysis, the confidence intervals and related tests for rdr. avg. FOM in each treatment

msAnovaEachRdrFRRC

The mean squares table of the ANOVA of the pseudovalues for each reader (based on data for the specified reader)

ciDiffTrtEachRdrFRRC

The confidence intervals and related tests of the FOM differences between pairs of treatments for each reader

FTestStatsRRFC

As for FTestStatsRRRC except that this is for random-reader fixed-case (RRFC) analysis

ciDiffTrtRRFC

For RRFC analysis, the confidence intervals and related test statistics for the FOM differences between pairs of treatments

ciAvgRdrEachTrtRRFC

For RRFC analysis, the confidence intervals and related tests for reader averaged FOM in each treatment

For method = "ORH" the return list has 14 members:
fomArray

Figure of merit array: see return of UtilFigureOfMerit

meanSquares

List with 3 members: msT, msR, msTR

varComp

The variance components of the OR figure of merit model; 6 values, listed in the following order: varR, varTR, cov1, cov2, cov3 and var

FTestStatsRRRC

Results of the F-test for RRRC random reader random case analysis; contains the following items: fRRRC - the value of the F-statistic, ndfRRRC - the numerator degrees of freedom, ddfRRRC - the denominator degrees of freedom and pRRRC - the pvalue

ciDiffTrtRRRC

Same as in DBMH method

ciAvgRdrEachTrtRRRC

Same as in DBMH method

FTestStatsFRRC

Same as FTestStatsRRRC, but now treating reader as a fixed effect

ciDiffTrtFRRC

Same as in DBMH method

ciAvgRdrEachTrtFRRC

Same as in DBMH method

ciDiffTrtEachRdrFRRC

Same as in DBMH method

varCovEachRdr

Var and Cov1 estimates for each reader

FTestStatsRRFC

Same as FTestStatsRRRC, but now treating case as a fixed effect

ciDiffTrtRRFC

Same as in DBMH method

ciAvgRdrEachTrtRRFC

Same as in DBMH method

References

Dorfman DD, Berbaum KS, Metz CE (1992) ROC characteristic rating analysis: Generalization to the Population of Readers and Patients with the Jackknife method, Invest. Radiol. 27, 723-731.

Obuchowski NA, Rockette HE (1995) Hypothesis Testing of the Diagnostic Accuracy for Multiple Diagnostic Tests: An ANOVA Approach with Dependent Observations, Communications in Statistics: Simulation and Computation 24, 285-308.

Hillis SL (2014) A marginal-mean ANOVA approach for analyzing multireader multicase radiological imaging data, Statistics in medicine 33, 330-360.

Chakraborty DP (2017) Observer Performance Methods for Diagnostic Imaging - Foundations, Modeling, and Applications with R-Based Examples, CRC Press, Boca Raton, FL. https://www.crcpress.com/Observer-Performance-Methods-for-Diagnostic-Imaging-Foundations-Modeling/Chakraborty/p/book/9781482214840

Examples

StSignificanceTesting(dataset02,FOM = "Wilcoxon", method = "DBMH")

#> $fomArray
#>           Rdr0      Rdr1      Rdr2      Rdr3      Rdr4
#> Trt0 0.9196457 0.8587762 0.9038647 0.9731079 0.8297907
#> Trt1 0.9478261 0.9053140 0.9217391 0.9993559 0.9299517
#> 
#> $anovaY
#>   Source          SS   DF         MS
#> 1      T   0.5467634    1 0.54676344
#> 2      R   1.7493072    4 0.43732680
#> 3      C  44.8462969  113 0.39686988
#> 4     TR   0.2512700    4 0.06281749
#> 5     TC  11.2828335  113 0.09984808
#> 6     RC  29.1544793  452 0.06450106
#> 7    TRC  18.0671646  452 0.03997160
#> 8  Total 105.8981150 1139         NA
#> 
#> $anovaYi
#>   Source  DF       Trt0       Trt1
#> 1      R   4 0.35141967 0.14872462
#> 2      C 113 0.33629430 0.16042367
#> 3     RC 452 0.06043607 0.04403659
#> 
#> $varComp
#>          varR       varC        varTR     varTC      varRC    varErr
#> 1 0.001534999 0.02724923 0.0002004025 0.0119753 0.01226473 0.0399716
#> 
#> $FTestStatsRRRC
#>      fRRRC ndfRRRC  ddfRRRC      pRRRC
#> 1 4.456319       1 15.25967 0.05166569
#> 
#> $ciDiffTrtRRRC
#>     TrtDiff    Estimate     StdErr       DF         t      PrGTt    CILower
#> 1 Trt0-Trt1 -0.04380032 0.02074862 15.25967 -2.110999 0.05166569 -0.0879595
#>        CIUpper
#> 1 0.0003588544
#> 
#> $ciAvgRdrEachTrtRRRC
#>   Treatment      Area     StdErr       DF   CILower   CIUpper
#> 1      Trt0 0.8970370 0.03317360 12.74465 0.8252236 0.9688505
#> 2      Trt1 0.9408374 0.02156637 12.71019 0.8941378 0.9875369
#> 
#> $FTestStatsFRRC
#>      fFRRC ndfFRRC ddfFRRC      pFRRC
#> 1 5.475953       1     113 0.02103497
#> 
#> $ciDiffTrtFRRC
#>   Treatment    Estimate     StdErr  DF         t      PrGTt     CILower
#> 1 Trt0-Trt1 -0.04380032 0.01871748 113 -2.340075 0.02103497 -0.08088303
#>        CIUpper
#> 1 -0.006717613
#> 
#> $ciAvgRdrEachTrtFRRC
#>   Treatment      Area     StdErr  DF   CILower   CIUpper
#> 1      Trt0 0.8970370 0.02428971 113 0.8489147 0.9451593
#> 2      Trt1 0.9408374 0.01677632 113 0.9076004 0.9740743
#> 
#> $msAnovaEachRdrFRRC
#>   Source  DF       Rdr0       Rdr1       Rdr2       Rdr3      Rdr4
#> 1      T   1 0.04526554 0.12344893 0.01821116 0.03927054 0.5718372
#> 2      C 113 0.12224980 0.21274824 0.13653391 0.01728937 0.1660528
#> 3     TC 113 0.03709953 0.03943181 0.05552040 0.01704235 0.1106404
#> 
#> $ciDiffTrtEachRdrFRRC
#>   Reader Treatment    Estimate     StdErr  DF          t      PrGTt     CILower
#> 1   Rdr0 Trt0-Trt1 -0.02818035 0.02551213 113 -1.1045864 0.27168532 -0.07872449
#> 2   Rdr1 Trt0-Trt1 -0.04653784 0.02630183 113 -1.7693768 0.07952866 -0.09864651
#> 3   Rdr2 Trt0-Trt1 -0.01787440 0.03120965 113 -0.5727202 0.56797197 -0.07970634
#> 4   Rdr3 Trt0-Trt1 -0.02624799 0.01729129 113 -1.5179891 0.13180998 -0.06050515
#> 5   Rdr4 Trt0-Trt1 -0.10016103 0.04405746 113 -2.2734182 0.02489049 -0.18744681
#>        CIUpper
#> 1  0.022363782
#> 2  0.005570821
#> 3  0.043957543
#> 4  0.008009174
#> 5 -0.012875255
#> 
#> $FTestStatsRRFC
#>   fRRFC ndfRRFC ddfRRFC      pRRFC
#> 1 8.704       1       4 0.04195875
#> 
#> $ciDiffTrtRRFC
#>   Treatment    Estimate     StdErr DF         t      PrGTt     CILower
#> 1 Trt0-Trt1 -0.04380032 0.01484629  4 -2.950254 0.04195875 -0.08502022
#>       CIUpper
#> 1 -0.00258042
#> 
#> $ciAvgRdrEachTrtRRFC
#>   Treatment      Area     StdErr DF   CILower   CIUpper
#> 1      Trt0 0.8970370 0.02482994  4 0.8280981 0.9659760
#> 2      Trt1 0.9408374 0.01615303  4 0.8959894 0.9856854
#> 

StSignificanceTesting(dataset02,FOM = "Wilcoxon", method = "ORH")

#> $fomArray
#>           Rdr0      Rdr1      Rdr2      Rdr3      Rdr4
#> Trt0 0.9196457 0.8587762 0.9038647 0.9731079 0.8297907
#> Trt1 0.9478261 0.9053140 0.9217391 0.9993559 0.9299517
#> 
#> $meanSquares
#>           msT       msR         msTR
#> 1 0.004796171 0.0038362 0.0005510306
#> 
#> $varComp
#>          varR        varTR         cov1         cov2         cov3          var
#> 1 0.001534999 0.0002004025 0.0003466137 0.0003440748 0.0002390284 0.0008022883
#> 
#> $FTestStatsRRRC
#>      fRRRC ndfRRRC  ddfRRRC      pRRRC
#> 1 4.456319       1 15.25967 0.05166569
#> 
#> $ciDiffTrtRRRC
#>   Treatment    Estimate     StdErr       DF         t      PrGTt    CILower
#> 1 Trt0-Trt1 -0.04380032 0.02074862 15.25967 -2.110999 0.05166569 -0.0879595
#>        CIUpper
#> 1 0.0003588544
#> 
#> $ciAvgRdrEachTrtRRRC
#>   Treatment      Area     StdErr       DF   CILower   CIUpper
#> 1      Trt0 0.8970370 0.03317360 12.74465 0.8252236 0.9688505
#> 2      Trt1 0.9408374 0.02156637 12.71019 0.8941378 0.9875369
#> 
#> $FTestStatsFRRC
#>      fFRRC ndfFRRC ddfFRRC      pFRRC
#> 1 5.475953       1     Inf 0.01927984
#> 
#> $ciDiffTrtFRRC
#>   Treatment    Estimate     StdErr  DF         t      PrGTt     CILower
#> 1 Trt0-Trt1 -0.04380032 0.01871748 Inf -2.340075 0.01927984 -0.08048591
#>       CIUpper
#> 1 -0.00711473
#> 
#> $ciAvgRdrEachTrtFRRC
#>   Treatment      Area     StdErr  DF   CILower   CIUpper
#> 1      Trt0 0.8970370 0.02428971 Inf 0.8494301 0.9446440
#> 2      Trt1 0.9408374 0.01677632 Inf 0.9079564 0.9737183
#> 
#> $ciDiffTrtEachRdrFRRC
#>   Reader Treatment    Estimate     StdErr  DF          t      PrGTt     CILower
#> 1   Rdr0 Trt0-Trt1 -0.02818035 0.02551213 Inf -1.1045864 0.26933885 -0.07818322
#> 2   Rdr1 Trt0-Trt1 -0.04653784 0.02630183 Inf -1.7693768 0.07683102 -0.09808848
#> 3   Rdr2 Trt0-Trt1 -0.01787440 0.03120965 Inf -0.5727202 0.56683414 -0.07904418
#> 4   Rdr3 Trt0-Trt1 -0.02624799 0.01729129 Inf -1.5179891 0.12901715 -0.06013829
#> 5   Rdr4 Trt0-Trt1 -0.10016103 0.04405746 Inf -2.2734182 0.02300099 -0.18651207
#>        CIUpper
#> 1  0.021822507
#> 2  0.005012792
#> 3  0.043295388
#> 4  0.007642316
#> 5 -0.013809995
#> 
#> $varCovEachRdr
#>   Reader          Var         Cov1
#> 1   Rdr0 0.0006989006 3.734661e-04
#> 2   Rdr1 0.0011060528 7.601598e-04
#> 3   Rdr2 0.0008423434 3.553224e-04
#> 4   Rdr3 0.0001505777 1.083399e-06
#> 5   Rdr4 0.0012135668 2.430368e-04
#> 
#> $FTestStatsRRFC
#>   fRRFC ndfRRFC ddfRRFC      pRRFC
#> 1 8.704       1       4 0.04195875
#> 
#> $ciDiffTrtRRFC
#>   Treatment    Estimate     StdErr DF         t      PrGTt     CILower
#> 1 Trt0-Trt1 -0.04380032 0.01484629  4 -2.950254 0.04195875 -0.08502022
#>       CIUpper
#> 1 -0.00258042
#> 
#> $ciAvgRdrEachTrtRRFC
#>   Treatment      Area     StdErr DF   CILower   CIUpper
#> 1      Trt0 0.8970370 0.02482994  4 0.8280981 0.9659760
#> 2      Trt1 0.9408374 0.01615303  4 0.8959894 0.9856854
#> 

# \donttest{
StSignificanceTesting(dataset05, FOM = "wAFROC")

#> $fomArray
#>           Rdr1      Rdr2      Rdr3      Rdr4     Rdr5      Rdr6      Rdr7
#> Trt1 0.7244681 0.8809693 0.8095745 0.6133176 0.577305 0.8493302 0.8927896
#> Trt2 0.8023641 0.9685579 0.8459811 0.7514184 0.720922 0.8718676 0.9369582
#>           Rdr8      Rdr9
#> Trt1 0.8026005 0.7639874
#> Trt2 0.8994878 0.8190307
#> 
#> $anovaY
#>   Source          SS   DF         MS
#> 1      T   2.5205406    1 2.52054062
#> 2      R  13.4963319    8 1.68704149
#> 3      C  70.1368924   91 0.77073508
#> 4     TR   0.6828268    8 0.08535334
#> 5     TC  34.2104485   91 0.37593899
#> 6     RC  91.6729348  728 0.12592436
#> 7    TRC  54.4372558  728 0.07477645
#> 8  Total 267.1572307 1655         NA
#> 
#> $anovaYi
#>   Source  DF      Trt1       Trt2
#> 1      R   8 1.1508797 0.62151513
#> 2      C  91 0.7536569 0.39301721
#> 3     RC 728 0.1044848 0.09621601
#> 
#> $varComp
#>          varR       varC        varTR     varTC      varRC     varErr
#> 1 0.008426849 0.01909157 0.0001149662 0.0334625 0.02557396 0.07477645
#> 
#> $FTestStatsRRRC
#>      fRRRC ndfRRRC  ddfRRRC      pRRRC
#> 1 6.521182       1 164.0528 0.01157058
#> 
#> $ciDiffTrtRRRC
#>     TrtDiff    Estimate     StdErr       DF         t      PrGTt    CILower
#> 1 Trt1-Trt2 -0.07802732 0.03055509 164.0528 -2.553661 0.01157058 -0.1383592
#>       CIUpper
#> 1 -0.01769539
#> 
#> $ciAvgRdrEachTrtRRRC
#>   Treatment      Area     StdErr       DF   CILower   CIUpper
#> 1      Trt1 0.7682602 0.04662591 19.57042 0.6708632 0.8656572
#> 2      Trt2 0.8462875 0.03330282 17.46510 0.7761670 0.9164081
#> 
#> $FTestStatsFRRC
#>      fFRRC ndfFRRC ddfFRRC      pFRRC
#> 1 6.704653       1      91 0.01119609
#> 
#> $ciDiffTrtFRRC
#>   Treatment    Estimate     StdErr DF         t      PrGTt   CILower
#> 1 Trt1-Trt2 -0.07802732 0.03013412 91 -2.589335 0.01119609 -0.137885
#>       CIUpper
#> 1 -0.01816959
#> 
#> $ciAvgRdrEachTrtFRRC
#>   Treatment      Area     StdErr DF   CILower   CIUpper
#> 1      Trt1 0.7682602 0.03016975 91 0.7083317 0.8281887
#> 2      Trt2 0.8462875 0.02178666 91 0.8030110 0.8895641
#> 
#> $msAnovaEachRdrFRRC
#>   Source DF      Rdr1       Rdr2       Rdr3      Rdr4      Rdr5       Rdr6
#> 1      T  1 0.2791181 0.35290151 0.06097033 0.8773051 0.9487890 0.02336505
#> 2      C 91 0.2828516 0.07427785 0.19669058 0.2882843 0.3349374 0.14146184
#> 3     TC 91 0.1358427 0.05289022 0.09472097 0.1244465 0.1502716 0.11912122
#>         Rdr7       Rdr8      Rdr9
#> 1 0.08973995 0.43180896 0.1393694
#> 2 0.06905361 0.15795870 0.2326140
#> 3 0.06797925 0.08507088 0.1438072
#> 
#> $ciDiffTrtEachRdrFRRC
#>   Reader Treatment    Estimate     StdErr DF          t       PrGTt    CILower
#> 1   Rdr1 Trt1-Trt2 -0.07789598 0.05434246 91 -1.4334276 0.155162626 -0.1858406
#> 2   Rdr2 Trt1-Trt2 -0.08758865 0.03390851 91 -2.5830873 0.011386438 -0.1549438
#> 3   Rdr3 Trt1-Trt2 -0.03640662 0.04537787 91 -0.8022989 0.424470369 -0.1265442
#> 4   Rdr4 Trt1-Trt2 -0.13810087 0.05201307 91 -2.6551185 0.009359286 -0.2414184
#> 5   Rdr5 Trt1-Trt2 -0.14361702 0.05715570 91 -2.5127329 0.013740646 -0.2571498
#> 6   Rdr6 Trt1-Trt2 -0.02253743 0.05088803 91 -0.4428827 0.658901563 -0.1236203
#> 7   Rdr7 Trt1-Trt2 -0.04416864 0.03844229 91 -1.1489595 0.253583707 -0.1205295
#> 8   Rdr8 Trt1-Trt2 -0.09688731 0.04300427 91 -2.2529698 0.026664234 -0.1823100
#> 9   Rdr9 Trt1-Trt2 -0.05504334 0.05591281 91 -0.9844495 0.327504771 -0.1661073
#>       CIUpper
#> 1  0.03004864
#> 2 -0.02023355
#> 3  0.05373095
#> 4 -0.03478329
#> 5 -0.03008424
#> 6  0.07854539
#> 7  0.03219225
#> 8 -0.01146462
#> 9  0.05602059
#> 
#> $FTestStatsRRFC
#>      fRRFC ndfRRFC ddfRRFC        pRRFC
#> 1 29.53066       1       8 0.0006202254
#> 
#> $ciDiffTrtRRFC
#>   Treatment    Estimate     StdErr DF         t        PrGTt    CILower
#> 1 Trt1-Trt2 -0.07802732 0.01435853  8 -5.434212 0.0006202254 -0.1111382
#>       CIUpper
#> 1 -0.04491648
#> 
#> $ciAvgRdrEachTrtRRFC
#>   Treatment      Area     StdErr DF   CILower   CIUpper
#> 1      Trt1 0.7682602 0.03728205  8 0.6822877 0.8542328
#> 2      Trt2 0.8462875 0.02739748  8 0.7831088 0.9094663
#> 

StSignificanceTesting(dataset05, FOM = "HrAuc", method = "DBMH")

#> $fomArray
#>           Rdr1      Rdr2      Rdr3     Rdr4      Rdr5      Rdr6      Rdr7
#> Trt1 0.8588652 0.8931442 0.8822695 0.730260 0.7141844 0.9290780 0.9189125
#> Trt2 0.8914894 0.9761229 0.8827423 0.848227 0.8399527 0.9089835 0.9524823
#>           Rdr8      Rdr9
#> Trt1 0.8791962 0.8144208
#> Trt2 0.9338061 0.8569740
#> 
#> $anovaY
#>   Source          SS   DF         MS
#> 1      T   1.1312035    1 1.13120350
#> 2      R   5.2762307    8 0.65952883
#> 3      C  55.8234912   91 0.61344496
#> 4     TR   0.8931472    8 0.11164340
#> 5     TC  20.7902271   91 0.22846403
#> 6     RC  71.3207550  728 0.09796807
#> 7    TRC  48.6128298  728 0.06677587
#> 8  Total 203.8478844 1655         NA
#> 
#> $anovaYi
#>   Source  DF       Trt1       Trt2
#> 1      R   8 0.56105777 0.21011446
#> 2      C  91 0.56306715 0.27884184
#> 3     RC 728 0.09095622 0.07378772
#> 
#> $varComp
#>          varR       varC        varTR      varTC     varRC     varErr
#> 1 0.002808115 0.01965493 0.0004876905 0.01796535 0.0155961 0.06677587
#> 
#> $FTestStatsRRRC
#>      fRRRC ndfRRRC  ddfRRRC      pRRRC
#> 1 4.138576       1 47.95164 0.04746731
#> 
#> $ciDiffTrtRRRC
#>     TrtDiff    Estimate     StdErr       DF         t      PrGTt    CILower
#> 1 Trt1-Trt2 -0.05227213 0.02569477 47.95164 -2.034349 0.04746731 -0.1039363
#>        CIUpper
#> 1 -0.000607988
#> 
#> $ciAvgRdrEachTrtRRRC
#>   Treatment      Area     StdErr       DF   CILower   CIUpper
#> 1      Trt1 0.8467034 0.03532405 27.12797 0.7742405 0.9191664
#> 2      Trt2 0.8989756 0.02239222 31.23396 0.8533202 0.9446309
#> 
#> $FTestStatsFRRC
#>      fFRRC ndfFRRC ddfFRRC      pFRRC
#> 1 4.951342       1      91 0.02854189
#> 
#> $ciDiffTrtFRRC
#>   Treatment    Estimate     StdErr DF         t      PrGTt     CILower
#> 1 Trt1-Trt2 -0.05227213 0.02349139 91 -2.225161 0.02854189 -0.09893489
#>        CIUpper
#> 1 -0.005609366
#> 
#> $ciAvgRdrEachTrtFRRC
#>   Treatment      Area     StdErr DF   CILower   CIUpper
#> 1      Trt1 0.8467034 0.02607744 91 0.7949038 0.8985031
#> 2      Trt2 0.8989756 0.01835117 91 0.8625232 0.9354279
#> 
#> $msAnovaEachRdrFRRC
#>   Source DF       Rdr1       Rdr2         Rdr3       Rdr4      Rdr5       Rdr6
#> 1      T  1 0.04895931 0.31673155 1.028341e-05 0.64014475 0.7276129 0.01857441
#> 2      C 91 0.15540136 0.06923532 1.518344e-01 0.27891232 0.2707407 0.09622574
#> 3     TC 91 0.08433040 0.05266761 5.920567e-02 0.09847573 0.1552121 0.05754136
#>         Rdr7       Rdr8       Rdr9
#> 1 0.05183866 0.13718324 0.08329561
#> 2 0.05829532 0.11009772 0.20644665
#> 3 0.06141383 0.06774519 0.12607906
#> 
#> $ciDiffTrtEachRdrFRRC
#>   Reader Treatment      Estimate     StdErr DF           t      PrGTt
#> 1   Rdr1 Trt1-Trt2 -0.0326241135 0.04281670 91 -0.76194840 0.44806112
#> 2   Rdr2 Trt1-Trt2 -0.0829787234 0.03383708 91 -2.45230142 0.01610224
#> 3   Rdr3 Trt1-Trt2 -0.0004728132 0.03587589 91 -0.01317913 0.98951373
#> 4   Rdr4 Trt1-Trt2 -0.1179669031 0.04626853 91 -2.54961427 0.01245695
#> 5   Rdr5 Trt1-Trt2 -0.1257683215 0.05808766 91 -2.16514697 0.03299256
#> 6   Rdr6 Trt1-Trt2  0.0200945626 0.03536805 91  0.56815578 0.57132850
#> 7   Rdr7 Trt1-Trt2 -0.0335697400 0.03653879 91 -0.91874248 0.36065911
#> 8   Rdr8 Trt1-Trt2 -0.0546099291 0.03837605 91 -1.42302100 0.15814934
#> 9   Rdr9 Trt1-Trt2 -0.0425531915 0.05235312 91 -0.81281100 0.41844785
#>       CILower     CIUpper
#> 1 -0.11767422  0.05242599
#> 2 -0.15019193 -0.01576552
#> 3 -0.07173587  0.07079024
#> 4 -0.20987365 -0.02606015
#> 5 -0.24115232 -0.01038432
#> 6 -0.05015973  0.09034885
#> 7 -0.10614956  0.03901008
#> 8 -0.13083924  0.02161939
#> 9 -0.14654623  0.06143985
#> 
#> $FTestStatsRRFC
#>      fRRFC ndfRRFC ddfRRFC      pRRFC
#> 1 10.13229       1       8 0.01293647
#> 
#> $ciDiffTrtRRFC
#>   Treatment    Estimate     StdErr DF         t      PrGTt     CILower
#> 1 Trt1-Trt2 -0.05227213 0.01642163  8 -3.183126 0.01293647 -0.09014048
#>       CIUpper
#> 1 -0.01440378
#> 
#> $ciAvgRdrEachTrtRRFC
#>   Treatment      Area     StdErr DF   CILower   CIUpper
#> 1      Trt1 0.8467034 0.02603087  8 0.7866762 0.9067307
#> 2      Trt2 0.8989756 0.01592989  8 0.8622412 0.9357100
#> 

StSignificanceTesting(dataset05, FOM = "SongA1", method = "DBMH")

#> $fomArray
#>           Rdr1      Rdr2      Rdr3      Rdr4      Rdr5      Rdr6      Rdr7
#> Trt1 0.8687943 0.8924350 0.8877069 0.7290780 0.7276596 0.9217494 0.9189125
#> Trt2 0.9066194 0.9761229 0.8957447 0.8468085 0.8394799 0.9170213 0.9513002
#>          Rdr8      Rdr9
#> Trt1 0.884870 0.8231678
#> Trt2 0.934279 0.8654846
#> 
#> $anovaY
#>   Source          SS   DF         MS
#> 1      T   1.1701879    1 1.17018790
#> 2      R   5.0736487    8 0.63420608
#> 3      C  52.9690122   91 0.58207706
#> 4     TR   0.6774732    8 0.08468415
#> 5     TC  21.3527068   91 0.23464513
#> 6     RC  70.1156297  728 0.09631268
#> 7    TRC  48.3642304  728 0.06643438
#> 8  Total 199.7228889 1655         NA
#> 
#> $anovaYi
#>   Source  DF       Trt1       Trt2
#> 1      R   8 0.51745098 0.20143926
#> 2      C  91 0.55736487 0.25935732
#> 3     RC 728 0.08915823 0.07358883
#> 
#> $varComp
#>         varR       varC        varTR      varTC      varRC     varErr
#> 1 0.00282415 0.01764187 0.0001983671 0.01869008 0.01493915 0.06643438
#> 
#> $FTestStatsRRRC
#>      fRRRC ndfRRRC  ddfRRRC      pRRRC
#> 1 4.627171       1 71.34537 0.03485704
#> 
#> $ciDiffTrtRRRC
#>     TrtDiff    Estimate     StdErr       DF         t      PrGTt    CILower
#> 1 Trt1-Trt2 -0.05316522 0.02471553 71.34537 -2.151086 0.03485704 -0.1024424
#>        CIUpper
#> 1 -0.003887996
#> 
#> $ciAvgRdrEachTrtRRRC
#>   Treatment      Area     StdErr       DF   CILower   CIUpper
#> 1      Trt1 0.8504859 0.03450229 29.02711 0.7799237 0.9210482
#> 2      Trt2 0.9036512 0.02162504 29.55901 0.8594593 0.9478430
#> 
#> $FTestStatsFRRC
#>      fFRRC ndfFRRC ddfFRRC      pFRRC
#> 1 4.987054       1      91 0.02798952
#> 
#> $ciDiffTrtFRRC
#>   Treatment    Estimate     StdErr DF         t      PrGTt   CILower
#> 1 Trt1-Trt2 -0.05316522 0.02380705 91 -2.233171 0.02798952 -0.100455
#>       CIUpper
#> 1 -0.00587544
#> 
#> $ciAvgRdrEachTrtFRRC
#>   Treatment      Area     StdErr DF   CILower   CIUpper
#> 1      Trt1 0.8504859 0.02594506 91 0.7989493 0.9020226
#> 2      Trt2 0.9036512 0.01769840 91 0.8684955 0.9388069
#> 
#> $msAnovaEachRdrFRRC
#>   Source DF       Rdr1       Rdr2        Rdr3      Rdr4      Rdr5        Rdr6
#> 1      T  1 0.06581381 0.32216890 0.002971905 0.6375816 0.5751742 0.001028341
#> 2      C 91 0.12888035 0.06996783 0.142350596 0.2792899 0.2731340 0.100446128
#> 3     TC 91 0.08779674 0.05347638 0.060003871 0.1003674 0.1468245 0.055170190
#>         Rdr7       Rdr8       Rdr9
#> 1 0.04825232 0.11229739 0.08237267
#> 2 0.05900943 0.10675232 0.19274794
#> 3 0.06217614 0.06671879 0.13358615
#> 
#> $ciDiffTrtEachRdrFRRC
#>   Reader Treatment     Estimate     StdErr DF          t      PrGTt     CILower
#> 1   Rdr1 Trt1-Trt2 -0.037825059 0.04368781 91 -0.8658035 0.38887487 -0.12460553
#> 2   Rdr2 Trt1-Trt2 -0.083687943 0.03409589 91 -2.4544874 0.01601086 -0.15141525
#> 3   Rdr3 Trt1-Trt2 -0.008037825 0.03611692 91 -0.2225501 0.82438441 -0.07977965
#> 4   Rdr4 Trt1-Trt2 -0.117730496 0.04671082 91 -2.5204119 0.01346397 -0.21051579
#> 5   Rdr5 Trt1-Trt2 -0.111820331 0.05649635 91 -1.9792489 0.05081050 -0.22404338
#> 6   Rdr6 Trt1-Trt2  0.004728132 0.03463166 91  0.1365263 0.89170683 -0.06406341
#> 7   Rdr7 Trt1-Trt2 -0.032387707 0.03676486 91 -0.8809418 0.38067008 -0.10541659
#> 8   Rdr8 Trt1-Trt2 -0.049408983 0.03808423 91 -1.2973607 0.19778650 -0.12505863
#> 9   Rdr9 Trt1-Trt2 -0.042316785 0.05388921 91 -0.7852553 0.43434361 -0.14936108
#>         CIUpper
#> 1  0.0489554120
#> 2 -0.0159606387
#> 3  0.0637040018
#> 4 -0.0249451990
#> 5  0.0004027206
#> 6  0.0735196731
#> 7  0.0406411782
#> 8  0.0262406608
#> 9  0.0647275057
#> 
#> $FTestStatsRRFC
#>      fRRFC ndfRRFC ddfRRFC       pRRFC
#> 1 13.81826       1       8 0.005894272
#> 
#> $ciDiffTrtRRFC
#>   Treatment    Estimate     StdErr DF         t       PrGTt     CILower
#> 1 Trt1-Trt2 -0.05316522 0.01430214  8 -3.717292 0.005894272 -0.08614601
#>       CIUpper
#> 1 -0.02018444
#> 
#> $ciAvgRdrEachTrtRRFC
#>   Treatment      Area     StdErr DF   CILower   CIUpper
#> 1      Trt1 0.8504859 0.02499882  8 0.7928386 0.9081333
#> 2      Trt2 0.9036512 0.01559757  8 0.8676831 0.9396192
#> 

StSignificanceTesting(dataset05, FOM = "SongA2", method = "DBMH")

#> $fomArray
#>           Rdr1      Rdr2      Rdr3      Rdr4      Rdr5      Rdr6      Rdr7
#> Trt1 0.8524823 0.8907801 0.8825059 0.7241135 0.7248227 0.9033097 0.9167849
#> Trt2 0.8427896 0.9756501 0.8806147 0.8373522 0.8326241 0.8803783 0.9494090
#>           Rdr8      Rdr9
#> Trt1 0.8782506 0.8170213
#> Trt2 0.9234043 0.8453901
#> 
#> $anovaY
#>   Source          SS   DF         MS
#> 1      T   0.7285249    1 0.72852493
#> 2      R   5.0697330    8 0.63371662
#> 3      C  55.5554986   91 0.61049999
#> 4     TR   0.9356828    8 0.11696035
#> 5     TC  21.4719085   91 0.23595504
#> 6     RC  76.9761232  728 0.10573643
#> 7    TRC  55.5779319  728 0.07634331
#> 8  Total 216.3154029 1655         NA
#> 
#> $anovaYi
#>   Source  DF       Trt1       Trt2
#> 1      R   8 0.49486075 0.25581621
#> 2      C  91 0.56700925 0.27944578
#> 3     RC 728 0.09374477 0.08833498
#> 
#> $varComp
#>          varR      varC        varTR      varTC      varRC     varErr
#> 1 0.002648713 0.0191751 0.0004414895 0.01773464 0.01469656 0.07634331
#> 
#> $FTestStatsRRRC
#>      fRRRC ndfRRRC  ddfRRRC     pRRRC
#> 1 2.634123       1 44.73314 0.1116172
#> 
#> $ciDiffTrtRRRC
#>     TrtDiff    Estimate     StdErr       DF         t     PrGTt     CILower
#> 1 Trt1-Trt2 -0.04194904 0.02584663 44.73314 -1.622998 0.1116172 -0.09401541
#>      CIUpper
#> 1 0.01011732
#> 
#> $ciAvgRdrEachTrtRRRC
#>   Treatment      Area     StdErr       DF   CILower   CIUpper
#> 1      Trt1 0.8433412 0.03419405 30.61872 0.7735668 0.9131157
#> 2      Trt2 0.8852903 0.02323288 24.41783 0.8373833 0.9331972
#> 
#> $FTestStatsFRRC
#>      fFRRC ndfFRRC ddfFRRC      pFRRC
#> 1 3.087558       1      91 0.08225646
#> 
#> $ciDiffTrtFRRC
#>   Treatment    Estimate     StdErr DF         t      PrGTt     CILower
#> 1 Trt1-Trt2 -0.04194904 0.02387341 91 -1.757145 0.08225646 -0.08937064
#>       CIUpper
#> 1 0.005472555
#> 
#> $ciAvgRdrEachTrtFRRC
#>   Treatment      Area     StdErr DF   CILower   CIUpper
#> 1      Trt1 0.8433412 0.02616856 91 0.7913606 0.8953219
#> 2      Trt2 0.8852903 0.01837103 91 0.8487985 0.9217821
#> 
#> $msAnovaEachRdrFRRC
#>   Source DF        Rdr1       Rdr2         Rdr3      Rdr4      Rdr5       Rdr6
#> 1      T  1 0.004321602 0.33133399 0.0001645345 0.5898589 0.5345727 0.02418915
#> 2      C 91 0.153019237 0.07132243 0.1512020679 0.2781543 0.2733314 0.14335394
#> 3     TC 91 0.121957698 0.05453287 0.0655964105 0.1076593 0.1496636 0.07322754
#>         Rdr7       Rdr8       Rdr9
#> 1 0.04895931 0.09378726 0.03702027
#> 2 0.06107056 0.11176465 0.21317284
#> 3 0.06212606 0.07357930 0.13835878
#> 
#> $ciDiffTrtEachRdrFRRC
#>   Reader Treatment     Estimate     StdErr DF           t      PrGTt
#> 1   Rdr1 Trt1-Trt2  0.009692671 0.05149033 91  0.18824255 0.85110566
#> 2   Rdr2 Trt1-Trt2 -0.084869976 0.03443105 91 -2.46492558 0.01558087
#> 3   Rdr3 Trt1-Trt2  0.001891253 0.03776253 91  0.05008279 0.96016618
#> 4   Rdr4 Trt1-Trt2 -0.113238771 0.04837788 91 -2.34071369 0.02143187
#> 5   Rdr5 Trt1-Trt2 -0.107801418 0.05703995 91 -1.88992824 0.06195068
#> 6   Rdr6 Trt1-Trt2  0.022931442 0.03989866 91  0.57474217 0.56688389
#> 7   Rdr7 Trt1-Trt2 -0.032624113 0.03675005 91 -0.88772967 0.37702653
#> 8   Rdr8 Trt1-Trt2 -0.045153664 0.03999438 91 -1.12900037 0.26186515
#> 9   Rdr9 Trt1-Trt2 -0.028368794 0.05484341 91 -0.51726898 0.60622347
#>       CILower      CIUpper
#> 1 -0.09258655  0.111971891
#> 2 -0.15326303 -0.016476925
#> 3 -0.07311938  0.076901886
#> 4 -0.20933549 -0.017142051
#> 5 -0.22110428  0.005501441
#> 6 -0.05632234  0.102185228
#> 7 -0.10562358  0.040375355
#> 8 -0.12459758  0.034290249
#> 9 -0.13730849  0.080570901
#> 
#> $FTestStatsRRFC
#>     fRRFC ndfRRFC ddfRRFC     pRRFC
#> 1 6.22882       1       8 0.0371869
#> 
#> $ciDiffTrtRRFC
#>   Treatment    Estimate     StdErr DF        t     PrGTt     CILower
#> 1 Trt1-Trt2 -0.04194904 0.01680812  8 -2.49576 0.0371869 -0.08070864
#>        CIUpper
#> 1 -0.003189446
#> 
#> $ciAvgRdrEachTrtRRFC
#>   Treatment      Area     StdErr DF   CILower   CIUpper
#> 1      Trt1 0.8433412 0.02444704  8 0.7869662 0.8997162
#> 2      Trt2 0.8852903 0.01757717  8 0.8447572 0.9258233
#> 

StSignificanceTesting(dataset05, FOM = "wJAFROC1", method = "DBMH")

#> $fomArray
#>           Rdr1      Rdr2      Rdr3      Rdr4      Rdr5      Rdr6      Rdr7
#> Trt1 0.7427729 0.8822271 0.7890264 0.6459490 0.6378739 0.8107655 0.8780643
#> Trt2 0.8097441 0.9692800 0.8720899 0.7603492 0.7558010 0.8720706 0.9432046
#>           Rdr8      Rdr9
#> Trt1 0.8102451 0.7751503
#> Trt2 0.8843085 0.8335646
#> 
#> $anovaY
#>   Source          SS   DF         MS
#> 1      T   2.7113227    1 2.71132271
#> 2      R   9.4725180    8 1.18406475
#> 3      C  63.9011774   91 0.70221074
#> 4     TR   0.1800666    8 0.02250832
#> 5     TC  29.3856670   91 0.32291942
#> 6     RC  71.5832344  728 0.09832862
#> 7    TRC  44.0119542  728 0.06045598
#> 8  Total 221.2459403 1655         NA
#> 
#> $anovaYi
#>   Source  DF       Trt1       Trt2
#> 1      R   8 0.70533444 0.50123863
#> 2      C  91 0.65733668 0.36779348
#> 3     RC 728 0.08466051 0.07412409
#> 
#> $varComp
#>          varR      varC         varTR     varTC      varRC     varErr
#> 1 0.006106977 0.0189677 -0.0004124746 0.0291626 0.01893632 0.06045598
#> 
#> $FTestStatsRRRC
#>      fRRRC ndfRRRC  ddfRRRC       pRRRC
#> 1 9.514356       1 1282.353 0.002082295
#> 
#> $ciDiffTrtRRRC
#>     TrtDiff    Estimate     StdErr       DF         t       PrGTt   CILower
#> 1 Trt1-Trt2 -0.08092644 0.02623619 1282.353 -3.084535 0.002082295 -0.132397
#>       CIUpper
#> 1 -0.02945588
#> 
#> $ciAvgRdrEachTrtRRRC
#>   Treatment      Area     StdErr       DF   CILower   CIUpper
#> 1      Trt1 0.7746749 0.03928729 26.26448 0.6939583 0.8553916
#> 2      Trt2 0.8556014 0.03098441 20.12031 0.7909938 0.9202090
#> 
#> $FTestStatsFRRC
#>      fFRRC ndfFRRC ddfFRRC      pFRRC
#> 1 8.396283       1      91 0.00471003
#> 
#> $ciDiffTrtFRRC
#>   Treatment    Estimate     StdErr DF         t      PrGTt    CILower  CIUpper
#> 1 Trt1-Trt2 -0.08092644 0.02792845 91 -2.897634 0.00471003 -0.1364029 -0.02545
#> 
#> $ciAvgRdrEachTrtFRRC
#>   Treatment      Area     StdErr DF   CILower   CIUpper
#> 1      Trt1 0.7746749 0.02817596 91 0.7187069 0.8306430
#> 2      Trt2 0.8556014 0.02107593 91 0.8137366 0.8974661
#> 
#> $msAnovaEachRdrFRRC
#>   Source DF      Rdr1       Rdr2       Rdr3      Rdr4      Rdr5       Rdr6
#> 1      T  1 0.2063163 0.34859741 0.31737924 0.6020212 0.6397126 0.17288261
#> 2      C 91 0.2556614 0.06883084 0.16478901 0.2221556 0.2794581 0.13383119
#> 3     TC 91 0.1225766 0.04391986 0.07947756 0.1284352 0.1231786 0.08844306
#>         Rdr7       Rdr8       Rdr9
#> 1 0.19518991 0.25232759 0.15696248
#> 2 0.06328581 0.15760943 0.14321827
#> 3 0.06355408 0.07241945 0.08456282
#> 
#> $ciDiffTrtEachRdrFRRC
#>   Reader Treatment    Estimate     StdErr DF         t       PrGTt    CILower
#> 1   Rdr1 Trt1-Trt2 -0.06697117 0.05162083 91 -1.297367 0.197784260 -0.1695096
#> 2   Rdr2 Trt1-Trt2 -0.08705288 0.03089951 91 -2.817290 0.005939808 -0.1484310
#> 3   Rdr3 Trt1-Trt2 -0.08306352 0.04156649 91 -1.998329 0.048665773 -0.1656302
#> 4   Rdr4 Trt1-Trt2 -0.11440025 0.05284003 91 -2.165030 0.033001802 -0.2193605
#> 5   Rdr5 Trt1-Trt2 -0.11792707 0.05174743 91 -2.278897 0.025012189 -0.2207170
#> 6   Rdr6 Trt1-Trt2 -0.06130512 0.04384832 91 -1.398118 0.165476116 -0.1484044
#> 7   Rdr7 Trt1-Trt2 -0.06514030 0.03717002 91 -1.752496 0.083056811 -0.1389740
#> 8   Rdr8 Trt1-Trt2 -0.07406337 0.03967790 91 -1.866615 0.065175577 -0.1528786
#> 9   Rdr9 Trt1-Trt2 -0.05841428 0.04287566 91 -1.362411 0.176430685 -0.1435815
#>         CIUpper
#> 1  0.0355672603
#> 2 -0.0256748087
#> 3 -0.0004967975
#> 4 -0.0094400136
#> 5 -0.0151371678
#> 6  0.0257941841
#> 7  0.0086933771
#> 8  0.0047519097
#> 9  0.0267529529
#> 
#> $FTestStatsRRFC
#>      fRRFC ndfRRFC ddfRRFC        pRRFC
#> 1 120.4587       1       8 4.219684e-06
#> 
#> $ciDiffTrtRRFC
#>   Treatment    Estimate      StdErr DF         t        PrGTt     CILower
#> 1 Trt1-Trt2 -0.08092644 0.007373461  8 -10.97537 4.219684e-06 -0.09792967
#>       CIUpper
#> 1 -0.06392321
#> 
#> $ciAvgRdrEachTrtRRFC
#>   Treatment      Area     StdErr DF   CILower   CIUpper
#> 1      Trt1 0.7746749 0.02918652  8 0.7073707 0.8419792
#> 2      Trt2 0.8556014 0.02460408  8 0.7988643 0.9123385
#> 

StSignificanceTesting(dataset05, FOM = "JAFROC1", method = "DBMH")

#> $fomArray
#>           Rdr1      Rdr2      Rdr3      Rdr4      Rdr5      Rdr6      Rdr7
#> Trt1 0.7471444 0.8856853 0.7941231 0.6552137 0.6369749 0.8118091 0.8819086
#> Trt2 0.7970707 0.9636146 0.8769344 0.7471444 0.7546979 0.8672623 0.9367170
#>           Rdr8      Rdr9
#> Trt1 0.8145726 0.7784635
#> Trt2 0.8849484 0.8221260
#> 
#> $anovaY
#>   Source          SS   DF         MS
#> 1      T   2.1238485    1 2.12384848
#> 2      R   9.6334648    8 1.20418310
#> 3      C  53.0953405   91 0.58346528
#> 4     TR   0.2070386    8 0.02587983
#> 5     TC  22.1540037   91 0.24345059
#> 6     RC  64.2655140  728 0.08827680
#> 7    TRC  42.5649592  728 0.05846835
#> 8  Total 194.0441693 1655         NA
#> 
#> $anovaYi
#>   Source  DF       Trt1       Trt2
#> 1      R   8 0.70206158 0.52800134
#> 2      C  91 0.49992187 0.32699400
#> 3     RC 728 0.07299324 0.07375192
#> 
#> $varComp
#>         varR       varC         varTR      varTC      varRC     varErr
#> 1 0.00624182 0.01723368 -0.0003542231 0.02055358 0.01490423 0.05846835
#> 
#> $FTestStatsRRRC
#>      fRRRC ndfRRRC  ddfRRRC       pRRRC
#> 1 10.07222       1 531.0852 0.001592393
#> 
#> $ciDiffTrtRRRC
#>     TrtDiff   Estimate     StdErr       DF         t       PrGTt    CILower
#> 1 Trt1-Trt2 -0.0716245 0.02256831 531.0852 -3.173676 0.001592393 -0.1159586
#>       CIUpper
#> 1 -0.02729039
#> 
#> $ciAvgRdrEachTrtRRRC
#>   Treatment      Area     StdErr       DF   CILower   CIUpper
#> 1      Trt1 0.7784328 0.03692580 20.68807 0.7015708 0.8552948
#> 2      Trt2 0.8500573 0.03071695 17.51430 0.7853949 0.9147197
#> 
#> $FTestStatsFRRC
#>      fFRRC ndfFRRC ddfFRRC       pFRRC
#> 1 8.723941       1      91 0.003996482
#> 
#> $ciDiffTrtFRRC
#>   Treatment   Estimate     StdErr DF         t       PrGTt    CILower
#> 1 Trt1-Trt2 -0.0716245 0.02424964 91 -2.953632 0.003996482 -0.1197934
#>       CIUpper
#> 1 -0.02345557
#> 
#> $ciAvgRdrEachTrtFRRC
#>   Treatment      Area     StdErr DF   CILower   CIUpper
#> 1      Trt1 0.7784328 0.02457174 91 0.7296241 0.8272416
#> 2      Trt2 0.8500573 0.01987260 91 0.8105828 0.8895318
#> 
#> $msAnovaEachRdrFRRC
#>   Source DF      Rdr1       Rdr2       Rdr3      Rdr4      Rdr5       Rdr6
#> 1      T  1 0.1146613 0.27935657 0.31545509 0.3887579 0.6374995 0.14145267
#> 2      C 91 0.2116076 0.07525673 0.13676648 0.1749548 0.2118838 0.11838259
#> 3     TC 91 0.1224560 0.03179289 0.06403499 0.1252785 0.1076908 0.07779831
#>         Rdr7       Rdr8      Rdr9
#> 1 0.13818220 0.22782684 0.0876950
#> 2 0.06296669 0.12338744 0.1744736
#> 3 0.04238798 0.05957697 0.0801810
#> 
#> $ciDiffTrtEachRdrFRRC
#>   Reader Treatment    Estimate     StdErr DF          t       PrGTt    CILower
#> 1   Rdr1 Trt1-Trt2 -0.04992631 0.05159541 91 -0.9676501 0.335783893 -0.1524143
#> 2   Rdr2 Trt1-Trt2 -0.07792926 0.02628973 91 -2.9642473 0.003873016 -0.1301506
#> 3   Rdr3 Trt1-Trt2 -0.08281135 0.03731039 91 -2.2195253 0.028936227 -0.1569238
#> 4   Rdr4 Trt1-Trt2 -0.09193073 0.05218664 91 -1.7615759 0.081499599 -0.1955931
#> 5   Rdr5 Trt1-Trt2 -0.11772292 0.04838495 91 -2.4330480 0.016927435 -0.2138337
#> 6   Rdr6 Trt1-Trt2 -0.05545321 0.04112502 91 -1.3484054 0.180874551 -0.1371430
#> 7   Rdr7 Trt1-Trt2 -0.05480840 0.03035585 91 -1.8055299 0.074298492 -0.1151066
#> 8   Rdr8 Trt1-Trt2 -0.07037583 0.03598821 91 -1.9555244 0.053589042 -0.1418620
#> 9   Rdr9 Trt1-Trt2 -0.04366249 0.04175003 91 -1.0458074 0.298419986 -0.1265938
#>        CIUpper
#> 1  0.052561644
#> 2 -0.025707940
#> 3 -0.008698849
#> 4  0.011731622
#> 5 -0.021612150
#> 6  0.026236602
#> 7  0.005489772
#> 8  0.001110334
#> 9  0.039268819
#> 
#> $FTestStatsRRFC
#>      fRRFC ndfRRFC ddfRRFC        pRRFC
#> 1 82.06579       1       8 1.766134e-05
#> 
#> $ciDiffTrtRRFC
#>   Treatment   Estimate      StdErr DF         t        PrGTt     CILower
#> 1 Trt1-Trt2 -0.0716245 0.007906431  8 -9.059017 1.766134e-05 -0.08985676
#>       CIUpper
#> 1 -0.05339224
#> 
#> $ciAvgRdrEachTrtRRFC
#>   Treatment      Area     StdErr DF   CILower   CIUpper
#> 1      Trt1 0.7784328 0.02911873  8 0.7112849 0.8455807
#> 2      Trt2 0.8500573 0.02525238  8 0.7918252 0.9082894
#> 

StSignificanceTesting(dataset05, FOM = "JAFROC", method = "DBMH")

#> $fomArray
#>           Rdr1      Rdr2      Rdr3      Rdr4      Rdr5      Rdr6      Rdr7
#> Trt1 0.7301318 0.8843691 0.8141243 0.6248588 0.5789077 0.8506591 0.8960452
#> Trt2 0.7892655 0.9629002 0.8512241 0.7384181 0.7197740 0.8670433 0.9306968
#>           Rdr8      Rdr9
#> Trt1 0.8073446 0.7676083
#> Trt2 0.8996234 0.8077213
#> 
#> $anovaY
#>   Source         SS   DF         MS
#> 1      T   1.918202    1 1.91820230
#> 2      R  13.544116    8 1.69301448
#> 3      C  57.884536   91 0.63609380
#> 4     TR   0.628953    8 0.07861913
#> 5     TC  26.574936   91 0.29203226
#> 6     RC  80.717060  728 0.11087508
#> 7    TRC  51.769491  728 0.07111194
#> 8  Total 233.037294 1655         NA
#> 
#> $anovaYi
#>   Source  DF       Trt1       Trt2
#> 1      R   8 1.12617576 0.64545785
#> 2      C  91 0.58359246 0.34453361
#> 3     RC 728 0.08980408 0.09218294
#> 
#> $varComp
#>          varR       varC        varTR     varTC      varRC     varErr
#> 1 0.008557784 0.01690547 8.159989e-05 0.0245467 0.01988157 0.07111194
#> 
#> $FTestStatsRRRC
#>      fRRRC ndfRRRC  ddfRRRC      pRRRC
#> 1 6.403839       1 116.1292 0.01272572
#> 
#> $ciDiffTrtRRRC
#>     TrtDiff    Estimate     StdErr       DF         t      PrGTt    CILower
#> 1 Trt1-Trt2 -0.06806863 0.02689842 116.1292 -2.530581 0.01272572 -0.1213437
#>       CIUpper
#> 1 -0.01479354
#> 
#> $ciAvgRdrEachTrtRRRC
#>   Treatment      Area     StdErr       DF   CILower   CIUpper
#> 1      Trt1 0.7726721 0.04423210 16.55345 0.6791584 0.8661858
#> 2      Trt2 0.8407407 0.03292886 15.47824 0.7707429 0.9107385
#> 
#> $FTestStatsFRRC
#>     fFRRC ndfFRRC ddfFRRC      pFRRC
#> 1 6.56846       1      91 0.01202163
#> 
#> $ciDiffTrtFRRC
#>   Treatment    Estimate     StdErr DF         t      PrGTt    CILower
#> 1 Trt1-Trt2 -0.06806863 0.02655922 91 -2.562901 0.01202163 -0.1208253
#>       CIUpper
#> 1 -0.01531201
#> 
#> $ciAvgRdrEachTrtFRRC
#>   Treatment      Area     StdErr DF   CILower   CIUpper
#> 1      Trt1 0.7726721 0.02654848 91 0.7199368 0.8254074
#> 2      Trt2 0.8407407 0.02039861 91 0.8002214 0.8812601
#> 
#> $msAnovaEachRdrFRRC
#>   Source DF      Rdr1       Rdr2       Rdr3      Rdr4      Rdr5      Rdr6
#> 1      T  1 0.1608526 0.28368796 0.06331422 0.5932031 0.9127923 0.0123483
#> 2      C 91 0.2322144 0.08004552 0.16456727 0.2254749 0.2520536 0.1286303
#> 3     TC 91 0.1332565 0.04008551 0.07921145 0.1178111 0.1331187 0.1030952
#>         Rdr7      Rdr8       Rdr9
#> 1 0.05523374 0.3917067 0.07401641
#> 2 0.06980407 0.1242606 0.24604383
#> 3 0.04544688 0.0722169 0.13668564
#> 
#> $ciDiffTrtEachRdrFRRC
#>   Reader Treatment    Estimate     StdErr DF          t       PrGTt     CILower
#> 1   Rdr1 Trt1-Trt2 -0.05913371 0.05382267 91 -1.0986766 0.274807495 -0.16604583
#> 2   Rdr2 Trt1-Trt2 -0.07853107 0.02951990 91 -2.6602762 0.009227523 -0.13716872
#> 3   Rdr3 Trt1-Trt2 -0.03709981 0.04149684 91 -0.8940394 0.373659300 -0.11952819
#> 4   Rdr4 Trt1-Trt2 -0.11355932 0.05060741 91 -2.2439267 0.027262577 -0.21408473
#> 5   Rdr5 Trt1-Trt2 -0.14086629 0.05379483 91 -2.6185840 0.010342550 -0.24772312
#> 6   Rdr6 Trt1-Trt2 -0.01638418 0.04734131 91 -0.3460864 0.730076685 -0.11042187
#> 7   Rdr7 Trt1-Trt2 -0.03465160 0.03143208 91 -1.1024278 0.273182818 -0.09708757
#> 8   Rdr8 Trt1-Trt2 -0.09227872 0.03962237 91 -2.3289548 0.022075779 -0.17098370
#> 9   Rdr9 Trt1-Trt2 -0.04011299 0.05451080 91 -0.7358725 0.463700899 -0.14839200
#>       CIUpper
#> 1  0.04777841
#> 2 -0.01989343
#> 3  0.04532857
#> 4 -0.01303391
#> 5 -0.03400946
#> 6  0.07765351
#> 7  0.02778437
#> 8 -0.01357374
#> 9  0.06816601
#> 
#> $FTestStatsRRFC
#>      fRRFC ndfRRFC ddfRRFC       pRRFC
#> 1 24.39867       1       8 0.001135776
#> 
#> $ciDiffTrtRRFC
#>   Treatment    Estimate     StdErr DF         t       PrGTt     CILower
#> 1 Trt1-Trt2 -0.06806863 0.01378047  8 -4.939501 0.001135776 -0.09984645
#>       CIUpper
#> 1 -0.03629082
#> 
#> $ciAvgRdrEachTrtRRFC
#>   Treatment      Area     StdErr DF   CILower   CIUpper
#> 1      Trt1 0.7726721 0.03687975  8 0.6876273 0.8577170
#> 2      Trt2 0.8407407 0.02792022  8 0.7763566 0.9051249
#> 

# }