Calculate empirical figures of merit (FOMs) for specified dataset

Calculate the specified empirical figure of merit for each treatment-reader combination in the ROC, FROC, ROI or LROC dataset

UtilFigureOfMerit(dataset, FOM = "wAFROC", FPFValue = 0.2)

Arguments

dataset
The dataset to be analyzed, RJafroc-package

FOM
The figure of merit; the default is "wAFROC"

FPFValue
Only needed for LROC data; where to evaluate a partial curve based figure of merit. The default is 0.2.

Value

An c(I, J) array, where the row names are modalityID's of the treatments and column names are the readerID's of the readers.

Details

The allowed FOMs depend on the dataType field of the dataset object.

For a dataType = "ROC" dataset only FOM = "Wilcoxon" is allowed.

For a dataType = "FROC" dataset the following FOMs are allowed:

FOM = "AFROC1"

FOM = "AFROC"

FOM = "wAFROC1"

FOM = "wAFROC" (the default)

FOM = "HrAuc"

FOM = "SongA1"

FOM = "SongA2"

FOM = "HrSe"

FOM = "HrSp"

FOM = "MaxLLF"

FOM = "MaxNLF"

FOM = "MaxNLFAllCases"

FOM = "ExpTrnsfmSp"

"MaxLLF", "MaxNLF" and "MaxNLFAllCases" correspond to ordinate, and abscissa, respectively, of the highest point on the FROC operating characteristic obtained by counting all the marks. The "ExpTrnsfmSp" FOM is described in the paper by Popescu. Given the large number of FOMs possible with FROC data, it is appropriate to make a recommendation: it is recommended that one use the wAFROC FOM.

For a dataType = "ROI" dataset only FOM = "ROI" is allowed.

For a dataType = "LROC" dataset the following FOMs are allowed:

FOM = "Wilcoxon" for ROC data inferred from LROC data

FOM = "PCL" the probability of correct localization at specified FPFValue

FOM = "ALROC" the area under the LROC from zero to specified FPFValue

FPFValue is the desired FPF at which to evaluate PCL or ALROC; the default is 0.2.

References

Chakraborty DP (2017) Observer Performance Methods for Diagnostic Imaging - Foundations, Modeling, and Applications with R-Based Examples, CRC Press, Boca Raton, FL. https://www.crcpress.com/Observer-Performance-Methods-for-Diagnostic-Imaging-Foundations-Modeling/Chakraborty/p/book/9781482214840

Chakraborty DP, Berbaum KS (2004) Observer studies involving detection and localization: modeling, analysis, and validation, Medical Physics, 31(8), 1--18.

Song T, Bandos AI, Rockette HE, Gur D (2008) On comparing methods for discriminating between actually negative and actually positive subjects with FROC type data, Medical Physics 35 1547--1558.

Popescu LM (2011) Nonparametric signal detectability evaluation using an exponential transformation of the FROC curve, Medical Physics, 38(10), 5690.

Obuchowski NA, Lieber ML, Powell KA (2000) Data Analysis for Detection and Localization of Multiple Abnormalities with Application to Mammography, Acad Radiol, 7:7 553--554.

Swensson RG (1996) Unified measurement of observer performance in detecting and localizing target objects on images, Med Phys 23:10, 1709--1725.

Examples

UtilFigureOfMerit(dataset02, FOM = "Wilcoxon") # ROC data

#>           Rdr - 0   Rdr - 1   Rdr - 2   Rdr - 3   Rdr - 4
#> Trt - 0 0.9196457 0.8587762 0.9038647 0.9731079 0.8297907
#> Trt - 1 0.9478261 0.9053140 0.9217391 0.9993559 0.9299517

UtilFigureOfMerit(DfFroc2Roc(dataset01), FOM = "Wilcoxon") # FROC dataset, converted to ROC

#>            Rdr - 1   Rdr - 2   Rdr - 3   Rdr - 4   Rdr - 5
#> Trt - BT 0.7959972 0.8652271 0.8411751 0.8311096 0.8465590
#> Trt - DM 0.6689490 0.7211493 0.7841175 0.7979284 0.7061096

UtilFigureOfMerit(dataset01) # FROC dataset, default wAFROC FOM

#>            Rdr - 1   Rdr - 2   Rdr - 3   Rdr - 4   Rdr - 5
#> Trt - BT 0.7602704 0.8406191 0.8171524 0.8153090 0.8278324
#> Trt - DM 0.6425854 0.7049977 0.7518434 0.7724426 0.6836962

UtilFigureOfMerit(datasetCadLroc, FOM = "Wilcoxon") #LROC data

#>          Rdr - 1   Rdr - 2   Rdr - 3   Rdr - 4   Rdr - 5   Rdr - 6  Rdr - 7
#> Trt - 1 0.628125 0.7585417 0.7494792 0.8670313 0.8005208 0.7686979 0.786875
#>         Rdr - 8   Rdr - 9  Rdr - 10
#> Trt - 1    0.84 0.7652083 0.7913542

UtilFigureOfMerit(datasetCadLroc, FOM = "PCL") #LROC data

#>           Rdr - 1   Rdr - 2   Rdr - 3 Rdr - 4 Rdr - 5   Rdr - 6   Rdr - 7
#> Trt - 1 0.5916667 0.6945313 0.6458333 0.80625   0.725 0.6598214 0.7684524
#>         Rdr - 8 Rdr - 9 Rdr - 10
#> Trt - 1  0.7375   0.675    0.675

UtilFigureOfMerit(datasetCadLroc, FOM = "ALROC") #LROC data

#>           Rdr - 1   Rdr - 2   Rdr - 3   Rdr - 4   Rdr - 5   Rdr - 6   Rdr - 7
#> Trt - 1 0.1005382 0.1075944 0.1088715 0.1447396 0.1189063 0.1000335 0.1345511
#>           Rdr - 8   Rdr - 9  Rdr - 10
#> Trt - 1 0.1221875 0.1060417 0.1054167

UtilFigureOfMerit(datasetROI, FOM = "ROI") #ROI data

#>           Rdr - 1   Rdr - 2   Rdr - 3   Rdr - 4   Rdr - 5
#> Trt - 1 0.9057239 0.8842834 0.8579279 0.9350207 0.8352103
#> Trt - 2 0.9297186 0.9546035 0.8937652 0.9531716 0.8770076
# NOT RUN {
UtilFigureOfMerit(dataset02, FOM = "wAFROC") #error
UtilFigureOfMerit(dataset01, FOM = "Wilcoxon") #error
# }