Calculate empirical figures of merit (FOMs) for specified dataset

Calculate the specified empirical figure of merit for each treatment-reader combination in the ROC, FROC, ROI or LROC dataset

UtilFigureOfMerit(dataset, FOM = "wAFROC", FPFValue = 0.2)

Arguments

dataset

The dataset to be analyzed, RJafroc-package

FOM

The figure of merit; the default is "wAFROC"

FPFValue

Only needed for LROC data and FOM = "PCL" or "ALROC"; where to evaluate a partial curve based figure of merit. The default is 0.2.

Value

An c(I, J) dataframe, where the row names are modalityID's of the treatments and column names are the readerID's of the readers.

Details

The allowed FOMs depend on the dataType field of the dataset object.

For dataset$descriptions$design = "SPLIT-PLOT-C", end-point based FOMs (e.g., "MaxLLF") are not allowed. For dataset$descriptions$type = "ROC" only FOM = "Wilcoxon" is allowed. For dataset$descriptions$type = "FROC" the following FOMs are allowed:

FOM = "AFROC1" (use only if zero normal cases)

FOM = "AFROC"

FOM = "wAFROC1" (use only if zero normal cases)

FOM = "wAFROC" (the default)

FOM = "HrAuc"

FOM = "SongA1"

FOM = "SongA2"

FOM = "HrSe" (an example of an end-point based FOM)

FOM = "HrSp" (another example)

FOM = "MaxLLF" (do:)

FOM = "MaxNLF" (do:)

FOM = "MaxNLFAllCases" (do:)

FOM = "ExpTrnsfmSp"

"MaxLLF", "MaxNLF" and "MaxNLFAllCases" correspond to ordinate, and abscissa, respectively, of the highest point on the FROC operating characteristic obtained by counting all the marks. The "ExpTrnsfmSp" FOM is described in the paper by Popescu. Given the large number of FOMs possible with FROC data, it is appropriate to make a recommendation: it is recommended that one use the wAFROC FOM whenever possible.

For dataType = "ROI" dataset only FOM = "ROI" is allowed.

For dataType = "LROC" dataset the following FOMs are allowed:

FOM = "Wilcoxon" for ROC data inferred from LROC data

FOM = "PCL" the probability of correct localization at specified FPFValue

FOM = "ALROC" the area under the LROC from zero to specified FPFValue

FPFValue The FPF at which to evaluate PCL or ALROC; the default is 0.2; only needed for LROC data.

References

Chakraborty DP (2017) Observer Performance Methods for Diagnostic Imaging - Foundations, Modeling, and Applications with R-Based Examples, CRC Press, Boca Raton, FL. https://www.crcpress.com/Observer-Performance-Methods-for-Diagnostic-Imaging-Foundations-Modeling/Chakraborty/p/book/9781482214840

Chakraborty DP, Berbaum KS (2004) Observer studies involving detection and localization: modeling, analysis, and validation, Medical Physics, 31(8), 1--18.

Song T, Bandos AI, Rockette HE, Gur D (2008) On comparing methods for discriminating between actually negative and actually positive subjects with FROC type data, Medical Physics 35 1547--1558.

Popescu LM (2011) Nonparametric signal detectability evaluation using an exponential transformation of the FROC curve, Medical Physics, 38(10), 5690.

Obuchowski NA, Lieber ML, Powell KA (2000) Data Analysis for Detection and Localization of Multiple Abnormalities with Application to Mammography, Acad Radiol, 7:7 553--554.

Swensson RG (1996) Unified measurement of observer performance in detecting and localizing target objects on images, Med Phys 23:10, 1709--1725.

Examples

UtilFigureOfMerit(dataset02, FOM = "Wilcoxon") # ROC data
#>            rdr0       rdr1       rdr2       rdr3       rdr4
#> trt0 0.91964573 0.85877617 0.90386473 0.97310789 0.82979066
#> trt1 0.94782609 0.90531401 0.92173913 0.99935588 0.92995169
UtilFigureOfMerit(DfFroc2Roc(dataset01), FOM = "Wilcoxon") # FROC dataset, converted to ROC
#>            rdr1       rdr2       rdr3       rdr4       rdr5
#> trt1 0.79599719 0.86522706 0.84117509 0.83110955 0.84655899
#> trt2 0.66894897 0.72114934 0.78411751 0.79792837 0.70610955
UtilFigureOfMerit(dataset01) # FROC dataset, default wAFROC FOM
#>             rdr1       rdr2       rdr3       rdr4       rdr5
#> trtBT 0.76027037 0.84061915 0.81715239 0.81530899 0.82783240
#> trtDM 0.64258544 0.70499766 0.75184340 0.77244265 0.68369616
UtilFigureOfMerit(datasetCadLroc, FOM = "Wilcoxon") #LROC data
#>            rdr1      rdr2       rdr3       rdr4     rdr5       rdr6       rdr7
#> trt1 0.81692708 0.8415625 0.84119792 0.89973958 0.838125 0.85635417 0.87869792
#>            rdr8       rdr9    rdr10
#> trt1 0.85838542 0.79703125 0.826875
UtilFigureOfMerit(datasetCadLroc, FOM = "PCL") #LROC data
#>            rdr1       rdr2 rdr3    rdr4  rdr5       rdr6       rdr7   rdr8
#> trt1 0.59166667 0.69453125 0.65 0.80625 0.725 0.65982143 0.76845238 0.7375
#>       rdr9 rdr10
#> trt1 0.675 0.675
UtilFigureOfMerit(datasetCadLroc, FOM = "ALROC") #LROC data
#>            rdr1      rdr2       rdr3       rdr4       rdr5       rdr6
#> trt1 0.10053819 0.1075944 0.10890625 0.14473958 0.11890625 0.10003348
#>            rdr7      rdr8       rdr9      rdr10
#> trt1 0.13455109 0.1221875 0.10604167 0.10541667
UtilFigureOfMerit(datasetROI, FOM = "ROI") #ROI data
#>            rdr1       rdr2       rdr3       rdr4       rdr5
#> trt1 0.90572391 0.88428345 0.85792794 0.93502071 0.83521034
#> trt2 0.92971864 0.95460351 0.89376524 0.95317156 0.87700762
 # these are meant to illustrate conditions which will throw an error
## UtilFigureOfMerit(dataset02, FOM = "wAFROC") #error
## UtilFigureOfMerit(dataset01, FOM = "Wilcoxon") #error