Calculate empirical figures of merit (FOMs) for factorial dataset, standard one-treatment or two-treatment cross-modality

Calculate the specified empirical figure of merit for each modality-reader combination in a standard (1T) or cross-modality (2T) dataset

UtilFigureOfMerit(dataset, FOM = "wAFROC", FPFValue = 0.2)

Arguments

dataset

The dataset to be analyzed, RJafroc-package

FOM

The figure of merit; the default is "wAFROC"

FPFValue

Only needed for LROC data and FOM = "PCL" or "ALROC"; where to evaluate a partial curve based figure of merit. The default is 0.2.

Value

For standard IT dataset: A c(I, J) dataframe, where the row names are modalityID's of the treatments and column names are the readerID's of the readers. For cross-modality dataset: A list containing two data frames are returned: * c(I2, J) data frame, FOMs averaged over the first modality, where the row names are modality IDS of the second modality * c(I1, J) data frames, FOMs averaged over the second modality, where the row names are modality IDs of the first modality, * For either 1T or 2T the column names are the readerID's.

Details

The allowed FOMs depend on the dataType field of the dataset object.

For dataset$descriptions$type = "ROC" only FOM = "Wilcoxon" is allowed. For dataset$descriptions$type = "FROC" the following FOMs are allowed:

FOM = "AFROC1" (use only if no non-diseased cases are available)

FOM = "AFROC"

FOM = "wAFROC1" (use only if no non-diseased cases are available)

FOM = "wAFROC" (the default)

FOM = "HrAuc"

FOM = "HrSe" (example of an end-point based FOM)

FOM = "HrSp" (do:)

FOM = "MaxLLF" (do:)

FOM = "MaxNLF" (do:)

FOM = "MaxNLFAllCases" (do:)

"MaxLLF", "MaxNLF" and "MaxNLFAllCases" correspond to ordinate, and abscissa, respectively, of the highest point on the FROC operating characteristic obtained by counting all the marks. Given the number of FOMs possible with FROC data, it is appropriate to make a recommendation: it is recommended the wAFROC FOM be used whenever possible. One should use the wAFROC1 FOM only if the dataset has no non-diseased cases.

For dataType = "ROI" dataset only FOM = "ROI" is allowed.

For dataType = "LROC" dataset the following FOMs are allowed:

FOM = "Wilcoxon" for ROC data inferred from LROC data

FOM = "PCL" the probability of correct localization at specified FPFValue

FOM = "ALROC" the area under the LROC from zero to specified FPFValue

FPFValue The FPF at which to evaluate PCL or ALROC; the default is 0.2; only needed for LROC data. For cross-modality analysis ROI and LROC datasets are not supported.

References

Chakraborty DP (2017) Observer Performance Methods for Diagnostic Imaging - Foundations, Modeling, and Applications with R-Based Examples, CRC Press, Boca Raton, FL. https://www.routledge.com/Observer-Performance-Methods-for-Diagnostic-Imaging-Foundations-Modeling/Chakraborty/p/book/9781482214840

Examples

res <- UtilFigureOfMerit(dataset02, FOM = "Wilcoxon") # ROC data
res <- UtilFigureOfMerit(dataset01) # FROC dataset, default wAFROC FOM
res <- UtilFigureOfMerit(datasetX, FOM = "wAFROC")