Lesion distribution matrix

The lesion distribution matrix for a dataset.

UtilLesionDistr(dataset)

Arguments

dataset
The dataset

Value

The lesion distribution matrix

Details

lesDistr The lesion distribution matrix, an [1:nRow,2] array, where nRow is the number of unique values of lesions per case in the dataset. The first column contains the number of lesions. The second column contains the fraction of diseased cases with the number of lesions indicated in the first column. See PlotRsmOperatingCharacteristics for a function that depends on lesDistr. See Chapter00Vignette2 for more details.

Examples

UtilLesionDistr (dataset01) # FROC data

#>      [,1]       [,2]
#> [1,]    1 0.93258427
#> [2,]    2 0.06741573

UtilLesionDistr (dataset02) # ROC data

#>      [,1] [,2]
#> [1,]    1    1

UtilLesionDistr (datasetCadLroc) # LROC data

#>      [,1] [,2]
#> [1,]    1    1