Lesion weights distribution

The lesion weights distribution, assuming equal weights. Needed to plot empirical curves for FROC data (ROC, wAFROC).

UtilLesionWeightsDistr(dataset)

Arguments

dataset
The supplied dataset.

Value

lesWghtDistr The lesion weights distribution, an [1:nRow,1:(maxLL+1)] array, where nRow is the number of lesions per case in the dataset and maxLL is the maximum number of lesions per case in the dataset. The first column contains the number of lesions per case, while the remaining columns contain the weights (equal weighting assumed). Missing values are filled with -Inf. This parameter is not to be confused with the lesionWeight list member in an FROC dataset, which enumerates the weights of lesions on individual cases. See PlotRsmOperatingCharacteristics for a function that depends on lesWghtDistr.See Chapter00Vignette2 for fuller explanation.

Examples

UtilLesionWeightsDistr (dataset11) # FROC data

#>       [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#>  [1,]    1 1.000000000        -Inf        -Inf        -Inf        -Inf
#>  [2,]    2 0.500000000 0.500000000        -Inf        -Inf        -Inf
#>  [3,]    3 0.333333333 0.333333333 0.333333333        -Inf        -Inf
#>  [4,]    4 0.250000000 0.250000000 0.250000000 0.250000000        -Inf
#>  [5,]    5 0.200000000 0.200000000 0.200000000 0.200000000 0.200000000
#>  [6,]    6 0.166666667 0.166666667 0.166666667 0.166666667 0.166666667
#>  [7,]    7 0.142857143 0.142857143 0.142857143 0.142857143 0.142857143
#>  [8,]    8 0.125000000 0.125000000 0.125000000 0.125000000 0.125000000
#>  [9,]    9 0.111111111 0.111111111 0.111111111 0.111111111 0.111111111
#> [10,]   10 0.100000000 0.100000000 0.100000000 0.100000000 0.100000000
#> [11,]   11 0.090909091 0.090909091 0.090909091 0.090909091 0.090909091
#> [12,]   12 0.083333333 0.083333333 0.083333333 0.083333333 0.083333333
#> [13,]   16 0.062500000 0.062500000 0.062500000 0.062500000 0.062500000
#> [14,]   20 0.050000000 0.050000000 0.050000000 0.050000000 0.050000000
#>              [,7]        [,8]        [,9]       [,10]       [,11]       [,12]
#>  [1,]        -Inf        -Inf        -Inf        -Inf        -Inf        -Inf
#>  [2,]        -Inf        -Inf        -Inf        -Inf        -Inf        -Inf
#>  [3,]        -Inf        -Inf        -Inf        -Inf        -Inf        -Inf
#>  [4,]        -Inf        -Inf        -Inf        -Inf        -Inf        -Inf
#>  [5,]        -Inf        -Inf        -Inf        -Inf        -Inf        -Inf
#>  [6,] 0.166666667        -Inf        -Inf        -Inf        -Inf        -Inf
#>  [7,] 0.142857143 0.142857143        -Inf        -Inf        -Inf        -Inf
#>  [8,] 0.125000000 0.125000000 0.125000000        -Inf        -Inf        -Inf
#>  [9,] 0.111111111 0.111111111 0.111111111 0.111111111        -Inf        -Inf
#> [10,] 0.100000000 0.100000000 0.100000000 0.100000000 0.100000000        -Inf
#> [11,] 0.090909091 0.090909091 0.090909091 0.090909091 0.090909091 0.090909091
#> [12,] 0.083333333 0.083333333 0.083333333 0.083333333 0.083333333 0.083333333
#> [13,] 0.062500000 0.062500000 0.062500000 0.062500000 0.062500000 0.062500000
#> [14,] 0.050000000 0.050000000 0.050000000 0.050000000 0.050000000 0.050000000
#>             [,13]  [,14]  [,15]  [,16]  [,17] [,18] [,19] [,20] [,21]
#>  [1,]        -Inf   -Inf   -Inf   -Inf   -Inf  -Inf  -Inf  -Inf  -Inf
#>  [2,]        -Inf   -Inf   -Inf   -Inf   -Inf  -Inf  -Inf  -Inf  -Inf
#>  [3,]        -Inf   -Inf   -Inf   -Inf   -Inf  -Inf  -Inf  -Inf  -Inf
#>  [4,]        -Inf   -Inf   -Inf   -Inf   -Inf  -Inf  -Inf  -Inf  -Inf
#>  [5,]        -Inf   -Inf   -Inf   -Inf   -Inf  -Inf  -Inf  -Inf  -Inf
#>  [6,]        -Inf   -Inf   -Inf   -Inf   -Inf  -Inf  -Inf  -Inf  -Inf
#>  [7,]        -Inf   -Inf   -Inf   -Inf   -Inf  -Inf  -Inf  -Inf  -Inf
#>  [8,]        -Inf   -Inf   -Inf   -Inf   -Inf  -Inf  -Inf  -Inf  -Inf
#>  [9,]        -Inf   -Inf   -Inf   -Inf   -Inf  -Inf  -Inf  -Inf  -Inf
#> [10,]        -Inf   -Inf   -Inf   -Inf   -Inf  -Inf  -Inf  -Inf  -Inf
#> [11,]        -Inf   -Inf   -Inf   -Inf   -Inf  -Inf  -Inf  -Inf  -Inf
#> [12,] 0.083333333   -Inf   -Inf   -Inf   -Inf  -Inf  -Inf  -Inf  -Inf
#> [13,] 0.062500000 0.0625 0.0625 0.0625 0.0625  -Inf  -Inf  -Inf  -Inf
#> [14,] 0.050000000 0.0500 0.0500 0.0500 0.0500  0.05  0.05  0.05  0.05

UtilLesionWeightsDistr (dataset02) # ROC data

#>      [,1] [,2]
#> [1,]    1    1

UtilLesionWeightsDistr (datasetCadLroc) # LROC data

#>      [,1] [,2]
#> [1,]    1    1