Calculate mean squares for factorial dataset

Calculates the mean squares used in the DBM and ORH methods for factorial dataset

UtilMeanSquares(dataset, FOM = "Wilcoxon", FPFValue = 0.2, method = "DBM")

Arguments

dataset

The dataset to be analyzed, see RJafroc-package.

FOM

The figure of merit to be used in the calculation. The default is "FOM_wAFROC". See UtilFigureOfMerit.

FPFValue

Only needed for LROC data and FOM = "PCL" or "ALROC"; where to evaluate a partial curve based figure of merit. The default is 0.2.

method

The method, in which the mean squares are calculated. The two valid choices are "DBM" (default) and "OR".

Value

A list containing the mean squares

Details

For DBM method, msT, msTR, msTC, msTRC will not be available if the dataset contains only one modality (NAs are returned). Similarly, msR, msTR, msRC, msTRC NAs are returned for single reader dataset. For ORH method, msT, msR, msTR will be returned for multiple reader multiple modality dataset. msT is not available for single modality dataset and msR is not available for single reader dataset.

Examples

result <- UtilMeanSquares(dataset02, FOM = "Wilcoxon")
result <- UtilMeanSquares(dataset05, FOM = "wAFROC", method = "OR")