Generate a formatted report file

Generate a formatted report of the analysis and save to a text file

UtilOutputReport (dataset, DataFileName, DataFileFormat, delimiter = ",",
   dataDescription = "MyData", ReportFileName, ReportFileFormat = "txt",
   stMethod = "DBMH", FOM = "wJAFROC", alpha = 0.05, covEstMethod = "Jackknife",
   nBoots = 200, renumber = FALSE, overwrite = TRUE)

Arguments

dataset
The dataset object to be analyzed (not the file name), RJafroc-package.

DataFileName
A string variable containing the name of the data file to be analyzed, see DfReadDataFile and "Details".

DataFileFormat
The format of the data specified in DataFileName: see DfReadDataFile and "Details". Must be specified if DataFileName is specified.

delimiter
See DfReadDataFile.

dataDescription
A description of the data, default is "MyData"

ReportFileName
The file name of the output report file. If this parameter is missing, the function will use DataFileName or dataDescription followed by the underscore separated concatenation of stMethod and FOM as the output report file.

ReportFileFormat
The format of the output report. The two available formats are "txt" (the default) and "xlsx", corresponding to a formatted text file and an Excel file, respectively.

stMethod
The significance testing method, "ORH" or "DBMH".

FOM
See StSignificanceTesting.

alpha
See StSignificanceTesting.

covEstMethod
See StSignificanceTesting.

nBoots
See StSignificanceTesting.

renumber
A logical variable: if TRUE, consecutive integers (staring from 1) will be used as the treatment and reader IDs in the output report. Otherwise, treatment and reader IDs in the original data file will be used. This option may be needed for aesthetics.

overwrite
A logical variable: if FALSE, a warning will be issued if the report file already exists and the program will wait until the user inputs "y" or "n" to determine whether to overwrite the existing file. If TRUE, an existing file will be silently overwritten.

Value

A formatted report of the data analysis, patterned roughly on that of OR-DBM MRMC V2.5.

Details

See examples

Examples

UtilOutputReport(dataset = dataset03, stMethod = "DBMH", FOM = "Wilcoxon",
   dataDescription = "MyROCData0", overwrite = TRUE)

#> The report has been saved to /Users/Dev/Documents/GitHub/rjafroc-master/docs/reference/MyROCData0_DBMH_Wilcoxon.txt.

# NOT RUN {
## Generate reports for a dataset object
UtilOutputReport(dataset = dataset02, stMethod = "DBMH", FOM = "Wilcoxon", 
             dataDescription = "MyROCData1", overwrite = TRUE)
             
UtilOutputReport(dataset = dataset02, stMethod = "DBMH", FOM = "Wilcoxon", 
dataDescription = "MyROCData2",ReportFileFormat = "xlsx", overwrite = TRUE)

UtilOutputReport(dataset = dataset02, stMethod = "ORH", FOM = "Wilcoxon", 
             dataDescription = "MyROCData3", overwrite = TRUE)
             
UtilOutputReport(dataset = dataset02, stMethod = "ORH", FOM = "Wilcoxon", 
dataDescription = "MyROCData4",ReportFileFormat = "xlsx", overwrite = TRUE)

## Generate report for a data file
fn <- system.file("extdata", "includedRocData.xlsx", 
package = "RJafroc", mustWork = TRUE)
UtilOutputReport(DataFileName = fn, DataFileFormat = "JAFROC", stMethod = "DBMH", FOM = "Wilcoxon",
             overwrite = TRUE, ReportFileFormat = "xlsx")
             
## Output report for an existing dataset
## UtilOutputReport(dataset = dataset05, stMethod = "DBMH", FOM = "Wilcoxon") # ERROR! as FOM is 
   incompatible with FROC data

UtilOutputReport(dataset = dataset05, stMethod = "ORH") # OK as default FOM is "wJAFROC"

UtilOutputReport(dataset = dataset05, stMethod = "DBMH", FOM = "HrAuc")

UtilOutputReport(dataset = dataset05, stMethod = "DBMH", FOM = "HrAuc", ReportFileFormat = "xlsx")
# }