Generate a formatted report file

Generate a formatted report of the analysis and save to a text file

UtilOutputReport(dataset, DataFileName, DataFileFormat, delimiter = ",",
  dataDescription = "MyData", ReportFileName, ReportFileFormat = "txt",
  method = "DBMH", FOM = "Wilcoxon", alpha = 0.05,
  covEstMethod = "Jackknife", nBoots = 200, renumber = FALSE,
  overwrite = TRUE)

Arguments

dataset
The dataset object to be analyzed (not the file name), see Dataset in RJafroc-package.

DataFileName
A string variable containing the name of the data file to be analyzed, in the form baseName.ext, see DfReadDataFile.

DataFileFormat
The format of the data specified in DataFileName: see DfReadDataFile. Must be specified if DataFileName is specified. Valid values are "JAFROC", "MRMC" or "iMRMC"

delimiter
used in text input data file, See DfReadDataFile.

dataDescription
Optional description of the data, default is "MyData"

ReportFileName
The file name of the output report file. If this parameter is missing, the function will use DataFileName or dataDescription followed by the underscore separated concatenation of method and FOM as the output report file.

ReportFileFormat
The format of the output report. The two available formats are "txt" (the default) and "xlsx", corresponding to a formatted text file or an Excel file, respectively.

method
The significance testing method, "ORH" or (the default) "DBMH".

FOM
The figure of merit; see StSignificanceTesting.

alpha
See StSignificanceTesting.

covEstMethod
See StSignificanceTesting; only needed for "ORH" analysis.

nBoots
See StSignificanceTesting; only needed for "ORH" analysis.

renumber
A logical variable: if TRUE, consecutive integers (staring from 1) will be used as the treatment and reader IDs in the output report. Otherwise, treatment and reader IDs in the original data file will be used. This option may be needed for aesthetics.

overwrite
A logical variable: if FALSE, a warning will be issued if the report file already exists and the program will wait until the user inputs "y" or "n" to determine whether to overwrite the existing file. If TRUE, an existing file will be silently overwritten.

Value

sigTestResult The object returned by StSignificanceTesting.

Details

A formatted report of the data analysis, patterned roughly on that of OR-DBM MRMC V2.5, is written to the output file.

Examples

UtilOutputReport(dataset = dataset03, overwrite = TRUE)

#> $fomArray
#>           Rdr - 0   Rdr - 1   Rdr - 2   Rdr - 3
#> Trt - 0 0.8534600 0.8649932 0.8573044 0.8152420
#> Trt - 1 0.8496156 0.8435097 0.8401176 0.8143374
#> 
#> $anovaY
#>   Source           SS  DF          MS
#> 1      T   0.02356541   1 0.023565410
#> 2      R   0.20521800   3 0.068406000
#> 3      C  52.52839868  99 0.530589886
#> 4     TR   0.01506079   3 0.005020264
#> 5     TC   6.41004881  99 0.064747968
#> 6     RC  39.24295381 297 0.132131158
#> 7    TRC  22.66007764 297 0.076296558
#> 8  Total 121.08532315 799          NA
#> 
#> $anovaYi
#>   Source  DF          0          1
#> 1      R   3 0.04926635 0.02415991
#> 2      C  99 0.29396753 0.30137032
#> 3     RC 297 0.10504787 0.10337984
#> 
#> $varComp
#>                  varComp
#> Var(R)      3.775568e-05
#> Var(C)      5.125091e-02
#> Var(T*R)   -7.127629e-04
#> Var(T*C)   -2.887147e-03
#> Var(R*C)    2.791730e-02
#> Var(Error)  7.629656e-02
#> 
#> $fRRRC
#> [1] 4.694058
#> 
#> $ddfRRRC
#> [1] 3
#> 
#> $pRRRC
#> [1] 0.1188379
#> 
#> $ciDiffTrtRRRC
#>   Treatment   Estimate      StdErr DF        t     PrGTt      CILower
#> 1     0 - 1 0.01085482 0.005010122  3 2.166577 0.1188379 -0.005089627
#>      CIUpper
#> 1 0.02679926
#> 
#> $ciAvgRdrEachTrtRRRC
#>   Treatment      Area     StdErr        DF   CILower   CIUpper
#> 1         0 0.8477499 0.02440215  70.12179 0.7990828 0.8964170
#> 2         1 0.8368951 0.02356642 253.64403 0.7904843 0.8833058
#> 
#> $fFRRC
#> [1] 0.363956
#> 
#> $ndf
#> [1] 1
#> 
#> $ddfFRRC
#> [1] 99
#> 
#> $pFRRC
#> [1] 0.547697
#> 
#> $ciDiffTrtFRRC
#>   Treatment   Estimate     StdErr DF         t    PrGTt     CILower    CIUpper
#> 1     0 - 1 0.01085482 0.01799277 99 0.6032876 0.547697 -0.02484675 0.04655638
#> 
#> $ciAvgRdrEachTrtFRRC
#>   Treatment      Area     StdErr DF   CILower   CIUpper
#> 1         0 0.8477499 0.02710939 99 0.7939590 0.9015408
#> 2         1 0.8368951 0.02744860 99 0.7824311 0.8913591
#> 
#> $ssAnovaEachRdr
#>   Source DF            0           1           2            3
#> 1      T  1 7.389761e-04  0.02307702  0.01476929 4.091217e-05
#> 2      C 99 2.018360e+01 22.12074893 21.21043057 2.825657e+01
#> 3     TC 99 9.064315e+00  7.94764631  6.06166901 5.996496e+00
#> 
#> $msAnovaEachRdr
#>   Source DF            0          1          2            3
#> 1      T  1 0.0007389761 0.02307702 0.01476929 4.091217e-05
#> 2      C 99 0.2038747746 0.22344191 0.21424677 2.854199e-01
#> 3     TC 99 0.0915587344 0.08027926 0.06122898 6.057067e-02
#> 
#> $ciDiffTrtEachRdr
#>   Reader Treatment     Estimate     StdErr DF          t     PrGTt     CILower
#> 1      0     0 - 1 0.0038444143 0.04279223 99 0.08983908 0.9285966 -0.08106465
#> 2      1     0 - 1 0.0214834916 0.04006975 99 0.53615233 0.5930559 -0.05802359
#> 3      2     0 - 1 0.0171867933 0.03499399 99 0.49113552 0.6244176 -0.05224888
#> 4      3     0 - 1 0.0009045681 0.03480536 99 0.02598933 0.9793182 -0.06815683
#>      CIUpper
#> 1 0.08875348
#> 2 0.10099057
#> 3 0.08662247
#> 4 0.06996596
#> 
#> $fRRFC
#> [1] 4.694058
#> 
#> $ddfRRFC
#> [1] 3
#> 
#> $pRRFC
#> [1] 0.1188379
#> 
#> $ciDiffTrtRRFC
#>   Treatment   Estimate      StdErr DF        t     PrGTt      CILower
#> 1     0 - 1 0.01085482 0.005010122  3 2.166577 0.1188379 -0.005089627
#>      CIUpper
#> 1 0.02679926
#> 
#> $ciAvgRdrEachTrtRRFC
#>   Treatment      Area     StdErr DF   CILower   CIUpper
#> 1         0 0.8477499 0.01109801  3 0.8124311 0.8830687
#> 2         1 0.8368951 0.00777173  3 0.8121620 0.8616282
#> 

# NOT RUN {
## Generate reports for a dataset object
UtilOutputReport(dataset = dataset02, method = "DBMH", 
             dataDescription = "MyROCData1", overwrite = TRUE)
             
UtilOutputReport(dataset = dataset02, method = "DBMH", 
dataDescription = "MyROCData2",ReportFileFormat = "xlsx", overwrite = TRUE)

UtilOutputReport(dataset = dataset02, method = "ORH", 
             dataDescription = "MyROCData3", overwrite = TRUE)
             
UtilOutputReport(dataset = dataset02, method = "ORH", 
dataDescription = "MyROCData4",ReportFileFormat = "xlsx", overwrite = TRUE)

## Generate report for a data file
fn <- system.file("extdata", "includedRocData.xlsx", 
package = "RJafroc", mustWork = TRUE)
UtilOutputReport(DataFileName = fn, DataFileFormat = "JAFROC", method = "DBMH",
             overwrite = TRUE, ReportFileFormat = "xlsx")
             
## Output report for an existing dataset
## UtilOutputReport(dataset05, method = "DBMH") # ERROR as FOM is 
   incompatible with FROC data

UtilOutputReport(dataset05, method = "ORH", FOM = "wAFROC")

UtilOutputReport(dataset05, method = "DBMH", FOM = "HrAuc")

UtilOutputReport(dataset05, method = "DBMH", FOM = "HrAuc", ReportFileFormat = "xlsx")
# }