Generate a text formatted report file or an Excel file

Generates a formatted report of the analysis and saves it to a text or an Excel file

UtilOutputReport(dataset, dataDescription = "MyDataDescription: ",
  ReportFileBaseName = NULL, ReportFileExt = "txt", method = "DBMH",
  FOM, alpha = 0.05, covEstMethod = "Jackknife", nBoots = 200,
  sequentialNames = FALSE, overWrite = FALSE)

Arguments

dataset
The dataset object to be analyzed (not the file name), see Dataset in RJafroc-package.

dataDescription
A short string describing the dataset, the default is "MyDataDescription: "; it is inserted, for reference, in the output file.

ReportFileBaseName
The file base name (sans the extension) for the desired report; the default is NULL, in which case the system generates a temporary text file, whose name is displayed.

ReportFileExt
The report file extension determines the type of output. txt, the default, for a text file, xlsx for an Excel file.

method
The significance testing method, "ORH" or (the default) "DBMH".

FOM
The figure of merit; see StSignificanceTesting.

alpha
See StSignificanceTesting; the default is 0.05.

covEstMethod
See StSignificanceTesting; only needed for method = "ORH"; the default is "Jackknife".

nBoots
See StSignificanceTesting; only needed for "ORH" analysis; the default is 200.

sequentialNames
A logical variable: if TRUE, consecutive integers (staring from 1) will be used as the treatment and reader IDs in the output report. Otherwise, treatment and reader IDs in the original data file will be used. This option may be needed for aesthetics. The default is FALSE.

overWrite
A logical variable: if FALSE, a warning will be issued if the report file already exists and the program will wait until the user inputs "y" or "n" to determine whether to overwrite the existing file. If TRUE, the existing file will be silently overwritten. The default is FALSE.

Value

sigTestResult The object returned by StSignificanceTesting.

Details

A formatted report of the data analysis is written to the output file in either text or Excel format.

Examples

 # text output is created in a temporary file
UtilOutputReport(dataset03, FOM = "Wilcoxon")

#> 
#> Output file name is: 	 /var/folders/d1/mx6dcbzx3v39r260458z2b200000gn/T//Rtmph9Ju36/RJafrocUtilOutputReportbfb61d0fd36b.txt 

#> $fomArray
#>           Rdr0      Rdr1      Rdr2      Rdr3
#> Trt0 0.8534600 0.8649932 0.8573044 0.8152420
#> Trt1 0.8496156 0.8435097 0.8401176 0.8143374
#> 
#> $anovaY
#>   Source           SS  DF          MS
#> 1      T   0.02356541   1 0.023565410
#> 2      R   0.20521800   3 0.068406000
#> 3      C  52.52839868  99 0.530589886
#> 4     TR   0.01506079   3 0.005020264
#> 5     TC   6.41004881  99 0.064747968
#> 6     RC  39.24295381 297 0.132131158
#> 7    TRC  22.66007764 297 0.076296558
#> 8  Total 121.08532315 799          NA
#> 
#> $anovaYi
#>   Source  DF       Trt0       Trt1
#> 1      R   3 0.04926635 0.02415991
#> 2      C  99 0.29396753 0.30137032
#> 3     RC 297 0.10504787 0.10337984
#> 
#> $varComp
#>           varR       varC         varTR        varTC     varRC     varErr
#> 1 3.775568e-05 0.05125091 -0.0007127629 -0.002887147 0.0279173 0.07629656
#> 
#> $FTestStatsRRRC
#>      fRRRC ndfRRRC ddfRRRC     pRRRC
#> 1 4.694058       1       3 0.1188379
#> 
#> $ciDiffTrtRRRC
#>     TrtDiff   Estimate      StdErr DF        t     PrGTt      CILower
#> 1 Trt0-Trt1 0.01085482 0.005010122  3 2.166577 0.1188379 -0.005089627
#>      CIUpper
#> 1 0.02679926
#> 
#> $ciAvgRdrEachTrtRRRC
#>   Treatment      Area     StdErr        DF   CILower   CIUpper
#> 1      Trt0 0.8477499 0.02440215  70.12179 0.7990828 0.8964170
#> 2      Trt1 0.8368951 0.02356642 253.64403 0.7904843 0.8833058
#> 
#> $FTestStatsFRRC
#>      fFRRC ndfFRRC ddfFRRC    pFRRC
#> 1 0.363956       1      99 0.547697
#> 
#> $ciDiffTrtFRRC
#>   Treatment   Estimate     StdErr DF         t    PrGTt     CILower    CIUpper
#> 1 Trt0-Trt1 0.01085482 0.01799277 99 0.6032876 0.547697 -0.02484675 0.04655638
#> 
#> $ciAvgRdrEachTrtFRRC
#>   Treatment      Area     StdErr DF   CILower   CIUpper
#> 1      Trt0 0.8477499 0.02710939 99 0.7939590 0.9015408
#> 2      Trt1 0.8368951 0.02744860 99 0.7824311 0.8913591
#> 
#> $msAnovaEachRdrFRRC
#>   Source DF         Rdr0       Rdr1       Rdr2         Rdr3
#> 1      T  1 0.0007389761 0.02307702 0.01476929 4.091217e-05
#> 2      C 99 0.2038747746 0.22344191 0.21424677 2.854199e-01
#> 3     TC 99 0.0915587344 0.08027926 0.06122898 6.057067e-02
#> 
#> $ciDiffTrtEachRdrFRRC
#>   Reader Treatment     Estimate     StdErr DF          t     PrGTt     CILower
#> 1   Rdr0 Trt0-Trt1 0.0038444143 0.04279223 99 0.08983908 0.9285966 -0.08106465
#> 2   Rdr1 Trt0-Trt1 0.0214834916 0.04006975 99 0.53615233 0.5930559 -0.05802359
#> 3   Rdr2 Trt0-Trt1 0.0171867933 0.03499399 99 0.49113552 0.6244176 -0.05224888
#> 4   Rdr3 Trt0-Trt1 0.0009045681 0.03480536 99 0.02598933 0.9793182 -0.06815683
#>      CIUpper
#> 1 0.08875348
#> 2 0.10099057
#> 3 0.08662247
#> 4 0.06996596
#> 
#> $FTestStatsRRFC
#>      fRRFC ndfRRFC ddfRRFC     pRRFC
#> 1 4.694058       1       3 0.1188379
#> 
#> $ciDiffTrtRRFC
#>   Treatment   Estimate      StdErr DF        t     PrGTt      CILower
#> 1 Trt0-Trt1 0.01085482 0.005010122  3 2.166577 0.1188379 -0.005089627
#>      CIUpper
#> 1 0.02679926
#> 
#> $ciAvgRdrEachTrtRRFC
#>   Treatment      Area     StdErr DF   CILower   CIUpper
#> 1      Trt0 0.8477499 0.01109801  3 0.8124311 0.8830687
#> 2      Trt1 0.8368951 0.00777173  3 0.8121620 0.8616282
#> 

# Excel output is created in a temporary file
UtilOutputReport(dataset03, FOM = "Wilcoxon", ReportFileExt = "xlsx")

#> 
#> Output file name is: 	 /var/folders/d1/mx6dcbzx3v39r260458z2b200000gn/T//Rtmph9Ju36/RJafrocUtilOutputReportbfb67a0124b4.xlsx 

#> $fomArray
#>           Rdr0      Rdr1      Rdr2      Rdr3
#> Trt0 0.8534600 0.8649932 0.8573044 0.8152420
#> Trt1 0.8496156 0.8435097 0.8401176 0.8143374
#> 
#> $anovaY
#>   Source           SS  DF          MS
#> 1      T   0.02356541   1 0.023565410
#> 2      R   0.20521800   3 0.068406000
#> 3      C  52.52839868  99 0.530589886
#> 4     TR   0.01506079   3 0.005020264
#> 5     TC   6.41004881  99 0.064747968
#> 6     RC  39.24295381 297 0.132131158
#> 7    TRC  22.66007764 297 0.076296558
#> 8  Total 121.08532315 799          NA
#> 
#> $anovaYi
#>   Source  DF       Trt0       Trt1
#> 1      R   3 0.04926635 0.02415991
#> 2      C  99 0.29396753 0.30137032
#> 3     RC 297 0.10504787 0.10337984
#> 
#> $varComp
#>           varR       varC         varTR        varTC     varRC     varErr
#> 1 3.775568e-05 0.05125091 -0.0007127629 -0.002887147 0.0279173 0.07629656
#> 
#> $FTestStatsRRRC
#>      fRRRC ndfRRRC ddfRRRC     pRRRC
#> 1 4.694058       1       3 0.1188379
#> 
#> $ciDiffTrtRRRC
#>     TrtDiff   Estimate      StdErr DF        t     PrGTt      CILower
#> 1 Trt0-Trt1 0.01085482 0.005010122  3 2.166577 0.1188379 -0.005089627
#>      CIUpper
#> 1 0.02679926
#> 
#> $ciAvgRdrEachTrtRRRC
#>   Treatment      Area     StdErr        DF   CILower   CIUpper
#> 1      Trt0 0.8477499 0.02440215  70.12179 0.7990828 0.8964170
#> 2      Trt1 0.8368951 0.02356642 253.64403 0.7904843 0.8833058
#> 
#> $FTestStatsFRRC
#>      fFRRC ndfFRRC ddfFRRC    pFRRC
#> 1 0.363956       1      99 0.547697
#> 
#> $ciDiffTrtFRRC
#>   Treatment   Estimate     StdErr DF         t    PrGTt     CILower    CIUpper
#> 1 Trt0-Trt1 0.01085482 0.01799277 99 0.6032876 0.547697 -0.02484675 0.04655638
#> 
#> $ciAvgRdrEachTrtFRRC
#>   Treatment      Area     StdErr DF   CILower   CIUpper
#> 1      Trt0 0.8477499 0.02710939 99 0.7939590 0.9015408
#> 2      Trt1 0.8368951 0.02744860 99 0.7824311 0.8913591
#> 
#> $msAnovaEachRdrFRRC
#>   Source DF         Rdr0       Rdr1       Rdr2         Rdr3
#> 1      T  1 0.0007389761 0.02307702 0.01476929 4.091217e-05
#> 2      C 99 0.2038747746 0.22344191 0.21424677 2.854199e-01
#> 3     TC 99 0.0915587344 0.08027926 0.06122898 6.057067e-02
#> 
#> $ciDiffTrtEachRdrFRRC
#>   Reader Treatment     Estimate     StdErr DF          t     PrGTt     CILower
#> 1   Rdr0 Trt0-Trt1 0.0038444143 0.04279223 99 0.08983908 0.9285966 -0.08106465
#> 2   Rdr1 Trt0-Trt1 0.0214834916 0.04006975 99 0.53615233 0.5930559 -0.05802359
#> 3   Rdr2 Trt0-Trt1 0.0171867933 0.03499399 99 0.49113552 0.6244176 -0.05224888
#> 4   Rdr3 Trt0-Trt1 0.0009045681 0.03480536 99 0.02598933 0.9793182 -0.06815683
#>      CIUpper
#> 1 0.08875348
#> 2 0.10099057
#> 3 0.08662247
#> 4 0.06996596
#> 
#> $FTestStatsRRFC
#>      fRRFC ndfRRFC ddfRRFC     pRRFC
#> 1 4.694058       1       3 0.1188379
#> 
#> $ciDiffTrtRRFC
#>   Treatment   Estimate      StdErr DF        t     PrGTt      CILower
#> 1 Trt0-Trt1 0.01085482 0.005010122  3 2.166577 0.1188379 -0.005089627
#>      CIUpper
#> 1 0.02679926
#> 
#> $ciAvgRdrEachTrtRRFC
#>   Treatment      Area     StdErr DF   CILower   CIUpper
#> 1      Trt0 0.8477499 0.01109801  3 0.8124311 0.8830687
#> 2      Trt1 0.8368951 0.00777173  3 0.8121620 0.8616282
#>