Utility for Obuchowski-Rockette variance components

Utility for Obuchowski-Rockette variance components

UtilVarComponentsOR(
  dataset,
  FOM,
  FPFValue = 0.2,
  covEstMethod = "jackknife",
  nBoots = 200,
  seed = NULL
)

Arguments

dataset
The dataset object

FOM
The figure of merit

FPFValue
Only needed for LROC data and FOM = "PCL" or "ALROC"; where to evaluate a partial curve based figure of merit. The default is 0.2.

covEstMethod
The covariance estimation method, "jackknife" (the default) or "bootstrap".

nBoots
Only needed for bootstrap covariance estimation method. The number of bootstraps, defaults to 200.

seed
Only needed for the bootstrap covariance estimation method. The initial seed for the random number generator, the default is NULL, as if no seed has been specified.

Value

A list object containing the following data.frames:

foms: the figures of merit for different treatment-reader combinations

TRanova: the OR treatment-reader ANOVA table

VarCom: the OR variance-components Cov1, Cov2, Cov3, Var and correlations rho1, rho2 and rho3

IndividualTrt: the individual treatment mean-squares, Var and Cov2 values

IndividualRdr: the individual reader mean-squares, Var and Cov1 values

Details

The variance components are obtained using StSignificanceTesting with method = "ORH".

Examples

## uses the default jackknife for covEstMethod
UtilVarComponentsOR(dataset02, FOM = "Wilcoxon")$foms

#> NULL

UtilVarComponentsOR(dataset02, FOM = "Wilcoxon")$TRanova

#>              SS DF            MS
#> T  0.0047961705  1 0.00479617053
#> R  0.0153448000  4 0.00383619999
#> TR 0.0022041225  4 0.00055103062

UtilVarComponentsOR(dataset02, FOM = "Wilcoxon")$VarCom

#>           Estimates       Rhos
#> VarR  0.00153499935         NA
#> VarTR 0.00020040252         NA
#> Cov1  0.00034661371 0.43203138
#> Cov2  0.00034407483 0.42886683
#> Cov3  0.00023902837 0.29793328
#> Var   0.00080228827         NA

UtilVarComponentsOR(dataset02, FOM = "Wilcoxon")$IndividualTrt

#>      DF   msREachTrt    varEachTrt   cov2EachTrt
#> trt0  4 0.0030826287 0.00101410277 0.00048396180
#> trt1  4 0.0013046019 0.00059047376 0.00020418785

UtilVarComponentsOR(dataset02, FOM = "Wilcoxon")$IndividualRdr

#>      DF    msTEachRdr    varEachRdr   cov1EachRdr
#> rdr0  1 0.00039706618 0.00069890056 3.7346610e-04
#> rdr1  1 0.00108288538 0.00110605283 7.6015978e-04
#> rdr2  1 0.00015974702 0.00084234345 3.5532241e-04
#> rdr3  1 0.00034447841 0.00015057773 1.0833995e-06
#> rdr4  1 0.00501611603 0.00121356676 2.4303685e-04

UtilVarComponentsOR(dataset02, FOM = "Wilcoxon",
   covEstMethod = "bootstrap", nBoots = 2000, seed = 100)$VarCom

#>           Estimates       Rhos
#> VarR  0.00155082365         NA
#> VarTR 0.00019451356         NA
#> Cov1  0.00032833333 0.41690698
#> Cov2  0.00033926769 0.43079107
#> Cov3  0.00023657230 0.30039181
#> Var   0.00078754579         NA

UtilVarComponentsOR(dataset02, FOM = "Wilcoxon", covEstMethod = "DeLong")$VarCom

#>           Estimates       Rhos
#> VarR  0.00153642538         NA
#> VarTR 0.00020458400         NA
#> Cov1  0.00034200896 0.43175729
#> Cov2  0.00033952653 0.42862343
#> Cov3  0.00023584965 0.29774017
#> Var   0.00079213245         NA