Dobbins 3 FROC dataset

This is referred to in the code as the "DOB3" dataset. This is a subset of DOB1 which includes data for lesions not-visible on CXR, but visible to truth panel on all treatments.

dataset13

Format

A list with 3 elements: $ratings, $lesions and $descriptions; $ratings contain 3 elements, $NL, $LL and $LL_IL as sub-lists; $lesions contain 3 elements, $perCase, $IDs and $weights as sub-lists; $descriptions contain 7 elements, $fileName, $type, $name, $truthTableStr, $design, $modalityID and $readerID as sub-lists;

rating$NL, num [1:4, 1:5, 1:158, 1:4], ratings of non-lesion localizations, NLs

rating$LL, num [1:4, 1:5, 1:106, 1:15], ratings of lesion localizations, LLs

rating$LL_ILNA, this placeholder is used only for LROC data

lesions$perCase, int [1:106], number of lesions per diseased case

lesions$IDs, num [1:106, 1:15], numeric labels of lesions on diseased cases

lesions$weights, num [1:106, 1:15], weights (or clinical importances) of lesions

descriptions$fileName, chr, "dataset13", base name of dataset in `data` folder

descriptions$type, chr "FROC", the data type

descriptions$name, chr "DOBBINS-3", the name of the dataset

descriptions$truthTableStr, num [1:4, 1:5, 1:158, 1:16], truth table structure

descriptions$design, chr "FCTRL", study design, factorial dataset

descriptions$modalityID, chr [1:4] "1" "2" "3" "4", modality label(s)

descriptions$readerID, chr [1:5] "1" "2" "3" "4" ..., reader labels

References

Dobbins III JT et al. Multi-Institutional Evaluation of Digital Tomosynthesis, Dual-Energy Radiography, and Conventional Chest Radiography for the Detection and Management of Pulmonary Nodules. Radiology. 2016;282(1):236-250.

Examples

res <- str(dataset13)
#> List of 3
#>  $ ratings     :List of 3
#>   ..$ NL   : num [1:4, 1:5, 1:158, 1:4] -Inf 3 -Inf 4 5 ...
#>   ..$ LL   : num [1:4, 1:5, 1:106, 1:15] -Inf -Inf -Inf -Inf -Inf ...
#>   ..$ LL_IL: logi NA
#>  $ lesions     :List of 3
#>   ..$ perCase: int [1:106] 3 1 7 2 3 6 10 1 4 5 ...
#>   ..$ IDs    : num [1:106, 1:15] 2 3 1 2 1 2 1 2 1 2 ...
#>   ..$ weights: num [1:106, 1:15] 0.333 1 0.143 0.5 0.333 ...
#>  $ descriptions:List of 7
#>   ..$ fileName     : chr "dataset13"
#>   ..$ type         : chr "FROC"
#>   ..$ name         : chr "DOBBINS-3"
#>   ..$ truthTableStr: num [1:4, 1:5, 1:158, 1:16] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ design       : chr "FCTRL"
#>   ..$ modalityID   : Named chr [1:4] "1" "2" "3" "4"
#>   .. ..- attr(*, "names")= chr [1:4] "1" "2" "3" "4"
#>   ..$ readerID     : Named chr [1:5] "1" "2" "3" "4" ...
#>   .. ..- attr(*, "names")= chr [1:5] "1" "2" "3" "4" ...