Nico Karssemeijer LROC dataset (CAD vs. radiologists)

This is the actual LROC data corresponding to dataset09, which was the inferred ROC data. Note that the LL field is split into two, LL, representing true positives where the lesions were correctly localized, and LL_IL, representing true positives where the lesions were incorrectly localized. The first reader is CAD and the remaining readers are radiologists.

datasetCadLroc

Format

A list with 3 elements: $ratings, $lesions and $descriptions; $ratings contain 3 elements, $NL, $LL and $LL_IL as sub-lists; $lesions contain 3 elements, $perCase, $IDs and $weights as sub-lists; $descriptions contain 7 elements, $fileName, $type, $name, $truthTableStr, $design, $modalityID and $readerID as sub-lists;

rating$NL, num [1, 1:10, 1:200, 1], ratings of localizations on normal cases

rating$LL, num [1, 1:10, 1:80, 1], ratings of correct localizations on abnormal cases

rating$LL_ILnum [1, 1:10, 1:80, 1], ratings of incorrect localizations on abnormal cases

lesions$perCase, int [1:80], number of lesions per diseased case

lesions$IDs, num [1:80, 1] , numeric labels of lesions on diseased cases

lesions$weights, num [1:80, 1], weights (or clinical importances) of lesions

descriptions$fileName, chr, "datasetCadLroc", base name of dataset in `data` folder

descriptions$type, chr "LROC", the data type

descriptions$name, chr "NICO-CAD-LROC", the name of the dataset

descriptions$truthTableStr, num [1:2, 1:4, 1:200, 1:2], truth table structure

descriptions$design, chr "FCTRL", study design, factorial dataset

descriptions$modalityID, chr "1", treatment label(s)

descriptions$readerID, chr [1:10] "1" "2" "3" "4" ..., reader labels

References

Hupse R et al. Standalone computer-aided detection compared to radiologists' performance for the detection of mammographic masses. Eur Radiol. 2013;23(1):93-100.

Examples

str(datasetCadLroc)
#> List of 3
#>  $ ratings     :List of 3
#>   ..$ NL   : num [1, 1:10, 1:200, 1] 28 0 14 0 16 0 31 0 0 0 ...
#>   ..$ LL   : num [1, 1:10, 1:80, 1] 0 0 0 0 0 0 0 0 67 0 ...
#>   ..$ LL_IL: num [1, 1:10, 1:80, 1] 29 12 13 10 41 67 61 51 0 0 ...
#>  $ lesions     :List of 3
#>   ..$ perCase: num [1:80] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ IDs    : num [1:80, 1] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ weights: num [1:80, 1] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ descriptions:List of 7
#>   ..$ fileName     : chr "datasetCadLroc"
#>   ..$ type         : chr "LROC"
#>   ..$ name         : chr "NICO-CAD-LROC"
#>   ..$ truthTableStr: num [1, 1:10, 1:200, 1:2] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ design       : chr "FCTRL"
#>   ..$ modalityID   : chr "1"
#>   ..$ readerID     : chr [1:10] "1" "2" "3" "4" ...