Simulated FROC CAD vs. RAD dataset

Simulated FROC CAD vs. RAD dataset suitable for checking code. It was generated from datasetCadLroc using SimulateFrocFromLrocData.R. The LROC paradigm always yields a single mark per case. Therefore the equivalent FROC will also have only one mark per case. The NL arrays of the two datasets are identical. The LL array is created by copying the LL (correct localiztion) array of the LROC dataset to the LL array of the FROC dataset, from diseased case index k2 = 1 to k2 = K2. Additionally, the LL_IL array of the LROC dataset is copied to the NL array of the FROC dataset, starting at case index k1 = K1+1 to k1 = K1+K2. Any zero ratings are replace by -Infs. The equivalent FROC dataset has the same HrAuc as the original LROC dataset. See example. The main use of this dataset & function is to test the CAD significance testing functions using CAD FROC datasets, which I currently don't have.

datasetCadSimuFroc

Format

A list with 3 elements: $ratings, $lesions and $descriptions; $ratings contain 3 elements, $NL, $LL and $LL_IL as sub-lists; $lesions contain 3 elements, $perCase, $IDs and $weights as sub-lists; $descriptions contain 7 elements, $fileName, $type, $name, $truthTableStr, $design, $modalityID and $readerID as sub-lists;

rating$NL, num [1, 1:10, 1:200, 1], ratings of non-lesion localizations, NLs

rating$LL, num [1, 1:10, 1:80, 1], ratings of lesion localizations, LLs

rating$LL_ILNA, this placeholder is used only for LROC data

lesions$perCase, int [1:80], number of lesions per diseased case

lesions$IDs, num [1:80, 1] , numeric labels of lesions on diseased cases

lesions$weights, num [1:80, 1], weights (or clinical importances) of lesions

descriptions$fileName, chr, "datasetCadSimuFroc", base name of dataset in `data` folder

descriptions$type, chr "LROC", the data type

descriptions$name, chr "NICO-CAD-LROC", the name of the dataset

descriptions$truthTableStr, num [1:2, 1:4, 1:200, 1:2], truth table structure

descriptions$design, chr "FCTRL", study design, factorial dataset

descriptions$modalityID, chr "1", treatment label(s)

descriptions$readerID, chr [1:10] "1" "2" "3" "4" ..., reader labels